==> 1_wrangling.Rmd <==
---
title: "1_Wrangling"
author: "JEON, CHEONGOK and KIM DAEHYUN"
date: '2021 3 9 '
output: html_document
---

# Aims

Getting raw data from gbif via rgbif package and conduct data pre-processing.
This will include followings.
1. get rid of points with low percise coordinate.
2. get rid of points that lack of bio-climate information.

After pre-processing, making divergent dataset to compare "with outlier(dispersal)" and "without outliers"
These two dataset will be compared to check whether remotely located outliers will effect the model outputs or not. This process was conducted on QGIS, so not included in this chapter but describing exporting, importing process.

To obtain proper percision, every points less than 5 decimal places were removed.
After bio climates were extracted, every points lack of bio-climate values were removed.
Chapter 1 include gbif data retriving process.
Chapter 2 include decimal places removal.
Chapter 3 include bio-climate NAs removal.

# CH.1 Get gbif data

```{r}
library(tidyverse)
library(readxl)
library(data.table)
```



```{r read target species list}
specieslist_mammal <- readxl::read_xlsx("species.xlsx", sheet = 1)
specieslist_reptile <- readxl::read_xlsx("species.xlsx", sheet = 2)
specieslist_amphibian <- readxl::read_xlsx("species.xlsx", sheet = 3)
specieslist_aves <- readxl::read_xlsx("species.xlsx", sheet = 4)
```


```{r mammal dataset}
#bind two mammal dataset
mammal.csv <- data.table::fread(file = "out/mammal/realout_mammal.csv")
reptile.csv. <- data.table::fread(file = "out/reptile/realout_reptile.csv")
amphibian.csv <- data.table::fread(file = "out/amphibian/realout_amphibian.csv")
mammal.csv <- data.table::fread(file = "out/mammal/realout_mammal.csv")

##coordinate NAs removal
mammal.csv <- filter(mammal.csv, complete.cases(mammal.csv$decimalLatitude))
reptile.csv <- filter(reptile.csv, complete.cases(reptile.csv$decimalLatitude))
amphibian.csv <- filter(amphibian.csv, complete.cases(amphibian.csv$decimalLatitude))

#decimal lower removal
decimalplaces <- function(x) {
  ifelse(abs(x - round(x)) > .Machine$double.eps^0.5,
         nchar(sub('^\\d+\\.', '', sub('0+$', '', as.character(x)))),
         0)
}

#filter long, lat lower than 3 decimal places
filter_longd3 <- function(x){filter(x, decimalplaces(decimalLongitude) >= 3)}
filter_lad3 <- function(x){filter(x, decimalplaces(decimalLatitude) >= 3)}

mammal.csv <- filter_longd3(mammal.csv)
mammal.csv <- filter_lad3(mammal.csv)

reptile.csv <- filter_longd3(reptile.csv)
reptile.csv <- filter_lad3(reptile.csv)

amphibian.csv <- filter_longd3(amphibian.csv)
amphibian.csv <- filter_lad3(amphibian.csv)

mammal.csv <- filter_longd3(mammal.csv)
mammal.csv <- filter_lad3(mammal.csv)

#add decimal palces column 
mutate_long <- function(x){mutate(x, decimalplaces_long = decimalplaces(decimalLongitude))}
mutate_lat <- function(x){mutate(x, decimalplaces_lat = decimalplaces(decimalLatitude))}

mammal.csv <- mutate_long(mammal.csv)
mammal.csv <- mutate_lat(mammal.csv)

reptile.csv <- mutate_long(reptile.csv)
reptile.csv <- mutate_lat(reptile.csv)

amphibian.csv <- mutate_long(amphibian.csv)
amphibian.csv <- mutate_lat(amphibian.csv)

#find duplicated points
unique.longlat <- function(x){dplyr::distinct(x, decimalLongitude, decimalLatitude, .keep_all = T)}

mammal.csv <- unique.longlat(mammal.csv)
reptile.csv <- unique.longlat(reptile.csv)
amphibian.csv <- unique.longlat(amphibian.csv)

#1st wrangling score
ggplot(mammal.csv, aes(x = species)) +
  geom_bar(stat = "count")+ 
  theme(axis.text.x=element_text(angle=45, hjust=1))
  
ggplot(reptile.csv, aes(x = species)) +
  geom_bar(stat = "count")+ 
  theme(axis.text.x=element_text(angle=45, hjust=1))

#save to list.


dt.mammal <- list()
for (i in 1:nrow(specieslist_mammal[,1])){
 dt.mammal[[i]] <- filter(mammal.csv, species == specieslist_mammal[[1]][i])
}

dt.reptile <- list()
for (i in 1:nrow(specieslist_reptile[,1])){
 dt.reptile[[i]] <- filter(reptile.csv, species == specieslist_reptile[[1]][i])
}

dt.amphibian <- list()
for (i in 1:nrow(specieslist_amphibian[,1])){
 dt.amphibian[[i]] <- filter(amphibian.csv, species == specieslist_amphibian[[1]][i])
}
```

> group_by(mammal.csv, species)
# A tibble: 984,384 x 52
# Groups:   species [35]

mammal은 35종류로 나눠 떨어지지만

> group_by(reptile.csv, species)
# A tibble: 787,244 x 52
# Groups:   species [106]

파충류는 애초에 96종을 요청했음에도 종이 106종이다. 아종이 분리된 것이다.
이들을 묶어야한다.

원인은 Ameiva ameiva였다. 

```{r}
reptile.csv %>% 
  group_by(species) %>%
  summarise(n = length(species))
```
다 더해도 얼마 안되므로(601) 묶기 귀찮으니 그냥 다 버린다.


map_int(dt.mammal, nrow)
[1]   3807   5517 108791   1643  12515 116916   4316  76928   6348   2376  31175  27682
[13]   4779  13967   8347    941  14988   3670  53991   9536  24653  92944  34289  31080
[25]  16946   3751   2664   7028   3084  41293  25914  33268  43682   7778 107777
```{r}
specieslist_mammal[16,1]
```
박쥐라서 중복 위치가 많았나보다. 조금 부족하지만 넘어간다.

map_int(dt.reptile, nrow)
[1]  6669 12821   108 23189 23335  1793 22379  4341  5187 10406  2785  5821  9006 16858
[15]  2942 21560     0  7326  5486  8903  9851  4744  5877  3296  2261  2261  4718  9328
[29]  1813  3283 10400   685  2015  1814  4511   163  3909  6618  3938  8696  3053  8894
[43]  3338  1833  9973  9654 12511  5102  8299  1983  1624  1613  2118  1933 13672  4554
[57]  9970 16386  5037 12509  7216  4268   547 19513  5406   673 19646  3999 15467  3222
[71]  7607  4441 39003  7222  5171   196  2506  9442  7113  1068 15816  2955  8097  6240
[85] 30793  7767 16408  3048 36902 21514  4356 10760  2363 14640  2277 24127
```{r}

specieslist_reptile[c(3,17,32,36,76),1]
```
백 단위가 몇종 보이긴 하지만 넘어간다. Cnemidophorus exsanguis가 0개다. 

```{r}
reptile.csv %>% 
  filter(taxonKey == 5227545)
```
Aspidoscelis exsanguis란다.. 
넣어주자..remove Ameiva too

```{r}
dt.reptile[[17]] <- reptile.csv %>% 
  filter(taxonKey == 5227545)

dt.reptile[[3]] <- NULL
```

map_int(dt.mammal, nrow)

 [1]   2009   3817 108759   1577  11105 116715   4269  76878   6348   2023  31150  27682
[13]   3643  12450   6676    860  14965   3276  50943   9536  24118  92891  34159  31080
[25]  16946   3690   2344   7025   3081  40339  25012  33204  43682   6629 106409

map_int(dt.reptile, nrow)

[1]  6669 12821 23189 23335  1793 22379  4341  5187 10406  2785  5821  9006 16858  2942
[15] 21560   350  7326  5486  8903  9851  4744  5877  3296  2261  2261  4718  9328  1813
[29]  3283 10400   685  2015  1814  4511   163  3909  6618  3938  8696  3053  8894  3338
[43]  1833  9973  9654 12511  5102  8299  1983  1624  1613  2118  1933 13672  4554  9970
[57] 16386  5037 12509  7216  4268   547 19513  5406   673 19646  3999 15467  3222  7607
[71]  4441 39003  7222  5171   196  2506  9442  7113  1068 15816  2955  8097  6240 30793
[85]  7767 16408  3048 36902 21514  4356 10760  2363 14640  2277 24127

이제 양서류를 해보자

map_int(dt.amphibian, nrow)
 [1]  3650  5869  2307  2440 26501  6934  3156  1998 46021 41418  2901  1672  1521   704  7879
[16]  3451  2725  3327  2564  8792  5121  1453 14405  7362 23759 28652 23035     0 10082 13624
[31]   456 12899  5778 10270 15027  3407   676   512  1688  8769  7528  2267 55643 16139 12288
[46]  1855  4162  5433 29245  1464

28번째 양서류 종이 0개가 나왔다. 
```{r}
specieslist_amphibian[28,1]
```

내가 요청했던건 리토리아 였고

```{r}
amphibian.csv %>% 
  group_by(species) %>%
  summarise(n = length(species))
```

받아온 자료에는 43번째에 Ranoidea aurea로 되어있었다. 종명이 개정된듯하다.
요청종 엑셀 목록을 수정한다.



24, 25번 종의 수가 겹친다. 엑셀에서 중복 열이 두개 있어서 그랬다.
같은 종이 두번 필터된 것이다. 하나를 제거한다.

포유류 35종, 파충류 94종. 양서류 50종. 503950
2,273,256개.

```{r}
#save to sp
library(sp)

sp.wgs84 <- function(x){
    SpatialPointsDataFrame(coords = x[,22:23], 
                  proj4string=CRS("+proj=longlat +datum=WGS84"), 
                  data = x[,1:52])
  }

sp_mammal <- list()
for (i in 1:length(dt.mammal)){
  sp_mammal[[i]] <- sp.wgs84(dt.mammal[[i]])
} 

sp_reptile <- list()
for (i in 1:length(dt.reptile)){
  sp_reptile[[i]] <- sp.wgs84(dt.reptile[[i]])
} 

sp_amphibian <- list()
for (i in 1:length(dt.amphibian)){
  sp_amphibian[[i]] <- sp.wgs84(dt.amphibian[[i]])
} 

#save to sf
library(sf)

sf_df <- function(x){sf::st_as_sf(x, 
                                  coords = c("decimalLongitude", "decimalLatitude"), 
                                  crs = 4326,
                                  remove = F)
}

sf_mammal <- map(dt.mammal, sf_df)
sf_reptile <- map(dt.reptile, sf_df)
sf_amphibian <- map(dt.amphibian, sf_df)

##cut geographical outliers - QGIS cut

sf_mammal.q<- list()
for (i in 1:length(sf_mammal)){
sf_mammal.q[[i]] <- sf_mammal[[i]][,53]
}

sf_reptile.q <- list()
for (i in 1:length(sf_reptile)){
sf_reptile.q[[i]] <- sf_reptile[[i]][,53]
}

sf_amphibian.q<- list()
for (i in 1:length(sf_amphibian)){
sf_amphibian.q[[i]] <- sf_amphibian[[i]][,53]
}

for (i in 1:length(sf_mammal)){
sf::write_sf(sf_mammal.q[[i]], paste0("shp/sf_mammal/",sf_mammal[[i]]$species[1],".shp"))
}

for (i in 1:length(sf_reptile.q)){
sf::write_sf(sf_reptile.q[[i]], paste0("shp/sf_reptile/",sf_reptile[[i]]$species[1],".shp"))
}

for (i in 1:length(sf_amphibian)){
sf::write_sf(sf_amphibian.q[[i]], paste0("shp/sf_amphibian/",sf_amphibian[[i]]$species[1],".shp"))
}

sf_mammal.q <- list()
for (i in 1:35){
sf_mammal.q[[i]] <- sf::st_read(dsn = paste0("shp/sf_mammal/",sf_mammal[[i]]$species[1],".shp"))
}

sf_reptile.q <- list()
for (i in 1:length(sf_reptile)){
sf_reptile.q[[i]] <- sf::st_read(dsn = paste0("shp/sf_reptile/",sf_reptile[[i]]$species[1],".shp"))
}

sf_amphibian.q <- list()
for (i in 1:length(sf_amphibian)){
sf_amphibian.q[[i]] <- sf::st_read(dsn = paste0("shp/sf_amphibian/",sf_amphibian[[i]]$species[1],".shp"))
}

# How much % points we cutted
( unlist(map(sf_mammal, nrow)) - unlist(map(sf_mammal.q, nrow)) ) / map_int(sf_mammal, nrow) *100
( unlist(map(sf_reptile, nrow)) - unlist(map(sf_reptile.q, nrow)) ) / map_int(sf_reptile, nrow) *100
( unlist(map(sf_amphibian, nrow)) - unlist(map(sf_amphibian.q, nrow)) ) / map_int(sf_amphibian, nrow) *100

# confirm ? to use sf_species.q ????
# Y? N?

```

```{r confirm}
sf_mammal <- sf_mammal.q
rm(sf_mammal.q)
rm(sf_mammal.)

sp_mammal <- map(sf_mammal.q, as_Spatial)

sf_reptile <- sf_reptile.q
rm(sf_reptile.q)
rm(sf_reptile.)

sp_reptile <- map(sf_reptile.q, as_Spatial)

sp_amphibian <- map(sf_amphibian.q, as_Spatial)



```

removed list from sf_reptile
17 Cnemidophorus exsanguis has zero record.
32 Emoia caeruleocauda 677
36 Gekko japonicus 162
63 Phyllodactylus tuberculosus 545 
66 Podarcis hispanicus 673
76 Sphaerodactylus macrolepis 196 

total 90 reptile species passed wrangling.




# CH.n Load bio-climate data
```{r bring data : bio climate}
library(raster)

biopath <- paste0("bios/", list.files(pattern="*.tif", path = "bios/"))
stackbio <- raster::stack(biopath)
bio1 <- raster::stack(biopath[1])
```



```{r data munging : random points for background}
library(dismo)

random_bg1x <- function(x,y){randomPoints(stackbio, n = length(x), p = x, ext = y, excludep = TRUE)}

# r is a raster, n is the number of points requested
fastRandomPoints <- function(r, n) {
  if(raster::nlayers(r) > 1) r <- r[[1]]
  v <- raster::getValues(r)
  v.notNA <- which(!is.na(v))
  x <- sample(v.notNA, n)
  pts <- raster::xyFromCell(r, x)
  return(pts)
}

bbox_mammal <- list()
for (i in 1:length(sf_mammal)){
  bbox_mammal[[i]] <- sf::st_bbox(sf_mammal.q[[i]])}

bbox_reptile <- list()
for (i in 1:length(sf_reptile)){
  bbox_reptile[[i]] <- sf::st_bbox(sf_reptile.q[[i]])}

bbox_amphibian <- list()
for (i in 1:length(sf_amphibian)){
  bbox_amphibian[[i]] <- sf::st_bbox(sf_amphibian.q[[i]])}

bbox_aves <- list()
for (i in 1:length(sf_aves)){
  bbox_aves[[i]] <- sf::st_bbox(sf_aves.q[[i]])}


library(furrr)
plan(multisession, workers = 50)
furrr::future_map2(a, b, random_bg1x, .progress = T)

a <- list()
a[[1]] <- sp_mammal[[1]]
b <- list()
b[[1]] <- bbox_mammal[[1]]

bg1x_mammal <- furrr::future_map2(sp_mammal, bbox_mammal, random_bg1x, .options = furrr_options(seed = TRUE))


bg1x_reptile <- furrr::future_map2(sp_reptile, bbox_reptile, random_bg1x, .options = furrr_options(seed = TRUE), .progress = T)

bg1x_amphibian <- furrr::future_map2(sp_amphibian, bbox_amphibian, random_bg1x, .options = furrr_options(seed = TRUE), .progress = T)


bg2x_amphibian <- furrr::future_map2(list_spatial_amphibian, bbox_amphibian, random_bg2x)
bg2x_aves <- furrr::future_map2(list_spatial_aves, bbox_aves, random_bg2x)
bg2x_mammal_pilot <- furrr::future_map2(sp_mammal_pilot, bbox_mammal_pilot, random_bg2x)

sf_bg <- function(x){sf::st_as_sf(x, coords = c("x", "y"), crs = 4326)}

bg1x_mammal. <- map(map(bg1x_mammal, as.data.frame), sf_bg)
bg1x_reptile. <- map(map(bg1x_reptile, as.data.frame), sf_bg)
bg1x_amphibian. <- map(map(bg1x_amphibian, as.data.frame), sf_bg)

for (i in 1:length(bg1x_mammal.)){
sf::write_sf(bg1x_mammal.[[i]], paste0("shp/sf_mammal/bg1x_mammal/bg1x_",sf_mammal[[i]]$species[1],".shp"))
}

for (i in 1:length(bg1x_reptile.)){
sf::write_sf(bg1x_reptile.[[i]], paste0("shp/sf_reptile/bg1x_reptile/bg1x_",sf_reptile[[i]]$species[1],".shp"))
}

for (i in 1:length(bg1x_amphibian.)){
sf::write_sf(bg1x_amphibian.[[i]], paste0("shp/sf_amphibian/bg1x_amphibian/bg1x_",sf_amphibian[[i]]$species[1],".shp"))
}



bg1x_reptile. <- map(map(bg1x_reptile, as.data.frame), sf_bg)

for (i in 1:length(bg1x_reptile.)){
sf::write_sf(bg1x_reptile.[[i]], paste0("shp/",i,".shp"))
}


```

```{r}
cores <- parallel::detectCores() - 1
cluster <- parallel::makeCluster(spec = cores)
doParallel::registerDoParallel(cl = cluster)

tic()
bg2x_mammal <- list()
foreach (i = 1:length(list_spatial_mammal), .packages='dismo') %dopar% {
  bg2x_mammal[[i]] <- random_bg2x(list_spatial_mammal[[i]], bbox_mammal[[i]])}
toc()

bg2x.pilot.df <- as.data.frame(bg2x_mammal_pilot[[1]])
bg2x.pilot.sf <- sf::st_as_sf(bg2x.pilot.df, coords = c("x", "y"), crs = 4326)
test2 <- sf_mammal[[1]]["geometry"]

sf::write_sf(bg2x.pilot.sf, "test.shp")
```



```{r d}
sf_bg <- function(x){sf::st_as_sf(x, coords = c("long", "lat"), crs = 4326)}
bg1x_mammal <- map(map(bg2x_mammal, as.data.frame), sf_bg)
bg1x_reptile <- map(map(bg2x_reptile, as.data.frame), sf_bg)
bg1x_amphibian <- map(map(bg2x_amphibian, as.data.frame), sf_bg)

for (i in 1:length(sf_mammal)){
  sf::write_sf(bg1x_mammal[[i]], paste0(sf_mammal[[i]]$species[1],"bg1x.shp"))
}

for (i in 1:length(sf_reptile)){
  sf::write_sf(bg1x_reptile[[i]], paste0(sf_reptile[[i]]$species[1],"_bg1x.shp"))
}

for (i in 1:length(sf_amphibian)){
  sf::write_sf(bg1x_amphibian[[i]], paste0(sf_amphibian[[i]]$species[1],"_bg1x.shp"))
}

# write shp
sfbg_amphibian <- map(map(sf_amphibian, as.data.frame), sf_bg)
for (i in 1:length(sfbg_amphibian)){
  sfbg_amphibian[[i]] <- sfbg_amphibian[[i]][,82] 
}

for (i in 1:length(sf_amphibian)){
  sf::write_sf(sfbg_amphibian[[i]], paste0(sf_amphibian[[i]]$species[1],"_sfd.shp"))
}


for (i in 1:length(sf_reptile)){
  sf::write_sf(sfbg_reptile[[i]], paste0("sf_reptile/shp/",sf_reptile[[i]]$species[1],"._sfd.shp"))
}

sfbg_reptile <- map(map(sf_reptile, as.data.frame), sf_bg)
for (i in 1:length(sfbg_reptile)){
  sfbg_reptile[[i]] <- sfbg_reptile[[i]][,82] 
}


```



# SDMtune:: 

멋진 건 줄 알고 비볐는데 진짜 원시적인 프로그램이다. ^^ 
데이터 세팅을 저장할수도 없고 다시 입력할수도 없다.

오로지 xy 좌표랑 래스터 스택만 가져오면 된다.


```{r}
bg1x_mammal <- paste0("C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_mammal/bg1x_mammal/", list.files(pattern="*.shp", path = "C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_mammal/bg1x_mammal"))

bg1x_reptile <- paste0("C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_reptile/bg1x_reptile/", list.files(pattern="*.shp", path = "C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_reptile/bg1x_reptile"))

bg1x_amphibian <- paste0("C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_amphibian/bg1x_amphibian/", list.files(pattern="*.shp", path = "C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_amphibian/bg1x_amphibian"))

bg1x_mammal <- map(bg1x_mammal, sf::st_read)
bg1x_reptile <- map(bg1x_reptile, sf::st_read)
bg1x_amphibian <- map(bg1x_amphibian, sf::st_read)  

for (i in 1:length(bg1x_mammal)){bg1x_mammal[[i]] <- sf::st_coordinates(bg1x_mammal[[i]])}
for (i in 1:length(bg1x_reptile)){bg1x_reptile[[i]] <- sf::st_coordinates(bg1x_reptile[[i]])}
for (i in 1:length(bg1x_amphibian)){bg1x_amphibian[[i]] <- sf::st_coordinates(bg1x_amphibian[[i]])}

```

```{r}
sfd_mammal <- list()
for (i in 1:length(sf_mammal.q)){
sfd_mammal[[i]] <- data.frame(st_coordinates(sf_mammal.q[[i]])) %>%
  mutate(species = sf_mammal[[i]][[10]][1])
}

sfd_reptile <- list()
for (i in 1:length(sf_reptile.q)){
sfd_reptile[[i]] <- data.frame(st_coordinates(sf_reptile.q[[i]])) %>%
  mutate(species = sf_reptile[[i]][[10]][1])
}


sfd_amphibian <- list()
for (i in 1:length(sf_amphibian.q)){
sfd_amphibian[[i]] <- data.frame(st_coordinates(sf_amphibian.q[[i]])) %>%
  mutate(species = specieslist_amphibian[[1]][i])
}

```







```{r model data : prepareSWD}

data.mammal_pilot <- list()
for (i in 1:length(sf_mammal_pilot)){data.mammal_pilot[[i]] <- SDMtune::prepareSWD(
  species = sf_mammal_pilot[[i]]$species[1], 
  p = sfd_mammal_pilot[[i]][,83:84], 
  a = bg2x_mammal_pilot[[i]], 
  env = stackbio)}

data.mammal <- list()
for (i in 1:length(sf_mammal)){data.mammal[[i]] <- SDMtune::prepareSWD(
  species = sf_mammal[[i]]$species[1], 
  p = sfd_mammal[[i]][,83:84], 
  a = bg1x_mammal[[i]], 
  env = stackbio)}

swdmaker <- function(x,y){SDMtune::prepareSWD(
  species = x$species[1], 
  p = x[,1:2], 
  a = y, 
  env = stackbio)}





data.mammal <- furrr::future_map2(sfd_mammal, bg1x_mammal, swdmaker, .options = furrr_options(seed = TRUE), .progress = T)
data.reptile <- furrr::future_map2(sfd_reptile, bg1x_reptile, swdmaker, .options = furrr_options(seed = TRUE), .progress = T)
data.amphibian <- furrr::future_map2(sfd_amphibian, bg1x_amphibian, swdmaker)


data.reptile <- list()
for (i in 1:length(sf_reptile)){data.reptile[[i]] <- SDMtune::prepareSWD(
  species = sf_reptile[[i]]$species[1], 
  p = sfd_reptile[[i]][,83:84], 
  a = bg1x_reptile[[i]], 
  env = stackbio)}

data.amphibian <- list()
for (i in 1:length(sf_amphibian.q)){data.amphibian[[i]] <- SDMtune::prepareSWD(
  species = sf_amphibian[[i]]$species[1], 
  p = sfd_amphibian[[i]][,83:84], 
  a = bg1x_amphibian[[i]], 
  env = stackbio)}



swd2csv(data, file_name = "data.csv")
```

# CH.n Save all data into rdata
```{r}

save(list_spatial_amphibian, file = "E:/sdmtune/rdata/list_spatial_amphibian.rdata")

save(list_xy)

save(sfd_mammal, file = "E:/sdmtune/rdata/sfd_mammal.rdata")
save(sfd_reptile, file = "E:/sdmtune/rdata/sfd_reptile.rdata")
save(sfd_amphibian, file = "E:/sdmtune/rdata/sfd_amphibian.rdata")


save(specieslist_mammal, file = "E:/sdmtune/rdata/specieslist_mammal.rdata")
save(specieslist_reptile, file = "E:/sdmtune/rdata/specieslist_reptile.rdata")
save(specieslist_amphibian, file = "E:/sdmtune/rdata/specieslist_amphibian.rdata")
save(specieslist_aves, file = "E:/sdmtune/rdata/specieslist_aves.rdata")

save(stackbio, file = "E:/sdmtune/rdata/stackbio.rdata")

save(t.t.mammal_0.2, file = "E:/sdmtune/rdata/t.t.mammal_0.2.rdata")
```

```{r}
save(bg1x_mammal, file = "E:/sdmtune/rdata/bg1x_mammal.rdata")
save(bg1x_reptile, file = "E:/sdmtune/rdata/bg1x_reptile.rdata")
save(bg1x_amphibian, file = "E:/sdmtune/rdata/bg1x_amphibian.rdata")

save(bbox_mammal, file = "E:/sdmtune/rdata/bbox_mammal.rdata")
save(bbox_reptile, file = "E:/sdmtune/rdata/bbox_reptile.rdata")
save(bbox_amphibian, file = "E:/sdmtune/rdata/bbox_amphibian.rdata")
save(bbox_aves, file = "E:/sdmtune/rdata/bbox_aves.rdata")

save(data_frame_mammal, file = "E:/sdmtune/rdata/data_frane_mammal.rdata")
save(data_frame_reptile, file = "E:/sdmtune/rdata/data_frame_reptile.rdata")
save(data_frame_amphibian, file = "E:/sdmtune/rdata/data_frame_amphibian.rdata")
save(data_frame_aves, file = "E:/sdmtune/rdata/data_frame_aves.rdata")


save(data.mammal, file = "E:/sdmtune/rdata/data.mammal.rdata")
save(data.reptile, file = "E:/sdmtune/rdata/data.reptile.rdata")
save(data.amphibian, file = "E:/sdmtune/rdata/data.amphibian.rdata")

save(list_longlat_mammal, file = "E:/sdmtune/rdata/list_longlat_mammal.rdata")
save(list_longlat_reptile, file = "E:/sdmtune/rdata/list_longlat_reptile.rdata")
save(list_longlat_amphibian, file = "E:/sdmtune/rdata/list_longlat_amphibian.rdata")
save(list_longlat_aves, file = "E:/sdmtune/rdata/list_longlat_aves.rdata")

save(list_spatial_mammal, file = "E:/sdmtune/rdata/list_spatial_mammal.rdata")
save(list_spatial_reptile, file = "E:/sdmtune/rdata/list_spatial_reptile.rdata")
save(list_spatial_amphibian, file = "E:/sdmtune/rdata/list_spatial_amphibian.rdata")
save(list_spatial_aves, file = "E:/sdmtune/rdata/list_spatial_aves.rdata")

save(list_xy_mammal, file = "E:/sdmtune/rdata/list_xy_mammal.rdata")
save(list_xy_reptile, file = "E:/sdmtune/rdata/list_xy_reptile.rdata")
save(list_xy_amphibian, file = "E:/sdmtune/rdata/list_xy_amphibian.rdata")
save(list_xy_aves, file = "E:/sdmtune/rdata/list_xy_aves.rdata")

save(sf_mammal, file = "E:/sdmtune/rdata/list_xy_mammal.rdata")
save(sf_reptile, file = "E:/sdmtune/rdata/list_xy_reptile.rdata")
save(sf_amphibian, file = "E:/sdmtune/rdata/list_xy_amphibian.rdata")
save(sf_aves, file = "E:/sdmtune/rdata/list_xy_aves.rdata")

save(sfd_mammal, file = "E:/sdmtune/rdata/list_xy_mammal.rdata")
save(sfd_reptile, file = "E:/sdmtune/rdata/list_xy_reptile.rdata")
save(sfd_amphibian, file = "E:/sdmtune/rdata/list_xy_amphibian.rdata")
save(sfd_aves, file = "E:/sdmtune/rdata/list_xy_aves.rdata")

save(specieslist_mammal, file = "E:/sdmtune/rdata/list_xy_mammal.rdata")
save(specieslist_reptile, file = "E:/sdmtune/rdata/list_xy_reptile.rdata")
save(specieslist_amphibian, file = "E:/sdmtune/rdata/list_xy_amphibian.rdata")
save(specieslist_aves, file = "E:/sdmtune/rdata/list_xy_aves.rdata")

save(stackbio, file = "E:/sdmtune/rdata/stackbio.rdata")

save(t.t.mammal_0.2, file = "E:/sdmtune/rdata/t.t.mammal_0.2.rdata")
```


```{r}
library(sf)
library(readr)
library(dplyr)
library(purrr)

data.mammal <- read_rds("data/data.mammal.rds")
data.reptile <- read_rds("data/data.reptile.rds")
data.amphibian <- read_rds("data/data.amphibian.rds")

name.mammal <- vector()
for(i in 1:length(data.mammal)){
 name.mammal[i] <- data.mammal[[i]]@species
}
name.reptile <- vector()
for(i in 1:length(data.reptile)){
 name.reptile[i] <- data.reptile[[i]]@species
}
name.amphibian <- vector()
for(i in 1:length(data.amphibian)){
 name.amphibian[i] <- data.amphibian[[i]]@species
}



ANURA <- sf::st_read(dsn = "IUCN_range/ANURA.shp")
CAUDATA <- sf::st_read(dsn = "IUCN_range/CAUDATA.shp")
CHAMELEONS <- sf::st_read(dsn = "IUCN_range/CHAMELEONS.shp")
CROCODILES_ALLIGATORS <- sf::st_read(dsn = "IUCN_range/CROCODILES_ALLIGATORS.shp")
GYMNOPHIONA <- sf::st_read(dsn = "IUCN_range/GYMNOPHIONA.shp")
MAMMALS_TERRESTRIAL <- sf::st_read(dsn = "IUCN_range/MAMMALS_TERRESTRIAL_ONLY.shp")
REPTILES <- sf::st_read(dsn = "IUCN_range/REPTILES.shp")


range.ANURA <- list()
for(i in 1:length(name.amphibian)){
range.ANURA[[i]] <- ANURA %>% dplyr::filter(binomial == name.amphibian[i])
}

range.CAUDATA <- list()
for(i in 1:length(name.amphibian)){
range.CAUDATA[[i]] <- CAUDATA %>% dplyr::filter(binomial == name.amphibian[i])
}

range.CHAMELONES <- list()
for(i in 1:length(name.reptile)){
range.CHAMELONES[[i]] <- CHAMELEONS %>% dplyr::filter(binomial == name.reptile[i])
}

range.CROCODILES_ALLIGATORS <- list()
for(i in 1:length(name.reptile)){
range.CROCODILES_ALLIGATORS[[i]] <- CROCODILES_ALLIGATORS %>% dplyr::filter(binomial == name.reptile[i])
}

range.GYMNOPHIONA <- list()
for(i in 1:length(name.amphibian)){
range.GYMNOPHIONA[[i]] <- GYMNOPHIONA %>% dplyr::filter(binomial == name.amphibian[i])
}

range.MAMMALS_TERRESTRIAL <- list()
for(i in 1:length(name.mammal)){
range.MAMMALS_TERRESTRIAL[[i]] <- MAMMALS_TERRESTRIAL %>% dplyr::filter(binomial == name.mammal[i])
}

range.REPTILES <- list()
for(i in 1:length(name.reptile)){
range.REPTILES[[i]] <- REPTILES %>% dplyr::filter(binomial == name.reptile[i])
}

map_int(range.ANURA, nrow)
map_int(range.CAUDATA, nrow)
map_int(range.GYMNOPHIONA, nrow)

map_int(range.CHAMELONES, nrow)
map_int(range.CROCODILES_ALLIGATORS, nrow)
map_int(range.REPTILES, nrow)

map_int(range.MAMMALS_TERRESTRIAL, nrow)

range.amphibian.check <- data.frame(ANURA = map_int(range.ANURA, nrow),
                              CAUDATA = map_int(range.CAUDATA, nrow),
                              GYMNOPHIONA = map_int(range.GYMNOPHIONA, nrow))

range.reptile.check <- data.frame(CHAMELONES = map_int(range.CHAMELONES, nrow),
                              CROCODILES_ALLIGATORS = map_int(range.CROCODILES_ALLIGATORS, nrow),
                              REPTILES = map_int(range.REPTILES, nrow))
r#all overlap. same thing. 
range.reptile <- range.REPTILES

#match. append
range.amphibian <- range.ANURA

for(i in which(range.amphibian.check$ANURA == 0)){
  range.amphibian[[i]] <- range.CAUDATA[[i]]
}

map_dbl(range.amphibian, nrow)

#ONLY 43. species dont have range

for(i in 1:length(data.mammal)){
sf::write_sf(range.MAMMALS_TERRESTRIAL[[i]], paste0("shp/range/range_",name.mammal[i],".shp"))
}

for(i in 1:length(data.reptile)){
sf::write_sf(range.reptile[[i]], paste0("shp/range_reptile/range_",name.reptile[i],".shp"))
}

for(i in 1:length(data.amphibian)){
sf::write_sf(range.amphibian[[i]], paste0("shp/range_amphibian/range_",name.amphibian[i],".shp"))
}

```

```{r}
which(map_dbl(range.MAMMALS_TERRESTRIAL, nrow) == 0) # 20

which(map_dbl(range.reptile, nrow) == 0) #  3  5  7 11 13 15 17 30 32 35 48 52 55 56 68 75 79 87 92
which(map_dbl(range.amphibian, nrow) == 0) # 43
```

```{r combine features}

range.mammal <- list()
for(i in 1:length(range.MAMMALS_TERRESTRIAL)){
  range.mammal[[i]] <- st_combine(range.MAMMALS_TERRESTRIAL[[i]])
}

for(i in 1:length(range.reptile)){
  range.reptile[[i]] <- st_combine(range.reptile[[i]])
}

for(i in 1:length(range.amphibian)){
  range.amphibian[[i]] <- st_combine(range.amphibian[[i]])
}



```
write range polygon that is combined, at range.mix_taxa
```{r}
for(i in 1:length(data.mammal)){
sf::write_sf(range.mammal[[i]], paste0("shp/range.mix_mammal/range_",name.mammal[i],".shp"))
}

for(i in 1:length(data.reptile)){
sf::write_sf(range.reptile[[i]], paste0("shp/range_reptile/range_",name.reptile[i],".shp"))
}

for(i in 1:length(data.amphibian)){
sf::write_sf(range.amphibian[[i]], paste0("shp/range_amphibian/range_",name.amphibian[i],".shp"))
}

```


```{r}
my.bg10k <- function(x){st_sample(x, size = 10000, 
                                  type = "random", 
                                  exact = T, 
                                  by_polygon = F 
                                  )}

my.bg30k <- function(x){st_sample(x, size = 30000, 
                                  type = "random", 
                                  exact = T, 
                                  by_polygon = F 
                                  )}

my.bg100k <- function(x){st_sample(x, size = 100000, 
                                  type = "random", 
                                  exact = T, 
                                  by_polygon = F 
)}



cropbio.path_reptile <- list.files(path = "D:/bio_crop/reptile")
cropbio_reptile <- list()
for(i in 1:length(cropbio.path_reptile)){
cropbio_reptile[[i]] <- raster::stack(paste0("D:/bio_crop/reptile/",cropbio.path_reptile[i]))}

my.random_bg30k <- function(x,y){randomPoints(stackbio, n = length(x), p = x, ext = y, excludep = TRUE)}



bg10k_mammal <- list()
for(i in 21:length(range.mammal)){
  bg10k_mammal[[i]] <- my.bg10k(range.mammal[[i]])}

#14 mustela got error so done in QGIS
#35 vulpes got error so done in QGIS
#20 dont have IUCN range
#Import 14 QGIS bg10k

bg10k_mammal[[14]] <- sf::st_read("shp/sf_mammal/bg10k_mammal/m14_mustela.shp")
bg10k_mammal[[35]] <- sf::st_read("shp/sf_mammal/bg10k_mammal/m35_vulpes.shp")
bg10k_mammal[[20]] <- sf::st_read("shp/sf_mammal/bg10k_mammal/m20_Ornithorhynchus.shp")

#convert to sf to extract from raster
for(i in 21:35){
  bg10k_mammal[[i]] <- sf::st_sf(bg10k_mammal[[i]])
}

```

20 have no IUCN list, using MCP
```{r}
library(adehabitatHR)

my.pseperate <- function(x){dplyr::bind_cols(dplyr::slice(x@coords, 1:which(x@pa==0)[1]-1),
                                             dplyr::slice(x@data, 1:which(x@pa==0)[1]-1))}

p.data.mammal <- map(data.mammal, my.pseperate)
p.data.reptile <- map(data.reptile, my.pseperate)
p.data.amphibian <- map(data.amphibian, my.pseperate)

sf_df <- function(x){sf::st_as_sf(x, 
                                  coords = c("X", "Y"), 
                                  crs = 4326,
                                  remove = F)
}

sf_p_mammal <- map(p.data.mammal, sf_df)
sp_p_mammal <- map(sf_p_mammal, as_Spatial)

sf_p_reptile <- map(p.data.reptile, sf_df)
sp_p_reptile <- map(sf_p_reptile, as_Spatial)

sf_p_amphibian <- map(p.data.amphibian, sf_df)
sp_p_amphibian <- map(sf_p_amphibian, as_Spatial)

mcp_20 <- mcp(sp_p_mammal[[20]])
sf_mcp_20 <- sf::st_as_sf(mcp_20)
range.mammal[[20]] <- sf_mcp_20
bg10k_mammal[[20]] <- my.bg10k(range.mammal[[20]])
bg10k_mammal[[20]] <- sf::st_sf(bg10k_mammal[[20]])

rgdal::writeOGR(mcp_20, "mcp_20.13k.shp", driver = "ESRI Shapefile", layer = "mcp_20.13k")


sp_bg10k_mammal <- map(bg10k_mammal, as_Spatial)
```



```{r}
my.extract <- function(x){raster::extract(stackbio, x)}

bg10k_mammal.bio <- future_map(sp_bg10k_mammal, my.extract)


```

```{r}
my.nacheck <- function(x){sum(is.na(x[,1]))}

map_dbl(bg10k_mammal.bio, my.nacheck)
map_dbl(bg10k_amphibian.bio, my.nacheck)
```
```{r}
for (i in 1:length(bg10k_mammal)){
sf::write_sf(bg10k_mammal[[i]], paste0("shp/sf_mammal/bg10k_mammal/",data.mammal[[i]]@species,".shp"))
}
```


reptile no range in 3  5  7 11 13 15 17 30 32 35 48 52 55 56 68 75 79 87 92
make range for them with MCP
```{r}
for(i in c(3, 5, 7, 11, 13, 15, 17, 30, 32, 35, 48, 52, 55, 56, 68, 75, 79, 87, 92)){
  range.reptile[[i]] <- sf::st_as_sfc(mcp(sp_p_reptile[[i]]))
}

range.amphibian[[43]] <- sf::st_as_sfc(mcp(sp_p_amphibian[[43]]))

```


```{r}
bg10k_reptile <- list()
for(i in 37:length(range.reptile)){
  bg10k_reptile[[i]] <- my.bg10k(range.reptile[[i]])}

bg30k_reptile <- list()
for(i in 1:length(range.reptile)){
  bg30k_reptile[[i]] <- my.bg30k(range.reptile[[i]])}

bg100k_reptile <- furrr::future_map(range.reptile, my.bg30k, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)


#12 is SEA TURTLE
#36 got error so done in QGIS
bg10k_reptile[[36]] <- sf::st_read("shp/sf_reptile/bg10k_reptile/bg10k_hemidactylus frenatus.shp")

#conver sf to extract
for(i in 13:94){
  bg10k_reptile[[i]] <- sf::st_sf(bg10k_reptile[[i]])
}

# - 12
bg10k_reptile[[12]] <- NULL

sp_bg10k_reptile <- map(bg10k_reptile, as_Spatial)

bg10k_reptile.bio <- future_map(sp_bg10k_reptile, my.extract)
```

```{r}
bg10k_amphibian <- furrr::future_map(range.amphibian, my.bg10k, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

sp_bg10k_amphibian <- map(bg10k_amphibian, as_Spatial)

bg10k_amphibian.bio <- furrr::future_map(sp_bg10k_amphibian, my.extract)
```


```{r}
map_dbl(bg10k_reptile.bio, my.nacheck)
```

```{r}

cropbio.path_mammal <- list.files(path = "D:/bio_crop/mammal")
cropbio_mammal <- list()
for(i in 1:length(cropbio.path_mammal)){
cropbio_mammal[[i]] <- raster::stack(paste0("D:/bio_crop/mammal/",cropbio.path_mammal[i]))}

cropbiodf.path <- vector()
for(i in 1:35){
cropbiodf.path[i] <- paste0("D:/bio_crop/mammal/",data.mammal[[i]]@species,".csv")
}

for(i in 1:length(cropbio_mammal)){
write_csv(raster::as.data.frame(cropbio_mammal[[i]]), cropbiodf.path[i])  
}
```

```{r}
bg10k_mammal.bio <- readr::read_rds("bg10k_mammal.bio.rds")
biocolnames <- c("wc2.1_30s_bio_01",
              "wc2.1_30s_bio_02",
              "wc2.1_30s_bio_03",
              "wc2.1_30s_bio_04",
              "wc2.1_30s_bio_05",
              "wc2.1_30s_bio_06",
              "wc2.1_30s_bio_07",
              "wc2.1_30s_bio_08",
              "wc2.1_30s_bio_09",
              "wc2.1_30s_bio_10",
              "wc2.1_30s_bio_11",
              "wc2.1_30s_bio_12",
              "wc2.1_30s_bio_13",
              "wc2.1_30s_bio_14",
              "wc2.1_30s_bio_15",
              "wc2.1_30s_bio_16",
              "wc2.1_30s_bio_17",
              "wc2.1_30s_bio_18",
              "wc2.1_30s_bio_19")

b.data.mammal <- list()
for(i in 1:length(bg10k_mammal.bio)){
  b.data.mammal[[i]] <- data.frame(wc2.1_30s_bio_01 = bg10k_mammal.bio[[i]][,1],
                                   wc2.1_30s_bio_02 = bg10k_mammal.bio[[i]][,2],
                                   wc2.1_30s_bio_03 = bg10k_mammal.bio[[i]][,3],
                                   wc2.1_30s_bio_04 = bg10k_mammal.bio[[i]][,4],
                                   wc2.1_30s_bio_05 = bg10k_mammal.bio[[i]][,5],
                                   wc2.1_30s_bio_06 = bg10k_mammal.bio[[i]][,6],
                                   wc2.1_30s_bio_07 = bg10k_mammal.bio[[i]][,7],
                                   wc2.1_30s_bio_08 = bg10k_mammal.bio[[i]][,8],
                                   wc2.1_30s_bio_09 = bg10k_mammal.bio[[i]][,9],
                                   wc2.1_30s_bio_10 = bg10k_mammal.bio[[i]][,10],
                                   wc2.1_30s_bio_11 = bg10k_mammal.bio[[i]][,11],
                                   wc2.1_30s_bio_12 = bg10k_mammal.bio[[i]][,12],
                                   wc2.1_30s_bio_13 = bg10k_mammal.bio[[i]][,13],
                                   wc2.1_30s_bio_14 = bg10k_mammal.bio[[i]][,14],
                                   wc2.1_30s_bio_15 = bg10k_mammal.bio[[i]][,15],
                                   wc2.1_30s_bio_16 = bg10k_mammal.bio[[i]][,16],
                                   wc2.1_30s_bio_17 = bg10k_mammal.bio[[i]][,17],
                                   wc2.1_30s_bio_18 = bg10k_mammal.bio[[i]][,18],
                                   wc2.1_30s_bio_19 = bg10k_mammal.bio[[i]][,19],
                                   species = data.mammal[[i]]@species,
                                   type = "background")}
for(i in 1:35){
b.data.mammal[[i]]  <- tidyr::pivot_longer(b.data.mammal[[i]], 
                                            cols = 1:19, 
                                            names_to = "bio", 
                                            values_to = "covari")}

  readr::read_csv(cropbiodf.path[35], 
                  col_names = biocolnames,
                  skip = 1,
                  col_types = cols(wc2.1_30s_bio_01  = "d",
                                   wc2.1_30s_bio_02  = "d",
                                   wc2.1_30s_bio_03  = "d",
                                   wc2.1_30s_bio_04  = "d",
                                   wc2.1_30s_bio_05  = "d",
                                   wc2.1_30s_bio_06  = "d",
                                   wc2.1_30s_bio_07  = "d",
                                   wc2.1_30s_bio_08  = "d",
                                   wc2.1_30s_bio_09  = "d",
                                   wc2.1_30s_bio_10  = "d",
                                   wc2.1_30s_bio_11  = "d",
                                   wc2.1_30s_bio_12  = "d",
                                   wc2.1_30s_bio_13  = "d",
                                   wc2.1_30s_bio_14  = "d",
                                   wc2.1_30s_bio_15  = "d",
                                   wc2.1_30s_bio_16  = "d",
                                   wc2.1_30s_bio_17  = "d",
                                   wc2.1_30s_bio_18  = "d",
                                   wc2.1_30s_bio_19  = "d" ))  %>% 
    mutate(species = data.mammal[[35]]@species,
           type = "range") %>% 
  tidyr::pivot_longer( 
                      cols = 1:19, 
                      names_to = "bio", 
                      values_to = "covari") %>%
  dplyr::bind_rows(b.data.mammal[[i]]) %>%
  write_csv(., paste0("D:/viz_rb/mammal/",data.mammal[[35]]@species,".csv"))
  
}

```

```{r}
data.mammal <- readr::read_rds("data/newdata/data.mammal.rds")
data.reptile <- readr::read_rds("data/newdata/data.reptile.rds")
data.amphibian <- readr::read_rds("data/newdata/data.amphibian.rds")
```


```{r}
library(geosphere)

my.aseperate <- function(x){dplyr::bind_cols(dplyr::slice(x@coords, which(x@pa==0)[1]:nrow(x@coords)),
                                             dplyr::slice(x@data, which(x@pa==0)[1]:nrow(x@data)))}

a.data.mammal <- map(data.mammal, my.aseperate)
a.data.reptile <- map(data.reptile, my.aseperate)
a.data.amphibian <- map(data.amphibian, my.aseperate)

geodistm <- function(x){as.dist(geosphere::distm(x[,1:2], fun = distGeo))}

a.geodistm.mammal <- furrr::future_map(a.data.mammal, geodistm)
a.geodistm.reptile <- furrr::future_map(a.data.reptile, geodistm)
a.geodistm.amphibian <- furrr::future_map(a.data.amphibian, geodistm)


```


make table and histogram
```{r}
library(tidyverse)
library(hrbrthemes)
library(viridis)
library(forcats)


for(i in 1:length(biovalue_mammal)){
biovalue_mammal[[i]] %>% mutate(species = ,
                                type = "range") -> biovalue_mammal[[i]]
}


for(i in 1:length(bg10k_mammal.bio)){   
bg10k_mammal.bio[[i]] %>% mutate(species = ,
                                 type = "background") -> bg10k_mammal.bio[[i]]
}

# Load dataset from github
data <- data.frame(species = ,
                   type = c("range", "background"),
                   bio1 = ,
                   bio2 = ,
                   bio3 = ,
                   bio4 = ,
                   bio5 = ,
                   bio6 = ,
                   bio7 = ,
                   bio8 = ,
                   bio9 = , 
                   bio10 = ,
                   bio11 = ,
                   bio12 = ,
                   bio13 = ,
                   bio14 = ,
                   bio15 = ,
                   bio16 = ,
                   bio17 = ,
                   bio18 = ,
                   bio19 = )

# plot
p <- data %>%
  mutate(text = fct_reorder(text, value)) %>%
  ggplot( aes(x=value, color=text, fill=text)) +
    geom_histogram(alpha=0.6, binwidth = 5) +
    scale_fill_viridis(discrete=TRUE) +
    scale_color_viridis(discrete=TRUE) +
    geom_density(col=3) +
    theme_ipsum() +
    theme(
      legend.position="none",
      panel.spacing = unit(0.1, "lines"),
      strip.text.x = element_text(size = 8)
    ) +
    xlab("") +
    ylab("Assigned Probability (%)") +
    facet_wrap(~text)


ggplot2.histogram(data=weight, xName='weight',
        groupName='sex', legendPosition="top",
        alpha=0.5, position="stack")
```




KS test 
```{r}
ks_mammal <- vector()
for(i in 1:length(biovalue_mammal)){
ks_mammal[i] <- ks.test(biovalue_mammal[[i]], na.omit(bg10k_mammal.bio[[i]][,1]))$p.value
}

testbg30 <- sf::read_sf("testbg30.shp")
testbg30 <- my.bg100k(range.mammal[[3]])
sp.testbg30 <- as_Spatial(testbg30)
testbg30.bio <- my.extract(sp.testbg30)
ks.test(biovalue_mammal[[3]], na.omit(testbg30.bio))

ks.boot(biovalue_mammal[[3]], na.omit(testbg30.bio), nboots = 10000)



ks.test(m1df$bio1, b)

j <- geodistm(m1dfcd)
```


```{r}
sample1 <- biovalue_mammal[[3]]
sample2 <- na.omit(testbg30.bio)
group <- c(rep("sample1", length(sample1)), rep("sample2", length(sample2)))
dat <- data.frame(KSD = c(sample1,sample2), group = group)
# create ECDF of data
cdf1 <- ecdf(biovalue_mammal[[3]])
cdf2 <- ecdf(na.omit(testbg30.bio)) 
# find min and max statistics to draw line between points of greatest distance
minMax <- seq(min(sample1, sample2), max(sample1, sample2), length.out=length(sample1)) 
x0 <- minMax[which( abs(cdf1(minMax) - cdf2(minMax)) == max(abs(cdf1(minMax) - cdf2(minMax))) )] 
y0 <- cdf1(x0) 
y1 <- cdf2(x0) 
```



```{r}
ggplot(dat, aes(x = KSD, group = group, color = group))+
  stat_ecdf(size=1) +
    theme_bw(base_size = 28) +
    theme(legend.position ="top") +
    xlab("Sample") +
    ylab("ECDF") +
    #geom_line(size=1) +
    geom_segment(aes(x = x0[1], y = y0[1], xend = x0[1], yend = y1[1]),
        linetype = "dashed", color = "red") +
    geom_point(aes(x = x0[1] , y= y0[1]), color="red", size=8) +
    geom_point(aes(x = x0[1] , y= y1[1]), color="red", size=8) +
    ggtitle("K-S Test: Sample 1 / Sample 2") +
    theme(legend.title=element_blank())
```
```{r}
cucconi.test <- function(x, y, method = c("permutation", "bootstrap")){

  # Implementation of the Cucconi test for the two-sample location-scale problem
  # A permutation/bootstrap distribution of the test statistic (C) under the
  # null hypothesis is used to calculate the p-value.
  # Reference: Marozzi (2013), p. 1302-1303

  m <- length(x)
  n <- length(y)
  C <- cucconi.teststat(x = x, y = y, m = m, n = n)
  
  if(method[1] == "permutation"){
    h0dist <- cucconi.dist.perm(x = x, y = y)
  }
  
  if(method[1] == "bootstrap"){
    h0dist <- cucconi.dist.boot(x = x, y = y)
  }
  
  p.value <- length(h0dist[h0dist >= C]) / length(h0dist)
  
  cat("\nCucconi two-sample location-scale test\n")
  cat("\nNull hypothesis: The locations and scales of the two population distributions are equal.\n")
  cat("Alternative hypothesis: The locations and/or scales of the two population distributions differ.\n")
  cat(paste("\nC = ", round(C, 3), ", p-value = ", round(p.value, 4), "\n\n", sep=""))
  
  return(list(C = C,
              method = method[1],
              p.value = p.value))
}
```


# BG corrupted. NEW DATA. NEW MODEL.
```{r}
data.mammal <- readr::read_rds("data/data.mammal.rds")
data.reptile <- readr::read_rds("data/data.reptile.rds") #reptile 12번 거북이가 실려있다. 후에 뺀다.
data.amphibian <- readr::read_rds("data/data.amphibian.rds")

```


```{r}
reaarange <- function(x){dplyr::select(x@data, c(1,12,13,14,15,16,17,18,19,2,3,4,5,6,7,8,9,10,11))}

for (i in 1:length(data.mammal)){
data.mammal[[i]]@data <- reaarange(data.mammal[[i]])
}

for (i in 1:length(data.reptile)){
data.reptile[[i]]@data <- reaarange(data.reptile[[i]])
}

for (i in 1:length(data.amphibian)){
data.amphibian[[i]]@data <- reaarange(data.amphibian[[i]])
}

for (i in 1:length(data.mammal)){
names(data.mammal[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                  "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                  "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                  "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                  "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}

for (i in 1:length(data.reptile)){
names(data.reptile[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                   "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                   "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                   "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                   "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}

for (i in 1:length(data.amphibian)){
names(data.amphibian[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                     "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                     "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                     "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                     "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}
```


```{r}
bg10kshp_mammal <- paste0("C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_mammal/bg10k_mammal/",
                          list.files(path = "C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_mammal/bg10k_mammal",
              pattern = "*.shp"))

bg10k_mammal <- list()
for(i in 1:length(data.mammal)){
bg10k_mammal[[i]] <-  sf::read_sf(bg10kshp_mammal[i])
}

bg10kcoord_mammal <- map(bg10k_mammal, st_coordinates)
bg10kcoord_mammal <- map(bg10kcoord_mammal, as.data.frame)
```

```{r}

# bg10k_reptile sf로 살아있다!

bg10kcoord_reptile <- map(bg10k_reptile, st_coordinates)
bg10kcoord_reptile <- map(bg10kcoord_reptile, as.data.frame)
```



load new background
```{r}
bg10k_mammal.bio <- readr::read_rds("bg10k_mammal.bio.rds")
bg10k_mammal.bio <- map(bg10k_mammal.bio, as.data.frame)
for(i in 1:length(bg10k_mammal.bio)){
  mutate(bg10k_mammal.bio[[i]],
         X = bg10kcoord_mammal[[i]]$X,
         Y = bg10kcoord_mammal[[i]]$Y) -> bg10k_mammal.bio[[i]]
}
bg10k_mammal.bio <- map(bg10k_mammal.bio, na.omit)
```

```{r}

bg10k_reptile.bio <- map(bg10k_reptile.bio, as.data.frame)

for(i in 1:length(bg10k_reptile.bio)){
  mutate(bg10k_reptile.bio[[i]],
         X = bg10kcoord_reptile[[i]]$X,
         Y = bg10kcoord_reptile[[i]]$Y) -> bg10k_reptile.bio[[i]]
}

bg10k_reptile.bio <- map(bg10k_reptile.bio, na.omit)
```

```{r}
bg10kcoord_amphibian <- map(bg10k_amphibian, st_coordinates)
bg10kcoord_amphibian <- map(bg10kcoord_amphibian, as.data.frame)

bg10k_amphibian.bio <- map(bg10k_amphibian.bio, as.data.frame)

for(i in 1:length(bg10k_amphibian.bio)){
  mutate(bg10k_amphibian.bio[[i]],
         X = bg10kcoord_amphibian[[i]]$X,
         Y = bg10kcoord_amphibian[[i]]$Y) -> bg10k_amphibian.bio[[i]]
}

bg10k_amphibian.bio <- map(bg10k_amphibian.bio, na.omit)
```



```{r}
my.pseperate <- function(x){dplyr::bind_cols(dplyr::slice(x@coords, 1:which(x@pa==0)[1]-1),
                                             dplyr::slice(x@data, 1:which(x@pa==0)[1]-1))}


p.data.mammal <- map(data.mammal, my.pseperate)

bg10kdata_mammal <- list()
for(i in 1:length(p.data.mammal)){
  bg10kdata_mammal[[i]] <- as.data.frame(
   rbind(p.data.mammal[[i]][,3:21], bg10k_mammal.bio[[i]][,1:19]))}

newcoord_mammal <- list()
for(i in 1:length(p.data.mammal)){
  newcoord_mammal[[i]] <-
   rbind(p.data.mammal[[i]][,1:2], bg10k_mammal.bio[[i]][,20:21])}

newpa_mammal <- list()
for(i in 1:length(p.data.mammal)){
newpa_mammal[[i]] <- c(rep(1, nrow(p.data.mammal[[i]])), 
                       rep(0, nrow(bg10k_mammal.bio[[i]])))}


for(i in 1:length(data.mammal)){
data.mammal[[i]]@data <- bg10kdata_mammal[[i]]
data.mammal[[i]]@coords <- newcoord_mammal[[i]]
data.mammal[[i]]@pa <- newpa_mammal[[i]]
}


```

```{r}
my.pseperate <- function(x){dplyr::bind_cols(dplyr::slice(x@coords, 1:which(x@pa==0)[1]-1),
                                             dplyr::slice(x@data, 1:which(x@pa==0)[1]-1))}

data.reptile[[12]] <- NULL

p.data.reptile <- map(data.reptile, my.pseperate)

bg10kdata_reptile <- list()
for(i in 1:length(p.data.reptile)){
  bg10kdata_reptile[[i]] <- as.data.frame(
   rbind(p.data.reptile[[i]][,3:21], bg10k_reptile.bio[[i]][,1:19]))}

newcoord_reptile <- list()
for(i in 1:length(p.data.reptile)){
  newcoord_reptile[[i]] <-
   rbind(p.data.reptile[[i]][,1:2], bg10k_reptile.bio[[i]][,20:21])}

newpa_reptile <- list()
for(i in 1:length(p.data.reptile)){
newpa_reptile[[i]] <- c(rep(1, nrow(p.data.reptile[[i]])), 
                       rep(0, nrow(bg10k_reptile.bio[[i]])))}


for(i in 1:length(data.reptile)){
data.reptile[[i]]@data <- bg10kdata_reptile[[i]]
data.reptile[[i]]@coords <- newcoord_reptile[[i]]
data.reptile[[i]]@pa <- newpa_reptile[[i]]
}

```

```{r}
my.pseperate <- function(x){dplyr::bind_cols(dplyr::slice(x@coords, 1:which(x@pa==0)[1]-1),
                                             dplyr::slice(x@data, 1:which(x@pa==0)[1]-1))}



p.data.amphibian <- map(data.amphibian, my.pseperate)

bg10kdata_amphibian <- list()
for(i in 1:length(p.data.amphibian)){
  bg10kdata_amphibian[[i]] <- as.data.frame(
   rbind(p.data.amphibian[[i]][,3:21], bg10k_amphibian.bio[[i]][,1:19]))}

newcoord_amphibian <- list()
for(i in 1:length(p.data.amphibian)){
  newcoord_amphibian[[i]] <-
   rbind(p.data.amphibian[[i]][,1:2], bg10k_amphibian.bio[[i]][,20:21])}

newpa_amphibian <- list()
for(i in 1:length(p.data.amphibian)){
newpa_amphibian[[i]] <- c(rep(1, nrow(p.data.amphibian[[i]])), 
                       rep(0, nrow(bg10k_amphibian.bio[[i]])))}


for(i in 1:length(data.amphibian)){
data.amphibian[[i]]@data <- bg10kdata_amphibian[[i]]
data.amphibian[[i]]@coords <- newcoord_amphibian[[i]]
data.amphibian[[i]]@pa <- newpa_amphibian[[i]]
}

```



```{r}
a <- vector()
for(i in 1:length(data.reptile)){
a[i] <- sum(data.reptile[[i]]@pa == 0)
}

a
```
```{r}
a <- vector()
for(i in 1:length(data.reptile)){
a[i] <- sum(data.reptile[[i]]@pa == 0)
}

a
```



```{r}
bg10k_mammal.xy <- list()
for(i in 1:length(bg10k_mammal.bio)){
bg10k_mammal.xy[[i]] <- bg10k_mammal.bio[[i]][,20:21]}

my.swdmaker <- function(x,y,z){SDMtune::prepareSWD(
  species = x@species, 
  p = y[,1:2], 
  a = z[,1:2], 
  env = stackbio)}

data.mammal <- furrr::future_map2(data.mammal, p.data.mammal, bg10k_mammal.xy,
                                  my.swdmaker, 
                                  .options = furrr_options(seed = TRUE), .progress = T)


```
