==> 2_modelling.Rmd <==
---
title: "2_modelling"
author: "JEON"
date: '2021 3 9 '
output: html_document
---

# Aims

After data wrangling, move on to modelling process.
```{r}
library(tidyverse)
library(SDMtune)
library(blockCV)
library(ENMeval)
library(raster)
library(sf)
library(sp)
library(zeallot)
library(rJava)
library(dismo)
```

```{r parallel computing}
library(parallel) #parallel computing
library(foreach)
library(doParallel)
library(doSNOW)
library(furrr)

cores <- parallel::detectCores() - 1
cluster <- parallel::makeCluster(spec = cores)
doParallel::registerDoParallel(cl = cluster)
future::plan(multisession, workers = 32)
stopCluster(cl)
future:::ClusterRegistry("stop")
```

#35, 93, 50 species
```{r}
data.mammal <- readr::read_rds("data/newdata/data.mammal.rds")
data.reptile <- readr::read_rds("data/newdata/data.reptile.rds")
data.amphibian <- readr::read_rds("data/newdata/data.amphibian.rds")

```

not only mydas sea turtle, lugubris also out..
```{r}
data.reptile[[47]] <- NULL
t.t.reptile_0.2[[47]] <- NULL
```
total 92 reptile species



```{r}
reaarange <- function(x){dplyr::select(x@data, c(1,12,13,14,15,16,17,18,19,2,3,4,5,6,7,8,9,10,11))}

for (i in 1:length(data.mammal)){
data.mammal[[i]]@data <- reaarange(data.mammal[[i]])
}

for (i in 1:length(data.reptile)){
data.reptile[[i]]@data <- reaarange(data.reptile[[i]])
}

for (i in 1:length(data.amphibian)){
data.amphibian[[i]]@data <- reaarange(data.amphibian[[i]])
}

for (i in 1:length(data.mammal)){
names(data.mammal[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                  "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                  "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                  "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                  "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}

for (i in 1:length(data.reptile)){
names(data.reptile[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                   "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                   "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                   "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                   "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}

for (i in 1:length(data.amphibian)){
names(data.amphibian[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                     "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                     "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                     "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                     "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}
```


```{r}
bg10kshp_mammal <- paste0("C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_mammal/bg10k_mammal/",
                          list.files(path = "C:/Users/lophital/Documents/rsac_x_overfit/shp/sf_mammal/bg10k_mammal",
              pattern = "*.shp"))

bg10k_mammal <- list()
for(i in 1:length(data.mammal)){
bg10k_mammal[[i]] <-  sf::read_sf(bg10kshp_mammal[i])
}

bg10kcoord_mammal <- map(bg10k_mammal, st_coordinates)
bg10kcoord_mammal <- map(bg10kcoord_mammal, as.data.frame)
```


load new background
```{r}
bg10k_mammal.bio <- readr::read_rds("bg10k_mammal.bio.rds")
bg10k_mammal.bio <- map(bg10k_mammal.bio, as.data.frame)
for(i in 1:length(bg10k_mammal.bio)){
  mutate(bg10k_mammal.bio[[i]],
         X = bg10kcoord_mammal[[i]]$X,
         Y = bg10kcoord_mammal[[i]]$Y) -> bg10k_mammal.bio[[i]]
}
bg10k_mammal.bio <- map(bg10k_mammal.bio, na.omit)
```





```{r}
my.pseperate <- function(x){dplyr::bind_cols(dplyr::slice(x@coords, 1:which(x@pa==0)[1]-1),
                                             dplyr::slice(x@data, 1:which(x@pa==0)[1]-1))}


p.data.mammal <- map(data.mammal, my.pseperate)

bg10kdata_mammal <- list()
for(i in 1:length(p.data.mammal)){
  bg10kdata_mammal[[i]] <- as.data.frame(
   rbind(p.data.mammal[[i]][,3:21], bg10k_mammal.bio[[i]][,1:19]))}

newcoord_mammal <- list()
for(i in 1:length(p.data.mammal)){
  newcoord_mammal[[i]] <-
   rbind(p.data.mammal[[i]][,1:2], bg10k_mammal.bio[[i]][,20:21])}

newpa_mammal <- list()
for(i in 1:length(p.data.mammal)){
newpa_mammal[[i]] <- c(rep(1, nrow(p.data.mammal[[i]])), 
                       rep(0, nrow(bg10k_mammal.bio[[i]])))}


for(i in 1:length(data.mammal)){
data.mammal[[i]]@data <- bg10kdata_mammal[[i]]
data.mammal[[i]]@coords <- newcoord_mammal[[i]]
data.mammal[[i]]@pa <- newpa_mammal[[i]]
}


```


```{r}
bg10k_mammal.xy <- list()
for(i in 1:length(bg10k_mammal.bio)){
bg10k_mammal.xy[[i]] <- bg10k_mammal.bio[[i]][,20:21]}

my.swdmaker <- function(x,y,z){SDMtune::prepareSWD(
  species = x@species, 
  p = y[,1:2], 
  a = z[,1:2], 
  env = stackbio)}

data.mammal <- furrr::future_map2(data.mammal, p.data.mammal, bg10k_mammal.xy,
                                  my.swdmaker, 
                                  .options = furrr_options(seed = TRUE), .progress = T)


```


PCNM model 65 species
```{r}
data.reptile <- readr::read_rds("data/newdata/data.pcnm.reptile.rds")
data.amphibian <- readr::read_rds("data/newdata/data.pcnm.amphibian.rds")
```




데이터를 어떻게 쪼개느냐에 따라 모델 성능이 갈릴 것이다.
random k fold는 가진 모든 출현 데이터를 k개의 가짓수로 train과 test 셋으로 나눈다.
1000개에 5 fold 라면, 200개는 test셋이고 800개는 train 셋인데 
각각의 샘플이 한번씩 test에 들어가도록 구성해야 하니 5개의 조합이 나온다.

학습 데이터 조합은 5 조합이고
새로운 테스트 아니 벨리데이션 데이터셋도 5개가 생기는 셈이다.


```{r t.t.가져오기}
t.t.mammal_0.2 <- read_rds("t.t/newt.t/t.t.mammal_0.2.rds")
t.t.reptile_0.2 <- read_rds("t.t/newt.t/t.t.reptile_0.2.rds")
t.t.amphibian_0.2 <- read_rds("t.t/newt.t/t.t.amphibian_0.2.rds")

```





```{r}
t.t.name <- function(x){c(paste0("train_", x$species[1]), paste0("test_", x$species[1]))}
t.t.name_mammal <- map(sf_mammal, t.t.name)

```

```{r include=FALSE}
#split 3, train, test, validation
c(train, val, test) %<-% trainValTest(data.mammal_pilot[[1]], val = 0.2, test = 0.2, only_presence = TRUE, seed = 4839)

```

```{r train model : }

library(SDMtune)

#split 2, train, test

# 20 % for test
t.t.mammal_0.2 <- list()
for (i in 1:length(data.mammal)){
  t.t.mammal_0.2[[i]] <- trainValTest(data.mammal[[i]], 
                                      test = 0.2, 
                                      only_presence = TRUE, 
                                      seed = 1)}

t.t.reptile_0.2 <- list()
for (i in 1:length(data.reptile)){t.t.reptile_0.2[[i]] <- trainValTest(data.reptile[[i]], test = 0.2, only_presence = TRUE, seed = 1)}

t.t.amphibian_0.2 <- list()
for (i in 1:length(data.amphibian)){t.t.amphibian_0.2[[i]] <- trainValTest(data.amphibian[[i]], test = 0.2, only_presence = TRUE, seed = 1)}





folds.rnd5k_mammal <- list()
for (i in 1:length(t.t.mammal_0.2)){rnd5k_mammal[[i]] <- randomFolds(t.t.mammal_0.2[[i]][[1]], k = 5, only_presence = TRUE, seed = 3667)}




#Crude
#default_model <- train(method = "Maxent", data = data)
#default_model <- train(method = "Maxent", data = data, fc = "lqph", reg = 1, iter = 500)

#without validation set
#c(train, test) %<-% trainValTest(data, test = 0.2, only_presence = TRUE, seed = 25)
#maxnet_model <- train("Maxnet", data = train)

#random 4 fold cross validation


#purrr
rnd4k <- function(x){randomFolds(x[[1]], 
                                 k = 4,
                                 only_presence = TRUE,
                                 seed = 1)}

rnd5k <- function(x){randomFolds(x[[1]], 
                                 k = 5,
                                 only_presence = TRUE,
                                 seed = 3668)}

folds.rnd4k_mammal <- map(t.t.mammal_0.2, rnd4k)
folds.rnd5k_mammal <- map(t.t.mammal_0.2, rnd5k)

folds.rnd4k_reptile <- map(t.t.reptile_0.2, rnd4k)

folds.rnd4k_amphibian <- map(t.t.amphibian_0.2, rnd4k)


model_rndcv4k_mammal <- list()
for (i in 1:length(rnd4k_mammal)){model_rndcv4k_mammal[[i]] <- train("Maxnet", data =t.t.mammal_0.2[[i]][[1]], folds = rnd4k_mammal[[i]])}


#purrr

maxnet.train <- function(x, y){SDMtune::train("Maxnet",
                                               data = x[[1]],
                                               folds = y)}
maxent.train <- function(x, y){SDMtune::train("Maxent",
                                               data = x[[1]],
                                               folds = y)}




model_rndcv4k_mammal <- list()
for(i in 1:length(t.t.mammal_0.2)){
  model_rndcv4k_mammal[[i]] <- maxnet.train(t.t.mammal_0.2[[i]], folds.rnd4k_mammal[[i]])
}


test_t.t <- t.t.mammal_0.2[[1]]
test_fold <- folds.rnd4k_mammal[[1]]
test_t.t[[1]]@data <- test_t.t[[1]]@data[,1:3]
test_t.t[[2]]@data <- test_t.t[[2]]@data[,1:3]

test_t.t2 <- t.t.mammal_0.2[[1]]
test_fold2 <- folds.rnd4k_mammal[[1]]
test_t.t2[[1]]@data <- test_t.t2[[1]]@data[,1:4]
test_t.t2[[2]]@data <- test_t.t2[[2]]@data[,1:4]

test_t.t3 <- t.t.mammal_0.2[[1]]
test_fold3 <- folds.rnd4k_mammal[[1]]
test_t.t3[[1]]@data <- test_t.t3[[1]]@data[,c(1,2,3,5)]
test_t.t3[[2]]@data <- test_t.t3[[2]]@data[,c(1,2,3,5)]

ent1 <- vector()
for(i in 1:4){
   ent1[i] <- test_model@models[[i]]@model@model$entropy
}

ent2 <- vector()
for(i in 1:4){
   ent2[i] <- test_model2@models[[i]]@model@model$entropy
}

ent3 <- vector()
for(i in 1:4){
   ent3[i] <- test_model3@models[[i]]@model@model$entropy
}

model_rndcv4k_mammal <- future_map2(t.t.mammal_0.2, 
                                    folds.rnd4k_mammal, 
                                    maxnet.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_rndcv4k_reptile <- future_map2(t.t.reptile_0.2, 
                                    folds.rnd4k_reptile, 
                                    maxnet.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_rndcv4k_reptile2 <- future_map2(t.t.reptile_0.2, 
                                    folds.rnd4k_reptile, 
                                    maxent.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_rndcv4k_amphibian <- future_map2(t.t.amphibian_0.2, 
                                    folds.rnd4k_amphibian, 
                                    maxnet.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_rndcv4k_amphibian2 <- future_map2(t.t.amphibian_0.2, 
                                    folds.rnd4k_amphibian, 
                                    maxent.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)


```
while PCNM reptile.. 39 stil got glmnet error
```{r}
t.t.reptile_0.2_without39 <- t.t.reptile_0.2[-39]
folds.rnd4k_reptile_without39 <- folds.rnd4k_reptile[-39]
model_rndcv4k_reptile <- future_map2(t.t.reptile_0.2_without39, 
                                    folds.rnd4k_reptile_without39, 
                                    maxnet.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model <- list()
for(i in 1:38){model[[i]] <- model_rndcv4k_reptile[[i]]}
model[[39]] <- model_rndcv4k_reptile2[[39]]
for(i in 39:64){model[[i+1]] <- model_rndcv4k_reptile[[i]]}

model_rndcv4k_reptile <- model
write_rds(model_rndcv4k_reptile, "model/newmodel/model_pcnm/model_rndcv4k_reptile.rds")
```

here 10th got glm error
```{r}
t.t.amphibian_0.2_without1023 <- t.t.amphibian_0.2[-c(10,23)]
folds.rnd4k_amphibian_without1023 <- folds.rnd4k_amphibian[-c(10,23)]
model_rndcv4k_amphibian <- future_map2(t.t.amphibian_0.2_without1023, 
                                    folds.rnd4k_amphibian_without1023, 
                                    maxnet.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

t.t.amphibian_0.2_without10 <- t.t.amphibian_0.2_without10[-22]#it was 23
folds.rnd4k_amphibian_without10 <- folds.rnd4k_amphibian_without10[-22]#it was 23
model_rndcv4k_amphibian <- future_map2(t.t.amphibian_0.2_without10, 
                                    folds.rnd4k_amphibian_without10, 
                                    maxnet.train, 
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T) 

model <- list()
for(i in 1:9){model[[i]] <- model_rndcv4k_amphibian[[i]]}
model[[10]] <- model_rndcv4k_amphibian2[[10]]
for(i in 10:21){model[[i+1]] <- model_rndcv4k_amphibian[[i]]}
model[[23]] <- model_rndcv4k_amphibian2[[23]]
for(i in 22:30){model[[i+2]] <- model_rndcv4k_amphibian[[i]]}
for(i in 1:32){print(model[[i]]@data@species)}

model_rndcv4k_amphibian <- model
write_rds(model_rndcv4k_amphibian, "model/newmodel/model_pcnm/model_rndcv4k_amphibian.rds")
```


```{r}
write_rds(t.t.reptile_0.2, "t.t/newt.t/t.t.pcnm.reptile_0.2.rds")
write_rds(t.t.amphibian_0.2, "t.t/newt.t/t.t.pcnm.amphibian_0.2.rds")


```



# spatial block cross validation model

```{r}
library(blockCV)
library(precrec)
```

Find proper size of spatial block.

load cropped raster
```{r}
bio_crop_mammal <- list()
for(i in 1:35){
list.files("D:/bio_crop/mammal", pattern = "*.tif", full.names = T)[i] %>% 
  raster::stack(.) -> bio_crop_mammal[[i]]}

bio_crop_reptile <- list()
for(i in 1:92){
list.files("D:/bio_crop/reptile2", pattern = "*.tif", full.names = T)[i] %>% 
  raster::stack(.) -> bio_crop_reptile[[i]]}

bio_crop_amphibian <- list()
for(i in 1:50){
list.files("D:/bio_crop/amphibian", pattern = "*.tif", full.names = T)[i] %>% 
  raster::stack(.) -> bio_crop_amphibian[[i]]}
```

Get the range of SA in range of habitat
```{r}
my.SARange <- function(x){blockCV::spatialAutoRange(rasterLayer = x,
                                          sampleNumber = 10000,
                                          doParallel = TRUE,
                                          showPlots = TRUE, 
                                          progress = T)}

SARange_mammal <- list()
for(i in 1:35){
  SARange_mammal[[i]] <- my.SARange(bio_crop_mammal[[i]])
}


SARange_reptile <- list()
for(i in 1:91){
  SARange_reptile[[i]] <- my.SARange(bio_crop_reptile[[i]])
}

SARange_amphibian <- list()
for(i in 1:50){
  SARange_amphibian[[i]] <- my.SARange(bio_crop_amphibian[[i]])
}

```

76 got error..Teira dugesii.. Tiny island lizzard. remove..
```{r}
SARange_reptile[[76]] <- NULL
bio_crop_reptile[[76]] <- NULL
data.reptile[[76]] <- NULL
t.t.reptile_0.2[[76]] <- NULL
```



```{r}
for(i in 77:92){
  SARange_reptile[[i]] <- my.SARange(bio_crop_reptile[[i]])
}

```




few of them fucked up.
In Conservative view, these are fucked up.
2, 3, 4, 5, 6, 7, 11, 15, 16, 17, 19, 20, 21, 23, 25, 26, 28, 29, 31, 32, 33, 34, 35
try more points
```{r}
test <- blockCV::spatialAutoRange(rasterLayer = bio_crop_mammal[[2]],
                                  sampleNumber = 10000,
                                  doParallel = TRUE,
                                  showPlots = TRUE,
                                  nCores = 60, 
                                  progress = TRUE)
```

more liberal, these are fucked up.
2, 3, 4, 5, 6, 7, 11, 15, 16, 17, 19, 20, 21, 28, 29, 33, 34

2 <- 1:4 1183532
3 <- 1:4 559611
4 <- 1:5 696445
5 <- 1:4 1622887(1404230)
6 <- 1:5 1312366
```{r}
blockCV::rangeExplorer(rasterLayer = bio_crop_mammal[[29]],
              speciesData = sf_train_mammal[[29]],
              species = "pa",
              rangeTable = NULL,
              minRange = 1000000, # limit the search domain
              maxRange = 3000000)
```

reptile fucked ups are



```{r}
library(automap)

plot(SARange_mammal[[5]]$variograms[[4]])
```
it seems first few rasters' ranges mean will works fine. 
so mean of 1:4 will be my range for those doesn't work well.

```{r}
good.range_mammal <- vector()
for(i in 1:length(SARange_mammal)){
  good.range_mammal[i] <- SARange_mammal[[i]]$range
}

for(i in c(2, 3, 4, 5, 6, 7, 11, 15, 16, 17, 19, 20, 21, 23, 25, 26, 28, 29, 31, 32, 33, 34, 35)){
  good.range_mammal[i] <- mean(SARange_mammal[[i]]$rangeTable$range[1:4])
}


good.range_reptile <- vector()
for(i in 1:length(SARange_reptile)){
  good.range_reptile[i] <- SARange_reptile[[i]]$range
}

for(i in c(1,2,7,8,9,10,11,12,13,15,16,17,19,20,22,23,24,25,26,31,32,33,34,35,38,39,42,43,44,45,47,48,51,52,54,56,57,58,60,61,62,66,67,69,70,71,73,76,78,80,81,83,84,86,87,91)){
  good.range_reptile[i] <- mean(SARange_reptile[[i]]$rangeTable$range[1:4])
}

good.range_amphibian <- vector()
for(i in 1:length(SARange_amphibian)){
  good.range_amphibian[i] <- SARange_amphibian[[i]]$range
}

for(i in c(2,4,5,6,9,11,12,16,17,18,20,23,25,27,28,29,30,31,33,34,35,39,40,41,42,49)){
  good.range_amphibian[i] <- mean(SARange_amphibian[[i]]$rangeTable$range[1:4])
}

```


```{r}
#10 keep getting error even low distance
blockCV::rangeExplorer(rasterLayer = bio_crop_reptile[[10]],
              speciesData = sf_train_reptile[[10]],
              species = "pa",
              rangeTable = NULL,
              minRange = 100000, # limit the search domain
              maxRange = 300000)
```

t.t as sf
it must include pa column
```{r}
tb_train_mammal <- list()
for(i in 1:length(t.t.mammal_0.2)){
tb_train_mammal[[i]] <- tibble(long = t.t.mammal_0.2[[i]][[1]]@coords$X,
                               lat = t.t.mammal_0.2[[i]][[1]]@coords$Y,
                               pa = t.t.mammal_0.2[[i]][[1]]@pa,
                               data = t.t.mammal_0.2[[i]][[1]]@data)
}

tb_train_reptile <- list()
for(i in 1:length(data.reptile)){
tb_train_reptile[[i]] <- tibble(long = t.t.reptile_0.2[[i]][[1]]@coords$X,
                               lat = t.t.reptile_0.2[[i]][[1]]@coords$Y,
                               pa = t.t.reptile_0.2[[i]][[1]]@pa,
                               data = t.t.reptile_0.2[[i]][[1]]@data)
}

tb_train_amphibian <- list()
for(i in 1:length(data.amphibian)){
tb_train_amphibian[[i]] <- tibble(long = t.t.amphibian_0.2[[i]][[1]]@coords$X,
                               lat = t.t.amphibian_0.2[[i]][[1]]@coords$Y,
                               pa = t.t.amphibian_0.2[[i]][[1]]@pa,
                               data = t.t.amphibian_0.2[[i]][[1]]@data)
}
  
sf_train_mammal <- list()
for(i in 1:length(tb_train_mammal)){
sf_train_mammal[[i]] <- sf::st_as_sf(tb_train_mammal[[i]],
                                coords = c(1,2), 
                                crs = 4326,
                                remove = F)
}

sf_train_reptile <- list()
for(i in 1:length(tb_train_reptile)){
sf_train_reptile[[i]] <- sf::st_as_sf(tb_train_reptile[[i]],
                                coords = c(1,2), 
                                crs = 4326,
                                remove = F)
}

sf_train_amphibian <- list()
for(i in 1:length(tb_train_amphibian)){
sf_train_amphibian[[i]] <- sf::st_as_sf(tb_train_amphibian[[i]],
                                coords = c(1,2), 
                                crs = 4326,
                                remove = F)
}
```

```{r}
folds.sp4k_mammal <- list()
folds.sp4k_mammal[[1]] <- blockCV::spatialBlock(speciesData = sf_train_mammal[[1]],
                                                  species = "pa",
                                                  rasterLayer = bio_crop_mammal[[1]],
                                                  theRange = good.range_mammal[1],
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)

my.spatialblock <- function(x, y, z){blockCV::spatialBlock(speciesData = x,
                                                  species = "pa",
                                                  rasterLayer = y,
                                                  theRange = z,
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)}

folds.sp4k_mammal <- furrr::future_pmap(sf_train_mammal,
                                         bio_crop_mammal,
                                         SARange_mammal)
folds.sp4k_mammal <- list()
for(i in 29:length(sf_train_mammal)){
folds.sp4k_mammal[[i]] <- blockCV::spatialBlock(speciesData = sf_train_mammal[[i]],
                                                  species = "pa",
                                                  rasterLayer = bio_crop_mammal[[i]],
                                                  theRange = good.range_mammal[i],
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)}

#11, 16, 20 is too small austarailia
#give it 1000000
good.range_mammal[11] <- 1000000
good.range_mammal[16] <- 1000000
good.range_mammal[20] <- 1000000
good.range_mammal[29] <- 2277277


folds.sp4k_reptile <- list()
for(i in 1:length(sf_train_reptile)){
folds.sp4k_reptile[[i]] <- blockCV::spatialBlock(speciesData = sf_train_reptile[[i]],
                                                  species = "pa",
                                                  rasterLayer = bio_crop_reptile[[i]],
                                                  theRange = good.range_reptile[i],
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)}



```
```{r}
#76 less block. decrease size to 1st range
good.range_reptile[76] <- 824164.8   

for(i in 76:length(sf_train_reptile)){
folds.sp4k_reptile[[i]] <- blockCV::spatialBlock(speciesData = sf_train_reptile[[i]],
                                                  species = "pa",
                                                  rasterLayer = bio_crop_reptile[[i]],
                                                  theRange = good.range_reptile[i],
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)}
```

load old folds
```{r}
folds.sp4k_reptile <- read_rds("folds/folds.sp4k_reptile.rds") #this is 91
folds.sp4k_amphibian <- read_rds("folds/folds.sp4k_amphibian.rds") #this is 50
```

my data is 65, 32.

find couples

sp block have 91. original as data

These are I cut from 91 to 70 to make distm
```{r}
folds.sp4k_reptile.copy <- folds.sp4k_reptile
folds.sp4k_reptile.copy[c(1,  5,  7,  8,  9, 10, 11, 12, 14, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 45, 46, 47, 48, 49, 50, 52, 53, 55, 57, 58, 59, 61, 62, 64, 66, 67, 68, 70, 71, 72, 73, 74, 75, 77, 78, 79, 81, 83, 86, 87, 88, 90)] -> folds.sp4k_reptile.copy
```

and these are I cut from 70 to 65 to get pcnm vectors that have at least 14
```{r}
folds.sp4k_reptile.copy[-c(22, 27, 58, 60, 61)] -> folds.sp4k_reptile.copy
```

about amphibian

```{r}
folds.sp4k_amphibian.copy <- folds.sp4k_amphibian
folds.sp4k_amphibian.copy[-c(5,9,10,23,25,26,27,29,31,33,34,42,44,45,49)] -> folds.sp4k_amphibian.copy
folds.sp4k_amphibian.copy[-c(11, 21, 25)] -> folds.sp4k_amphibian.copy
```




```{r}
nrow.spfold.reptile <- vector()
for(i in 1:length(folds.sp4k_reptile.copy)){
nrow.spfold.reptile[i] <- (folds.sp4k_reptile.copy[[i]]$records[1,2] + folds.sp4k_reptile.copy[[i]]$records[1,4])}

nrow.t.t.reptile <- vector()
for(i in 1:length(t.t.reptile_0.2)){
nrow.t.t.reptile[i] <- sum(t.t.reptile_0.2[[i]][[1]]@pa == 1)}

nrow.spfold.amphibian <- vector()
for(i in 1:length(folds.sp4k_amphibian.copy)){
nrow.spfold.amphibian[i] <- (folds.sp4k_amphibian.copy[[i]]$records[1,2] + folds.sp4k_amphibian.copy[[i]]$records[1,4])}

nrow.t.t.amphibian <- vector()
for(i in 1:length(t.t.amphibian_0.2)){
nrow.t.t.amphibian[i] <- sum(t.t.amphibian_0.2[[i]][[1]]@pa == 1)}

```

```{r}
nrow.spfold.reptile
```
```{r}
nrow.t.t.reptile
```
```{r}
which(nrow.spfold.reptile == nrow.t.t.reptile[7])
```




```{r}
nrow.spfold.amphibian
```

```{r}
nrow.t.t.amphibian
```


```{r}
model_spcv4k_mammal <- future_map2(t.t.mammal_0.2, 
                                    folds.sp4k_mammal, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_spcv4k_reptile <- future_map2(t.t.reptile_0.2, 
                                    folds.sp4k_reptile.copy, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_spcv4k_amphibian <- future_map2(t.t.amphibian_0.2, 
                                    folds.sp4k_amphibian.copy, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)


```


62 got glm error... try without it
```{r}
t.t.reptile_0.2.test <- t.t.reptile_0.2
folds.sp4k_reptile.test <- folds.sp4k_reptile

t.t.reptile_0.2.test[[62]] <- t.t.reptile_0.2.test[[1]]
folds.sp4k_reptile.test[[62]] <- folds.sp4k_reptile.test[[1]]

model_spcv4k_reptile.test <- future_map2(t.t.reptile_0.2.test, 
                                    folds.sp4k_reptile.test, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T) # it is done!
```

```{r}
t.t.62 <- t.t.reptile_0.2[[62]]
fold.62 <- folds.sp4k_reptile[[62]]

bgtest <- list(SDMtune::addSamplesToBg(t.t.62[[1]]),
               SDMtune::addSamplesToBg(t.t.62[[2]]))

model_spcv4k_reptile.test[[62]] <- maxnet.train(bgtest, fold.62)

```



```{r}
folds.sp4k_amphibian <- list()
for(i in 1:length(sf_train_amphibian)){
folds.sp4k_amphibian[[i]] <- blockCV::spatialBlock(speciesData = sf_train_amphibian[[i]],
                                                  species = "pa",
                                                  rasterLayer = bio_crop_amphibian[[i]],
                                                  theRange = good.range_amphibian[i],
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)}
```

amphibian sp cv 4k

```{r}
#49 got less block : 3
good.range_amphibian[49] <- 648399.2

for(i in 49:length(sf_train_amphibian)){
folds.sp4k_amphibian[[i]] <- blockCV::spatialBlock(speciesData = sf_train_amphibian[[i]],
                                                  species = "pa",
                                                  rasterLayer = bio_crop_amphibian[[i]],
                                                  theRange = good.range_amphibian[i],
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)}

```





```{r}
#cut only presence
for(i in 1:length(sf_train_mammal)){
sf_train_mammal[[i]] <- slice(sf_train_mammal[[i]], 1:which(t.t.mammal_0.2[[i]][[1]]@pa==0)[1]-1)
}

for(i in 1:length(sf_train_reptile)){
sf_train_reptile[[i]] <- slice(sf_train_reptile[[i]], 1:which(t.t.reptile_0.2[[i]][[1]]@pa==0)[1]-1)
}

for(i in 1:length(sf_train_amphibian)){
sf_train_amphibian[[i]] <- slice(sf_train_amphibian[[i]], 1:which(data.amphibian[[i]]@pa==0)[1]-1)
}

#env fold

folds.env_mammal <- list()
for(i in 1:length(sf_train_mammal)){
folds.env_mammal[[i]] <- my.env.block(rasterLayer = stackbio,
                                       speciesData = sf_train_mammal[[i]],
                                       species = "pa",
                                       k = 4,
                                       standardization = TRUE,
                                       rasterBlock = FALSE,
                                       ext = sf_train_mammal[[i]]$data)
}


folds.env_reptile <- list()
for(i in 1:length(sf_train_reptile)){
folds.env_reptile[[i]] <- my.env.block(rasterLayer = stackbio,
                                       speciesData = sf_train_reptile[[i]],
                                       species = "pa",
                                       k = 4,
                                       standardization = TRUE,
                                       rasterBlock = FALSE,
                                       ext = sf_train_reptile[[i]]$data)
}



folds.env_amphibian <- list()
for(i in 1:length(sf_train_amphibian)){
folds.env_amphibian[[i]] <- my.env.block(rasterLayer = stackbio,
                                       speciesData = sf_train_amphibian[[i]],
                                       species = "pa",
                                       k = 4,
                                       standardization = TRUE,
                                       rasterBlock = FALSE,
                                       ext = sf_train_amphibian[[i]]$data)
}


```

```{r}
a <- list()
for(i in 1:length(folds.env_reptile)){
  a[[i]] <- folds.env_reptile[[i]]$records
}


```






```{r}


folds.block_mammal <- list()
for(i in 1:length(data.mammal)){
  folds.block_mammal[[i]] <- get.block(occ = t.t.mammal_0.2[[i]][[1]]@data[t.t.mammal_0.2[[i]][[1]]@pa == 1,], 
                                       bg.coords = data.mammal[[i]]@data[data.mammal[[i]]@pa == 0, ])
}

folds.block_reptile <- list()
for(i in 1:length(data.reptile)){
  folds.block_reptile[[i]] <- ENMeval::get.block(occ = t.t.reptile_0.2[[i]][[1]]@data[t.t.reptile_0.2[[i]][[1]]@pa == 1,], 
                                       bg.coords = data.reptile[[i]]@data[data.reptile[[i]]@pa == 0, ])
}

folds.block_amphibian <- list()
for(i in 1:length(data.amphibian)){
  folds.block_amphibian[[i]] <- get.block(occ = t.t.amphibian_0.2[[i]][[1]]@data[t.t.amphibian_0.2[[i]][[1]]@pa == 1,], 
                                       bg.coords = data.amphibian[[i]]@data[data.amphibian[[i]]@pa == 0, ])
}



model_spcv4k_mammal <- future_map2(t.t.mammal_0.2, 
                                    folds.block_mammal, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_spcv4k_reptile <- future_map2(t.t.reptile_0.2, 
                                    folds.block_reptile, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_spcv4k_reptile <- list()
for(i in 1:length(t.t.reptile_0.2)){
  model_spcv4k_reptile[[i]] <- maxnet.train(t.t.reptile_0.2[[i]], folds.block_reptile[[i]])
}

model_spcv4k_amphibian <- future_map2(t.t.amphibian_0.2, 
                                    folds.block_amphibian, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)


```


cb_folds <- get.checkerboard1(occ = data@data[data@pa == 1, ], 
                              env = predictors, 
                              bg.coords = data@data[data@pa == 0, ], 
                              aggregation.factor = 4)

model <- train(method = "Maxnet", data = data, fc = "l", reg = 0.8, folds = cb_folds)

# 3. Environmental block


```{r}
library(raster)
biopath <- paste0("bios/", list.files(pattern="*.tif", path = "bios/"))
stackbio <- raster::stack("stackbio.tif")
```


things are fucked up... make again
```{r}
for(i in 29:length(sf_train_mammal)){
folds.sp4k_mammal[[i]] <- blockCV::spatialBlock(speciesData = sf_train_mammal[[i]],
                                                  species = "pa",
                                                  rasterLayer = bio_crop_mammal[[i]],
                                                  theRange = good.range_mammal[i],
                                                  k = 4, 
                                                  selection = "random", 
                                                  numLimit = 0, 
                                                  iteration = 100, 
                                                  showBlocks = TRUE, 
                                                  biomod2Format = TRUE, 
                                                  progress = TRUE, 
                                                  verbose = TRUE)}
```


```{r}

folds.env_mammal <- list()
for(i in 1:35){
folds.env_mammal[[i]] <- blockCV:::envBlock(rasterLayer = stackbio,
                         speciesData = sf_train_mammal[[i]],
                         species = "pa",
                         k = 4,
                         standardization = "normal", #표준화
                         rasterBlock = FALSE,
                         numLimit = 100,
                         biomod2Format = T,
                         verbose = T)  
}

```


```{r}


# Create spatial points data frame
sp_mammal <- list()
for (i in 1:length(t.t.mammal_0.2)){
sp_mammal[[i]] <- SpatialPointsDataFrame(t.t.mammal_0.2[[i]][[1]]@coords, 
                                         data = data.frame(t.t.mammal_0.2[[i]][[1]]@pa),
                                         proj4string = CRS("+proj=longlat +datum=WGS84"))
}

for (i in 1:length(sp_mammal)){
names(sp_mammal[[i]]) <- "pa"
}  



model_envcv4k_mammal <- future_map2(t.t.mammal_0.2, 
                                    folds.env_mammal, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_envcv4k_mammal <- list()
for(i in 20:length(t.t.mammal_0.2)){
  model_envcv4k_mammal[[i]] <- maxnet.train(t.t.mammal_0.2[[i]], folds.env_mammal[[i]])
}

model_envcv4k_mammal[[6]] <- maxnet.train(t.t.mammal_0.2[[6]], folds.env_mammal[[6]])

#3 got bound error
#6 got bound error
#8 got error
#19 got error
#22 got error 


model_envcv4k_reptile <- future_map2(t.t.reptile_0.2, 
                                    folds.env_reptile, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_envcv4k_amphibian <- future_map2(t.t.amphibian_0.2, 
                                    folds.env_amphibian, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

model_envcv4k_mammal <- future_map2(t.t.mammal_0.2, 
                                    folds.env_mammal, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)


#reptile 8 got error

t.t.reptile_0.2.tmp <- t.t.reptile_0.2[-8]
folds.env_reptile.tmp <- folds.env_reptile[-8]

t.t.reptile_0.2.tmp <- t.t.reptile_0.2.tmp[-14]
folds.env_reptile.tmp <- folds.env_reptile.tmp[-14]

model_envcv4k_reptile.tmp <- future_map2(t.t.reptile_0.2.tmp, 
                                    folds.env_reptile.tmp, 
                                    maxnet.train,
                                    .options = furrr_options(seed = TRUE),
                                    .progress = T)

```
```{r}

for(i in 9:length(t.t.reptile_0.2)){
  model_envcv4k_reptile[[i]] <- maxnet.train(t.t.reptile_0.2[[i]], folds.env_reptile[[i]])
}
```


































```{r}
new("standardGeneric", .Data = function (object, ...) 
standardGeneric("predict"), generic = "predict", package = "stats", 
  group = list(), valueClass = character(0), signature = "object", 
  default = new("derivedDefaultMethod", .Data = function (object, 
    ...) 
  UseMethod("predict"), target = new("signature", .Data = "ANY", 
    names = "object", package = "methods"), defined = new("signature", 
    .Data = "ANY", names = "object", package = "methods"), 
    generic = "predict"), skeleton = (new("derivedDefaultMethod", 
    .Data = function (object, ...) 
    UseMethod("predict"), target = new("signature", .Data = "ANY", 
      names = "object", package = "methods"), defined = new("signature", 
      .Data = "ANY", names = "object", package = "methods"), 
    generic = "predict"))(object, ...))

```







```{r}
predict.test <- function(x,y){
  SDMtune::predict(x,
                   data = y[[2]],
                   fun = "mean",
                   type = "cloglog")
}

predict.train <- function(x,y){
  SDMtune::predict(x,
                   data = y[[1]],
                   fun = "mean",
                   type = "cloglog")
}



test_spcv4k_mammal <- map2(model_spcv4k_mammal, 
                                   t.t.mammal_0.2, 
                                   predict.test 
                                   )

train_spcv4k_mammal <- map2(model_spcv4k_mammal, 
                                   t.t.mammal_0.2, 
                                   predict.train 
                                   )

test_rndcv4k_mammal <- map2(model_rndcv4k_mammal, 
                                   t.t.mammal_0.2, 
                                   predict.test 
                                   )

train_rndcv4k_mammal <- map2(model_rndcv4k_mammal, 
                                   t.t.mammal_0.2, 
                                   predict.train 
                                   )

test_envcv4k_mammal <- map2(model_envcv4k_mammal, 
                                   t.t.mammal_0.2, 
                                   predict.test 
                                   )


train_rndcv4k_reptile <- map2(model_rndcv4k_reptile, 
                                   t.t.reptile_0.2, 
                                   predict.train 
                                   )

test_rndcv4k_reptile <- map2(model_rndcv4k_reptile, 
                                   t.t.reptile_0.2, 
                                   predict.test 
                                   )

train_rndcv4k_amphibian <- map2(model_rndcv4k_amphibian, 
                                   t.t.amphibian_0.2, 
                                   predict.train 
                                   )

test_rndcv4k_amphibian <- map2(model_rndcv4k_amphibian, 
                                   t.t.amphibian_0.2, 
                                   predict.test 
                                   )


```








```{r}
rmse.train_rndcv4k_mammal <- vector()
for (i in 1:length(train_rndcv4k_mammal)){
 rmse.train_rndcv4k_mammal[i] <- sqrt(sum((1- train_rndcv4k_mammal[[i]])^2) / length(train_rndcv4k_mammal[[i]]))
}

rmse.test_rndcv4k_mammal <- vector()
for (i in 1:length(test_rndcv4k_mammal)){
 rmse.test_rndcv4k_mammal[i] <- sqrt(sum((1- test_rndcv4k_mammal[[i]])^2) / length(test_rndcv4k_mammal[[i]]))
}

rmse.train_spcv4k_mammal <- vector()
for (i in 1:length(train_spcv4k_mammal)){
 rmse.train_spcv4k_mammal[i] <- sqrt(sum((1- train_spcv4k_mammal[[i]])^2) / length(train_spcv4k_mammal[[i]]))
}

rmse.test_spcv4k_mammal <- vector()
for (i in 1:length(test_spcv4k_mammal)){
 rmse.test_spcv4k_mammal[i] <- sqrt(sum((1- test_spcv4k_mammal[[i]])^2) / length(test_spcv4k_mammal[[i]]))
}





rmse.train_rndcv4k_reptile <- vector()
for (i in 1:length(train_rndcv4k_reptile)){
 rmse.train_rndcv4k_reptile[i] <- sqrt(sum((1- train_rndcv4k_reptile[[i]])^2) / length(train_rndcv4k_reptile[[i]]))
}

rmse.test_rndcv4k_reptile <- vector()
for (i in 1:length(test_rndcv4k_reptile)){
 rmse.test_rndcv4k_reptile[i] <- sqrt(sum((1- test_rndcv4k_reptile[[i]])^2) / length(test_rndcv4k_reptile[[i]]))
}

rmse.train_spcv4k_reptile <- vector()
for (i in 1:length(train_spcv4k_reptile)){
 rmse.train_spcv4k_reptile[i] <- sqrt(sum((1- train_spcv4k_reptile[[i]])^2) / length(train_spcv4k_reptile[[i]]))
}

rmse.test_spcv4k_reptile <- vector()
for (i in 1:length(test_spcv4k_reptile)){
 rmse.test_spcv4k_reptile[i] <- sqrt(sum((1- test_spcv4k_reptile[[i]])^2) / length(test_spcv4k_reptile[[i]]))
}

rmse_rndcv4k_reptile <- vector()
for (i in 1:length(rndcv4k_reptile.df)){
 rmse_rndcv4k_reptile[i] <- sqrt(sum((1 - rndcv4k_reptile.df[[i]][,4])^2) / length(rndcv4k_reptile.df[[i]]))
}
```

```{r}
result_mammal <- readr::read_csv("result_mammal.csv")

n.points <- vector()
for(i in 1:length(data.mammal)){n.points[i] <- nrow(data.mammal[[i]]@data)}


result_mammal %>% mutate(rmse.test_rndcv4k = rmse.test_rndcv4k_mammal,
                         rmse.test_spcv4k = rmse.test_spcv4k_mammal,
                         rmse.train_rndcv4k = rmse.train_rndcv4k_mammal,
                         rmse.train_spcv4k = rmse.train_spcv4k_mammal,
                         n.points = n.points) -> result_mammal
readr::write_csv(result_mammal,"result_mammal.csv")


result_reptile <- readr::read_csv("result_reptile.csv")

n.points <- vector()
for(i in 1:length(data.reptile)){n.points[i] <- nrow(data.reptile[[i]]@data)}


result_reptile %>% mutate(rmse.test_rndcv4k = rmse.test_rndcv4k_reptile,
                          rmse.train_rndcv4k = rmse.train_rndcv4k_reptile,
                          rmse.all_rndcv4k = rmse_rndcv4k_reptile
                          ) -> result_reptile
readr::write_csv(result_reptile,"result_reptile.csv")


```





```{r}
qq <- ggplot(data = result_mammal) +
    geom_point(aes(x = rsac_rndcv4k_sph, y = rmse.train_rndcv4k_mammal, color = 1)) +
    geom_smooth(aes(x = rsac_rndcv4k_sph, y= rmse.train_rndcv4k_mammal), method = "lm")

qq <- ggplot(data = result_mammal) +
    geom_point(aes(x = nb_01, y = rmse.train_rndcv4k_mammal, color = 1)) +
    geom_smooth(aes(x = nb_01, y= rmse.train_rndcv4k_mammal), method = "lm") +
    geom_point(aes(x = nb_03, y = rmse.train_rndcv4k_mammal, color = 2)) +
    geom_smooth(aes(x = nb_03, y= rmse.train_rndcv4k_mammal), method = "lm") +
    geom_point(aes(x = nb_05, y = rmse.train_rndcv4k_mammal, color = 3)) +
    geom_smooth(aes(x = nb_05, y= rmse.train_rndcv4k_mammal), method = "lm") +
    geom_point(aes(x = nb_09, y = rmse.train_rndcv4k_mammal, color = 4)) +
    geom_smooth(aes(x = nb_09, y= rmse.train_rndcv4k_mammal), method = "lm")

```


```{r}
readr::write_rds(test_rndcv4k_mammal, file = "train_test/test_rndcv4k_mammal.rds")
readr::write_rds(test_spcv4k_mammal, file = "train_test/test_spcv4k_mammal.rds")
readr::write_rds(train_rndcv4k_mammal, file = "train_test/train_rndcv4k_mammal.rds")
readr::write_rds(train_spcv4k_mammal, file = "train_test/train_spcv4k_mammal.rds")

readr::write_rds(test_rndcv4k_mammal, file = "train_test/test_rndcv4k_mammal")
readr::write_rds(test_spcv4k_mammal, file = "train_test/test_rndcv4k_mammal")
readr::write_rds(train_rndcv4k_mammal, file = "train_test/train_rndcv4k_mammal")
readr::write_rds(train_spcv4k_mammal, file = "train_test/train_spcv4k_mammal")

readr::write_rds(test_rndcv4k_amphibian, file = "train_test/test_rndcv4k_amphibian")
readr::write_rds(train_rndcv4k_amphibian, file = "train_test/train_rndcv4k_amphibian")
```











































```{r predict with model : }
pred <- predict(default_model, data = data, type = "cloglog")
map <- predict(default_model, data = predictors, type = "cloglog")
map <- predict(default_model, data = predictors, type = "cloglog", file = "my_map", format = "GTiff")

p <- data@data[data@pa == 1, ]
pred <- predict(default_model, data = p, type = "cloglog")
tail(pred)
```

```{r}


test_xy_mammal <- list()
for (i in 1:length(t.t.mammal_0.2)){test_xy_mammal[[i]] <-
t.t.mammal_0.2[[i]][[2]]@coords %>% mutate(., 
                                           prediction = test_rndcv4k_mammal[[i]],
                                           pa = t.t.mammal_0.2[[i]][[2]]@pa) %>%
  filter(., pa == 1)

}

for (i in 1:length(test_xy_mammal)){
  write_csv(test_xy_mammal[[i]], file = paste0("model/test_xy_",t.t.mammal_0.2[[i]][[2]]@species,".csv"
  ))}

# add residual col

for (i in 1:length(test_xy_mammal)){
  test_xy_mammal[[i]] <- test_xy_mammal[[i]] %>% mutate(., 
                                 residual.1 = 1 - prediction)
}

for (i in 1:length(test_xy_reptile)){
  test_xy_reptile[[i]] <- test_xy_reptile[[i]] %>% mutate(., 
                                 residual.1 = 1 - prediction)
}

##

# fix sequence
for (i in 1:length(test_xy_mammal)){
  test_xy_mammal[[i]] <- test_xy_mammal[[i]][,c(1,2,3,5,4)]
}

for (i in 1:length(test_xy_reptile)){
  test_xy_reptile[[i]] <- test_xy_reptile[[i]][,c(1,2,3,5,4)]
}


#### reptile

test_rndcv4k_reptile <- list()
for (i in 1:length(model_rndcv4k_reptile)){
  test_rndcv4k_reptile[[i]] <- SDMtune::predict(model_rndcv4k_reptile[[i]],
                                               data = t.t.reptile_0.2[[i]][[2]],
                                               type = "cloglog")
}
test_xy_reptile <- list()
for (i in 1:length(t.t.reptile_0.2)){test_xy_reptile[[i]] <-
t.t.reptile_0.2[[i]][[2]]@coords %>% mutate(., 
                                           prediction = test_rndcv4k_reptile[[i]],
                                           pa = t.t.reptile_0.2[[i]][[2]]@pa) %>%
  filter(., pa == 1)

}

for (i in 1:length(test_xy_reptile)){
  write_csv(test_xy_reptile[[i]], file = paste0("model/test_xy_",t.t.reptile_0.2[[i]][[2]]@species,".csv"
                                               ))}
```

```{r}
cat("Training auc: ", auc(maxnet_model))
cat("Testing auc: ", auc(maxnet_model, test = test))

output <- data.frame(matrix(NA, nrow = 10, ncol = 3)) # Create an empty data.frame
colnames(output) <- c("seed", "trainAUC", "testAUC")
set.seed(25)
seeds <- sample.int(1000, 10) # Create 10 different random seeds
for (i in 1:length(seeds)) { # Loop through the seeds
  c(train, test) %<-% trainValTest(data, test = 0.2, seed = seeds[i], only_presence = TRUE) # Make the train/test split
  m <- train("Maxnet", data = train) # train the model
  # Populate the output data.frame
  output[i, 1] <- seeds[i]
  output[i, 2] <- auc(m)
  output[i, 3] <- auc(m, test = test)
}
```

```{r}
t.t.reptile_0.2 <- list()
for (i in 1:length(data.reptile)){t.t.reptile_0.2[[i]] <- trainValTest(data.reptile[[i]], test = 0.2, only_presence = TRUE, seed = 4839)}

rnd4k <- function(x){randomFolds(x[[1]], 
                                 k = 4,
                                 only_presence = TRUE,
                                 seed = 3668)}

folds.rnd4k_reptile <- map(t.t.reptile_0.2, rnd4k)

#h32 67-76
t.t.reptile_0.2.h32 <- list()
for (i in 67:76){
t.t.reptile_0.2.h32[[i-66]] <- t.t.reptile_0.2[[i]]
}

folds.rnd4k_reptile.h32 <- list()
for (i in 67:76){
folds.rnd4k_reptile.h32[[i-66]] <- folds.rnd4k_reptile[[i]]
}

maxnet.train <- function(x, y){train("Maxnet",
                                data = x[[1]],
                                folds = y)}

moedl_rndcv4k_reptile.h32 <- furrr::future_map2(t.t.reptile_0.2.h32, folds.rnd4k_reptile.h32, maxnet)

# h1
t.t.reptile_0.2.h1 <- list()
for (i in 15:25){
t.t.reptile_0.2.h1[[i-14]] <- t.t.reptile_0.2[[i]]
}

folds.rnd4k_reptile.h1 <- list()
for (i in 15:25){
folds.rnd4k_reptile.h1[[i-14]] <- folds.rnd4k_reptile[[i]]
}

moedl_rndcv4k_reptile.h1 <- furrr::future_map2(t.t.reptile_0.2.h1, folds.rnd4k_reptile.h1, maxnet)
```

```{r t.t.amphibian}
t.t.amphibian_0.2 <- list()
for (i in 1:length(data.amphibian)){t.t.amphibian_0.2[[i]] <- trainValTest(data.amphibian[[i]], test = 0.2, only_presence = TRUE, seed = 4839)}

rnd4k <- function(x){randomFolds(x[[1]], 
                                 k = 4,
                                 only_presence = TRUE,
                                 seed = 3668)}

folds.rnd4k_amphibian <- map(t.t.amphibian_0.2, rnd4k)
```

```{r}


h <- list(reg = seq(0.2, 1, 0.1))
exp_1 <- gridSearch(model, hypers = h, metric = "auc", test = val)
```
```{r it's modeling time dismo::kfold}

aves_na_TF <- map_lgl(extract_spatial_aves_furrr, na.omit)

fold5 <- function(x){kfold(x, k=5)}
fold4 <- function(x){kfold(x, k=4)}


fold5_amphibian <- map(list_spatial_amphibian, fold5)
fold5_aves <- map(list_spatial_aves, fold5)
fold5_mammal <- map(list_spatial_mammal_na, fold5)
fold5_reptile <- map(list_spatial_reptile, fold5)

test_fold5_amphibian <- list()
for (i in seq_along(1:length(list_spatial_amphibian))){ 
                      test_fold5_amphibian[[i]] <- list_spatial_amphibian[[i]][fold5_amphibian[[i]] == 1, ]}

test_fold5_aves <- list()
for (i in seq_along(1:length(list_spatial_aves))){ 
                      test_fold5_aves[[i]] <- list_spatial_aves[[i]][fold5_aves[[i]] == 1, ]}

test_fold5_mammal <- list()
for (i in seq_along(1:length(list_spatial_mammal))){ 
                      test_fold5_mammal[[i]] <- list_spatial_mammal_na[[i]][fold5_mammal[[i]] == 1, ]}

train_fold5_amphibian <- list()
for (i in seq_along(1:length(list_spatial_amphibian))){ 
                      train_fold5_amphibian[[i]] <- list_spatial_amphibian[[i]][fold5_amphibian[[i]] != 1, ]}

train_fold5_reptile <- list()
for (i in seq_along(1:length(list_spatial_reptile))){ 
                      train_fold5_reptile[[i]] <- list_spatial_reptile[[i]][fold5_reptile[[i]] != 1, ]}

train_fold5_mammal <- list()
for (i in seq_along(1:length(list_spatial_mammal_na))){ 
                      train_fold5_mammal[[i]] <- list_spatial_mammal_na[[i]][fold5_mammal[[i]] != 1, ]}

```

```{r save the model}
map(train_fold5_reptile, sp::as.)
writeOGR(train_fold5_mammal[[1]], dsn = getwd(), driver = "ESRI Shapefile")

savemodel <- function(x){save(x, file = model_mammal_names)}

map(model_mammal_maxent, savemodel)

model_mammal_names <- paste0("E:/model/mammal/model_",sci_MAMMAL,".rda")


for (i in 1:51){save(model_mammal_maxent[[i]], file =  model_mammal_names[[i]])}

save(model_mammal_maxent[[1]], file = model_mammal_names[1])

saveRDS(model_mammal_maxent, file = "E:/gmm.rda")

save(model_mammal_maxent, list = model_mammal_names, file = "E:/model")

```

```{r it's modeling time: dismo::maxent}

model_amphibian_maxent <-  map2(bios_amphibian ,train_fold5_amphibian, maxent)

model_aves_maxent <- map2(bios_aves, train_fold5_aves, maxent)



model_mammal_maxent <- future_map2(bios_mammal, train_fold5_mammal, maxent)
#Error in .local(x, p, ...) : only got:1random background point values; is there a layer with many NA values?


bios_mammal <- bios_mammal


model_reptile_maxent <- map2(bios_reptile, train_fold5_reptile, maxent)

```

```{r what was important varibles}
map(model_amphibian_maxent, plot)
map(model_aves_maxent, plot)
map(model_mammal_maxent, plot)
```

```{r AUC, ROC, }
rocplot <- function(x){plot(x, 'ROC')} 

```

```{r it's modeling time: dismo::predict}
```

```{r random points for background}
randompoints_amp1 <- randomPoints(bios_amphibian[[1]], 1000)
randompoints_ave1 <- randomPoints(aves_bios[[1]], 1000)

```

```{r }
# extract test point and make it data frame
pointdf_test_fold5_amphibian <- map2(bios_amphibian, test_fold5_amphibian, raster::extract) %>% map(data.frame)
pointdf_test_fold5_aves <- map2(aves_bios, test_fold5_aves, raster::extract) %>% map(data.frame)
pointdf_test_fold5_mammal <- map2(bios_mammal, test_fold5_mammal, raster::extract) %>% map(data.frame)

point_test_fold5_mammal <- map2(bios_mammal, test_fold5_mammal_sf, raster::extract)




#remove god damn NAs... 
extractNA_test_fold5_mammal <- map(test_fold5_mammal_sf,na.omit)



# predict = model + test data
predict <- dismo::predict()

predict_amphibian_maxent <- map2(model_amphibian_maxent, pointdf_test_fold5_amphibian, predict)
predict_aves_maxent <- map2(model_aves_maxent, pointdf_test_fold5_aves, predict) 
predict_mammal_maxent <- map2(model_mammal_maxent, pointdf_test_fold5_mammal, predict)


# as data frame
pred_df_amp <- map(predict_amphibian_maxent, as.data.frame)
pred_df_amp <- map(pred_df_amp) %>% mutate()
pred_df_ave <- map(predict_aves_maxent, as.data.frame)
pred_df_mam <- map(predict_mammal_maxent, as.data.frame)

#save it for future
predict_amphibian_names <- paste0("E:/pred/amphibian/predict_", list_origin_amphibian, ".csv")
predict_aves_names <- paste0("E:/pred/aves/predict_", list_origin_aves, ".csv")

map2(pred_df_amp, predict_amphibian_names, write_csv)
map2(pred_df_ave, predict_aves_names, write_csv)


hist(predict_mammal_maxent[[1]])
```
```{r}
map(predict_mammal_maxent, hist)
```

```{r predict don't have xy. prediction sp to sf, }
as(test_fold5_mammal[[1]], "sf")

test_fold5_mammal_sf <- map2(test_fold5_mammal, "sf", as)

test_fold5_mammal_sf <- test_fold5_mammal_sf %>% map2(point_test_fold5_mammal, mutate)

#convert df and give variable name "pred"
predict_mammal_maxent <- map(predict_mammal_maxent, as.data.frame)

rename_pred <- function(x){rename(x, "pred" = ".x[[i]]")}

predict_mammal_maxent <- map(predict_mammal_maxent, rename_pred)

sf_pred_fold5_mammal <- extractNA_test_fold5_mammal %>% map2(predict_mammal_maxent, mutate)

sf_pred_fold5_mammal <- sf_pred_fold5_mammal %>% map(mutate)

#bio, pred, resid. ALL WE NEED.

sf_resid_fold5_mammal <- list()
for (i in seq_along(1:length(sf_pred_fold5_mammal))){
  sf_resid_fold5_mammal[[i]] <- mutate(sf_pred_fold5_mammal[[i]], residual_1 = 1 - pred,
                                  residual_0.9 = 0.9 - pred,
                                  residual_0.8 = 0.8 - pred,
                                  residual_0.7 = 0.7 - pred,
                                  residual_0.6 = 0.6 - pred,
                                  residual_0.5 = 0.5 - pred)
}





```








add background 
```{r}
numb.back.mammal <- vector()
for(i in 1:length(data.mammal)){
  numb.back.mammal[i] <- sum(data.mammal[[i]]@pa==0)
}
```


```{r}
for(i in 1:length(folds.env_mammal)){
  folds.env_mammal[[i]]$folds[[1]][[1]][length(folds.env_mammal[[1]]$folds[[1]][[1]])+1:length(folds.env_mammal[[1]]$folds[[1]][[1]])+numb.back.mammal[1]] <- 3661:7463
  folds.env_mammal[[1]]$folds[[1]][[2]][3661:7463] <- 3661:7463
  
  folds.env_mammal[[1]]$folds[[2]][[1]][3661:7463] <- 3661:7463
  folds.env_mammal[[1]]$folds[[2]][[2]][3661:7463] <- 3661:7463
  
  folds.env_mammal[[1]]$folds[[3]][[1]][3661:7463] <- 3661:7463
  folds.env_mammal[[1]]$folds[[3]][[2]][3661:7463] <- 3661:7463
  
  folds.env_mammal[[1]]$folds[[4]][[1]][3661:7463] <- 3661:7463
  folds.env_mammal[[1]]$folds[[4]][[2]][3661:7463] <- 3661:7463


```



```{r}
my.env.block <- function (rasterLayer, speciesData, species = NULL, k = 5, standardization = "normal", 
          rasterBlock = TRUE, sampleNumber = 10000, biomod2Format = TRUE, 
          numLimit = 0, verbose = TRUE, ext) 
{
  if (methods::is(speciesData, "SpatialPoints")) {
    speciesData <- sf::st_as_sf(speciesData)
  }
  else if (!methods::is(speciesData, "sf")) {
    stop("speciesData should be a sf or SpatialPoints object")
  }
  if (!is.null(species)) {
    if (species %in% colnames(speciesData) == FALSE) {
      warning("There is no match between the columns name in 'speciesData' and 'species' argument (the response variable).\n")
      species <- NULL
    }
  }
  if (methods::is(rasterLayer, "Raster")) {
    if (raster::nlayers(rasterLayer) >= 1) {
      foldList <- list()
      foldNum <- rep(NA, nrow(speciesData))
      biomodTable <- data.frame(RUN1 = rep(TRUE, nrow(speciesData)))
      if (standardization == "standard") {
        rasterLayer <- standardize(rasterLayer)
      }
      else if (standardization == "normal") {
        rasterLayer <- normalize(rasterLayer)
      }
      else if (standardization == "none" || is.null(standardization)) {
        rasterLayer <- rasterLayer
      }
      if (rasterBlock == TRUE) {
        ext <- raster::extract(rasterLayer, speciesData)
        if (anyNA(ext)) {
          stop("The input raster layer does not cover all the species points.")
        }
        if (raster::ncell(rasterLayer) < 10 * sampleNumber) {
          rp <- length(which(!is.na(raster::values(rasterLayer[[1]]))))
          if (rp < sampleNumber) {
            sampleNumber <- rp
            message("The sampleNumber reduced to ", 
                    sampleNumber, "; the total number of available cells.\n")
          }
        }
        sampr <- raster::sampleRandom(rasterLayer, size = sampleNumber)
        sampr <- sampr[stats::complete.cases(sampr), 
        ]
        sampr <- rbind(ext, sampr)
        kms <- stats::kmeans(sampr, centers = k, iter.max = 500, 
                             nstart = 25)
        speciesData$fold <- kms$cluster[seq_len(nrow(speciesData))]
      }
      else {
        
        if (anyNA(ext)) {
          stop("The input raster layer does not cover all the species points.")
        }
        kms <- stats::kmeans(ext, centers = k, iter.max = 500, 
                             nstart = 25)
        speciesData$fold <- kms$cluster
      }
      if (is.null(species)) {
        trainTestTable <- data.frame(train = rep(0, 
                                                 k), test = 0)
      }
      else {
        cl <- sort(unique(speciesData[, species, drop = TRUE]))
        clen <- length(cl)
        trainTestTable <- as.data.frame(matrix(0, nrow = k, 
                                               ncol = clen * 2))
        names(trainTestTable) <- c(paste("train", cl, 
                                         sep = "_"), paste("test", cl, sep = "_"))
      }
      for (i in seq_len(k)) {
        testSet <- which(speciesData$fold == i)
        trainSet <- which(speciesData$fold != i)
        foldNum[testSet] <- i
        foldList[[i]] <- assign(paste0("fold", i), list(trainSet, 
                                                        testSet))
        if (is.null(species)) {
          trainTestTable$train[i] <- length(trainSet)
          trainTestTable$test[i] <- length(testSet)
        }
        else {
          countrain <- table(speciesData[trainSet, species, 
                                         drop = TRUE])
          countest <- table(speciesData[testSet, species, 
                                        drop = TRUE])
          trainTestTable[i, which(cl %in% names(countrain))] <- countrain
          trainTestTable[i, clen + which(cl %in% names(countest))] <- countest
        }
        if (biomod2Format == TRUE) {
          colm <- paste0("RUN", i)
          biomodTable[, colm] <- FALSE
          biomodTable[trainSet, colm] <- TRUE
        }
      }
      if (any(trainTestTable <= numLimit)) {
        zerofolds <- which(apply(trainTestTable, 1, 
                                 function(x) any(x <= numLimit)))
        if (length(zerofolds) > 1) {
          warning("The folds ", paste(zerofolds, collapse = ", "), 
                  " have class(es) with ", numLimit, " or less records")
        }
        else {
          warning("The fold ", zerofolds, " has class(es) with ", 
                  numLimit, " or less records")
        }
      }
      if (biomod2Format == TRUE) {
        biomodTable <- as.matrix(biomodTable)
        theList <- list(folds = foldList, foldID = foldNum, 
                        biomodTable = biomodTable, k = k, species = species, 
                        records = trainTestTable)
      }
      else {
        theList <- list(folds = foldList, foldID = foldNum, 
                        biomodTable = NULL, k = k, species = species, 
                        records = trainTestTable)
      }
    }
    else stop("'The raster layer is empty!'")
  }
  else stop("The input file is not a valid R raster file")
  if (verbose) 
    print(trainTestTable)
  class(theList) <- c("EnvironmentalBlock")
  return(theList)
}

```

case 1. 

3 climate variable from PC1, PC2, PC3, PC4


```{r}
library(FactoMineR)

PCA_mammal <- list()
for (i in 1:length(data.mammal)){
  PCA_mammal[[i]] <- PCA(data.mammal[[i]]@data)
} 

PCA_mammal.contrib3 <- vector()
for(i in 1:length(PCA_mammal)){
  PCA_mammal.contrib3[i] <- PCA_mammal[[i]]$eig[3,3]
}

PCA_mammal.contrib3
```

```{r}
PCA1234_mammal <- list()
for(i in 1:length(PCA_mammal)){
  PCA1234_mammal[[i]] <- c(names(tail(sort(PCA_mammal[[i]]$var$cos2[,1]), 1)),
                       names(tail(sort(PCA_mammal[[i]]$var$cos2[,2]), 1)),
                       names(tail(sort(PCA_mammal[[i]]$var$cos2[,3]), 1)),
                       names(tail(sort(PCA_mammal[[i]]$var$cos2[,4]), 1)))
}

factoextra::fviz_pca_var(PCA_mammal[[1]], repel = TRUE)


```


```{r}
library(vegan)
library(geosphere)
```


PCNM. Spatial filter

```{r}

geodistm <- function(x){as.dist(geosphere::distm(x@coords, fun = distGeo))}

distm_mammal <- list()
for(i in 1:3){
  distm_mammal[[i]] <- geodistm(data.mammal[[i]])
}


distm_mammal[[1]] <- geodistm(data.mammal[[1]])
distm_mammal[[2]] <- geodistm(data.mammal[[2]])
distm_mammal[[3]] <- geodistm(data.mammal[[3]])
distm_mammal[[4]] <- geodistm(data.mammal[[4]])
distm_mammal[[5]] <- geodistm(data.mammal[[5]])
distm_mammal[[6]] <- geodistm(data.mammal[[6]])
b <- as.dist(geosphere::distm(a, fun = distGeo))

b <- parDist(as.matrix(data.mammal[[3]]@coords), method = "euclidean")

parallelDist::parDist
geosphere::distGeo()
geosphere::distm()

# RcppArmadillo is used as dependency
library(RcppArmadillo)
# Use RcppXPtrUtils for simple usage of C++ external pointers
library(RcppXPtrUtils)
# compile user-defined function and return pointer (RcppArmadillo is used as dependency)
euclideanFuncPtr <- cppXPtr(
"double customDist(const arma::mat &A, const arma::mat &B) {
  return sqrt(arma::accu(arma::square(A - B)));
}", depends = c("RcppArmadillo"))
# distance matrix for user-defined euclidean distance function
# (note that method is set to "custom")
parDist(matrix(1:16, ncol=2), method="custom", func = euclideanFuncPtr)


distm_mammal <- future_map(data.mammal, geodistm)
distm_mammal <- as.dist()

```


```{r}
library(bigmemory)
library(bigdist)
library(disk.frame)
library(parallelDist)



a <- data.mammal[[3]]@coords 
 data.frame(label = 1:nrow(a),
                 latitude = a$Y,
                 longitude = a$X) -> a
 
write_tsv(a, "abc.txt")

a <- as.disk.frame(a, overwrite = T)

a[12,2]
```
```{r}
big.mammal.3 <- bigdist::bigdist(as.matrix(data.mammal[[3]]@coords), file = file.path("D:", "mammal.4"), method = "euclidean")
```


```{r}
library(parallel)
cl <- makePSOCKcluster(30)
setDefaultCluster(cl)
adder <- function(a, b) a + b
clusterExport(NULL, c('adder'))
parLapply(NULL, 1:8, function(z) adder(z, 100))

clusterEvalQ(NULL, library(MASS))
```







```{r}

```



###############################################
variable selection with PCNM

```{r}
SDMtune::varSel(model = ,metric = "auc",bg4cor = ,)
```

```{r}
maxent.train <- function(x, y){SDMtune::train("Maxent",
                                               data = x[[1]],
                                               folds = y)}

maxent.a1 <- maxent.train(data.amphibian.test, folds[[1]])

SDMtune::maxentVarImp()

```
