==> 3_spatial_dependency.Rmd <==
---
title: "3_spatial_dependency"
author: "JEON"
date: '2021 3 9 '
output: html_document
---

```{r}
library(tidyverse)
library(spdep)
```
```{r}
library(furrr)
future::plan(multisession, workers = 30)
```


```{r}
train_rndcv4k_mammal <- readr::read_rds(file = "train_test/train_rndcv4k_mammal.rds")
test_rndcv4k_mammal <- readr::read_rds(file = "train_test/test_rndcv4k_mammal.rds")
train_spcv4k_mammal <- readr::read_rds(file = "train_test/train_spcv4k_mammal.rds")
test_spcv4k_mammal <- readr::read_rds(file = "train_test/test_spcv4k_mammal.rds")

t.t.mammal_0.2 <- readr::read_rds(file = "t.t/t.t.mammal_0.2.rds")


train_rndcv4k_reptile <- readr::read_rds(file = "train_test/train_rndcv4k_reptile.rds")
test_rndcv4k_reptile <- readr::read_rds(file = "train_test/test_rndcv4k_reptile.rds")
t.t.reptile_0.2 <- readr::read_rds(file = "t.t/t.t.reptile_0.2.rds")

train_rndcv4k_amphibian <- readr::read_rds(file = "train_test/train_rndcv4k_amphibian.rds")
test_rndcv4k_amphibian <- readr::read_rds(file = "train_test/test_rndcv4k_amphibian.rds")
t.t.amphibian_0.2 <- readr::read_rds(file = "t.t/t.t.amphibian_0.2.rds")

```

it's hell
```{r}
rndcv4k_mammal.df <- list()
for(i in 1:length(train_rndcv4k_mammal.df)){
rndcv4k_mammal.df[[i]] <- dplyr::bind_rows(train_rndcv4k_mammal.df[[i]], test_rndcv4k_mammal.df[[i]])
}

rndcv4k_reptile.df <- list()
for(i in 1:length(train_rndcv4k_reptile.df)){
rndcv4k_reptile.df[[i]] <- dplyr::bind_rows(train_rndcv4k_reptile.df[[i]], test_rndcv4k_reptile.df[[i]])
}

sf_rndcv4k_mammal <- map(rndcv4k_mammal.df, sf_xy)


sf_rndcv4k_reptile <- map(rndcv4k_reptile.df, sf_xy)

vario_rndcv4k_mammal <- future_map(sf_rndcv4k_mammal, vario_resid, 
                                .options = furrr_options(seed = TRUE),
                                .progress = T)

vario_rndcv4k_reptile <- future_map(sf_rndcv4k_reptile, vario_resid, 
                                .options = furrr_options(seed = TRUE),
                                .progress = T)

sf_rndcv4k_reptile.good <- list()
sf_rndcv4k_reptile.good <- sf_rndcv4k_reptile[c(2,4,5,7,12,14,17,20,22,23,
                                                25,26,27,28,29,30,32,33,36,37,
                                                39,40,41,43,44,47,48,49,52,57,
                                                61,63,64,66,73,74,75,80,81,86,
                                                87,89,91,93)]

range.vm.sph.4kR.good <- list()
range.vm.sph.4kR.good <- range.vm.sph.4kR[c(2,4,5,7,12,14,17,20,22,23,
                                                25,26,27,28,29,30,32,33,36,37,
                                                39,40,41,43,44,47,48,49,52,57,
                                                61,63,64,66,73,74,75,80,81,86,
                                                87,89,91,93)]






dnn.sph.4kR.good <- furrr::future_map2(sf_rndcv4k_reptile.good, 
                                       range.vm.sph.4kR.good, 
                                       dnn,
                                       .options = furrr_options(seed = TRUE),
                                       .progress = T)

listw_reptile.good <- furrr::future_map(dnn.sph.4kR.good, 
                                           nb2listw_zero, 
                                           .options = furrr_options(seed = TRUE),
                                           .progress = T)


moran_rndsph_reptile   <- future_map2(sf_rndcv4k_reptile.good, 
                         listw_reptile.good, 
                         my.moran.test, 
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

```


```{r}
train_rndcv4k_mammal.df <- list()
for(i in 1:length(train_rndcv4k_mammal)){
train_rndcv4k_mammal.df[[i]] <- slice(data.frame(prediction = train_rndcv4k_mammal[[i]],
                                                long = t.t.mammal_0.2[[i]][[1]]@coords[,1],
                                                lat = t.t.mammal_0.2[[i]][[1]]@coords[,2],
                                                residual = 1 - train_rndcv4k_mammal[[i]])
                                                , 1:sum(t.t.mammal_0.2[[i]][[1]]@pa == 1))
}


test_rndcv4k_mammal.df <- list()
for(i in 1:length(test_rndcv4k_mammal)){
test_rndcv4k_mammal.df[[i]] <- slice(data.frame(prediction = test_rndcv4k_mammal[[i]],
                                   long = t.t.mammal_0.2[[i]][[2]]@coords[,1],
                                   lat = t.t.mammal_0.2[[i]][[2]]@coords[,2],
                                   residual = 1 - test_rndcv4k_mammal[[i]])
                                   , 1:sum(t.t.mammal_0.2[[i]][[2]]@pa == 1))
}



train_rndcv4k_reptile.df <- list()
for(i in 1:length(train_rndcv4k_reptile)){
train_rndcv4k_reptile.df[[i]] <- slice(data.frame(prediction = train_rndcv4k_reptile[[i]],
                                                long = t.t.reptile_0.2[[i]][[1]]@coords[,1],
                                                lat = t.t.reptile_0.2[[i]][[1]]@coords[,2],
                                                residual = 1 - train_rndcv4k_reptile[[i]])
                                                , 1:sum(t.t.reptile_0.2[[i]][[1]]@pa == 1))
}


test_rndcv4k_reptile.df <- list()
for(i in 1:length(test_rndcv4k_reptile)){
test_rndcv4k_reptile.df[[i]] <- slice(data.frame(prediction = test_rndcv4k_reptile[[i]],
                                   long = t.t.reptile_0.2[[i]][[2]]@coords[,1],
                                   lat = t.t.reptile_0.2[[i]][[2]]@coords[,2],
                                   residual = 1 - test_rndcv4k_reptile[[i]])
                                   , 1:sum(t.t.reptile_0.2[[i]][[2]]@pa == 1))
}


train_rndcv4k_amphibian.df <- list()
for(i in 1:length(train_rndcv4k_amphibian)){
train_rndcv4k_amphibian.df[[i]] <- slice(data.frame(prediction = train_rndcv4k_amphibian[[i]],
                                                long = t.t.amphibian_0.2[[i]][[1]]@coords[,1],
                                                lat = t.t.amphibian_0.2[[i]][[1]]@coords[,2],
                                                residual = 1 - train_rndcv4k_amphibian[[i]])
                                                , 1:sum(t.t.amphibian_0.2[[i]][[1]]@pa == 1))
}


test_rndcv4k_amphibian.df <- list()
for(i in 1:length(test_rndcv4k_amphibian)){
test_rndcv4k_amphibian.df[[i]] <- slice(data.frame(prediction = test_rndcv4k_amphibian[[i]],
                                   long = t.t.amphibian_0.2[[i]][[2]]@coords[,1],
                                   lat = t.t.amphibian_0.2[[i]][[2]]@coords[,2],
                                   residual = 1 - test_rndcv4k_amphibian[[i]])
                                   , 1:sum(t.t.amphibian_0.2[[i]][[2]]@pa == 1))
}





```




```{r semivariogram}
library(gstat)

vario_pred <- function(x){gstat::variogram(prediction~1, x)}
vario_resid <- function(x){gstat::variogram(residual~1, x)}

sf_xy <- function(x){sf::st_as_sf(x, coords = c("long", "lat"), crs = 4326)}
sf_train_rndcv4k_mammal <- map(train_rndcv4k_mammal.df, sf_xy)
sf_test_rndcv4k_mammal <- map(test_rndcv4k_mammal.df, sf_xy)

vario_rndcv4k_mammal <- future_map(sf_test_rndcv4k_mammal, vario_resid, 
                                .options = furrr_options(seed = TRUE),
                                .progress = T)

sf_train_rndcv4k_reptile <- map(train_rndcv4k_reptile.df, sf_xy)
sf_test_rndcv4k_reptile <- map(test_rndcv4k_reptile.df, sf_xy)

vario_rndcv4k_reptile <- future_map(sf_test_rndcv4k_reptile, vario_resid, 
                                .options = furrr_options(seed = TRUE),
                                .progress = T)

sf_train_rndcv4k_amphibian <- map(train_rndcv4k_amphibian.df, sf_xy)
sf_test_rndcv4k_amphibian <- map(test_rndcv4k_amphibian.df, sf_xy)

vario_rndcv4k_amphibian <- future_map(sf_test_rndcv4k_amphibian, vario_resid, 
                                .options = furrr_options(seed = TRUE),
                                .progress = T)



####
plot_variogram <- function(v, m) {
  preds = variogramLine(m, maxdist = max(v$dist))
  ggplot() + 
    geom_point(data = v, aes(x = dist, y = gamma, size=np)) +
    geom_line(data = preds, aes(x = dist, y = gamma)) +
    theme(legend.key.size = unit(0.00005, 'cm'))
}

v <- variogram(zinc ~ 1, meuse)
m <- fit.variogram(v, vgm(c("Exp", "Sph")))
plot_variogram(v, m) 
###

vario.plot.4km.sph <- map2(vario_rndcv4k_mammal, vmcv4k.sph_mammal, plot_variogram)
vario.plot.4km.gau <- map2(vario_mammal_cv4k, vmcv4k.gau_mammal, plot_variogram)
vario.plot.4km.exp <- map2(vario_mammal_cv4k, vmcv4k.exp_mammal, plot_variogram)
vario.plot.4km.mat <- map2(vario_mammal_cv4k, vmcv4k.mat_mammal, plot_variogram)
vario.plot.4km.exc <- map2(vario_mammal_cv4k, vmcv4k.exc_mammal, plot_variogram)
vario.plot.4km.cir <- map2(vario_mammal_cv4k, vmcv4k.cir_mammal, plot_variogram)
vario.plot.4km.lin <- map2(vario_mammal_cv4k, vmcv4k.lin_mammal, plot_variogram)
vario.plot.4km.bes <- map2(vario_mammal_cv4k, vmcv4k.bes_mammal, plot_variogram)
vario.plot.4km.pen <- map2(vario_mammal_cv4k, vmcv4k.pen_mammal, plot_variogram)
vario.plot.4km.per <- map2(vario_mammal_cv4k, vmcv4k.per_mammal, plot_variogram)
vario.plot.4km.wav <- map2(vario_mammal_cv4k, vmcv4k.wav_mammal, plot_variogram)
vario.plot.4km.hol <- map2(vario_mammal_cv4k, vmcv4k.hol_mammal, plot_variogram)
vario.plot.4km.spl <- map2(vario_mammal_cv4k, vmcv4k.spl_mammal, plot_variogram)
vario.plot.4km.leg <- map2(vario_mammal_cv4k, vmcv4k.leg_mammal, plot_variogram)
vario.plot.4km.err <- map2(vario_mammal_cv4k, vmcv4k.err_mammal, plot_variogram)



#ggplot variogram
gg.vario <- function(x){ggplot(x,aes(x=dist,y=gamma,size=np)) + geom_point()}

vario.plot.4km.sph <- map(vario_mammal_cv4k, gg.vario)


#Check variogram


vario_sph <- function(x){fit.variogram(x, vgm("Sph"))}
vario_exp <- function(x){fit.variogram(x, vgm("Exp"))}
vario_gau <- function(x){fit.variogram(x, vgm("Gau"))}
vario_mat <- function(x){fit.variogram(x, vgm("Mat"))}
vario_exc <- function(x){fit.variogram(x, vgm("Exc"))}
vario_stemat <- function(x){fit.variogram(x, vgm("Ste Mat"))}
vario_cir <- function(x){fit.variogram(x, vgm("Cir"))}
vario_lin <- function(x){fit.variogram(x, vgm("Lin"))}
vario_bes <- function(x){fit.variogram(x, vgm("Bes"))}
vario_pen <- function(x){fit.variogram(x, vgm("Pen"))}
vario_per <- function(x){fit.variogram(x, vgm("Per"))}
vario_wav <- function(x){fit.variogram(x, vgm("Wav"))}
vario_hol <- function(x){fit.variogram(x, vgm("Hol"))}
vario_log <- function(x){fit.variogram(x, vgm("Log"))}
vario_pow <- function(x){fit.variogram(x, vgm("Pow"))}
vario_spl <- function(x){fit.variogram(x, vgm("Spl"))}
vario_leg <- function(x){fit.variogram(x, vgm("Leg"))}
vario_err <- function(x){fit.variogram(x, vgm("Err"))}
vario_int <- function(x){fit.variogram(x, vgm("Int"))}


vmcv4k.sph_mammal <- map(vario_rndcv4k_mammal, vario_sph)
vmcv4k.exp_mammal <- map(vario_mammal_cv4k, vario_exp)
vmcv4k.gau_mammal <- map(vario_mammal_cv4k, vario_gau)
vmcv4k.mat_mammal <- map(vario_mammal_cv4k, vario_mat)
vmcv4k.exc_mammal <- map(vario_mammal_cv4k, vario_exc)
vmcv4k.stemat_mammal <- map(vario_mammal_cv4k, vario_stemat)
vmcv4k.cir_mammal <- map(vario_mammal_cv4k, vario_cir)
vmcv4k.lin_mammal <- map(vario_mammal_cv4k, vario_lin)
vmcv4k.bes_mammal <- map(vario_mammal_cv4k, vario_bes)
vmcv4k.pen_mammal <- map(vario_mammal_cv4k, vario_pen)
vmcv4k.per_mammal <- map(vario_mammal_cv4k, vario_per)
vmcv4k.wav_mammal <- map(vario_mammal_cv4k, vario_wav)
vmcv4k.hol_mammal <- map(vario_mammal_cv4k, vario_hol)
vmcv4k.log_mammal <- map(vario_mammal_cv4k, vario_log)
vmcv4k.pow_mammal <- map(vario_mammal_cv4k, vario_pow)
vmcv4k.spl_mammal <- map(vario_mammal_cv4k, vario_spl)
vmcv4k.leg_mammal <- map(vario_mammal_cv4k, vario_leg)
vmcv4k.err_mammal <- map(vario_mammal_cv4k, vario_err)
vmcv4k.int_mammal <- map(vario_mammal_cv4k, vario_int)

range.vm.sph.4kM <- map(vmcv4k.sph_mammal, c(3,2))

vmcv4k.sph_reptile <- map(vario_rndcv4k_reptile, vario_sph)
range.vm.sph.4kR <- map(vmcv4k.sph_reptile, c(3,2))

vmcv4k.sph_amphibian <- map(vario_rndcv4k_amphibian, vario_sph)
range.vm.sph.4kA <- map(vmcv4k.sph_amphibian, c(3,2))


range.vm.sph.4kM <- data.frame(range = unlist(map(vmcv4k.sph_mammal, c(3,2))))
range.vm.exp.4kM <- data.frame(range = unlist(map(vmcv4k.exp_mammal, c(3,2))))
range.vm.gau.4kM <- data.frame(range = unlist(map(vmcv4k.gau_mammal, c(3,2))))
range.vm.mat.4kM <- data.frame(range = unlist(map(vmcv4k.mat_mammal, c(3,2))))

plot.vario.sph <- function(x){plot(x, model=fit.variogram(x, vgm("Sph")))}
plot.vario.exp <- function(x){plot(x, model=fit.variogram(x, vgm("Exp")))}
plot.vario.gau <- function(x){plot(x, model=fit.variogram(x, vgm("Gau")))}
plot.vario.mat <- function(x){plot(x, model=fit.variogram(x, vgm("Mat")))}

vario.plot.4km.sph <- map(vario_mammal_cv4k, plot.vario.sph)
vario.plot.4km.gau <- map(vario_mammal_cv4k, plot.vario.gau)
vario.plot.4km.sph <- map(vario_mammal_cv4k, plot.vario)
vario.plot.4km.sph <- map(vario_mammal_cv4k, plot.vario)

vario.plot.4kR.sph <- map(vario_rndcv4k_reptile, plot.vario.sph)

library(gridExtra)

a<- grid.arrange(vario.plot.4km.sph[[1]], vario.plot.4km.sph[[2]], vario.plot.4km.sph[[3]], vario.plot.4km.sph[[4]],vario.plot.4km.sph[[5]], vario.plot.4km.sph[[6]],vario.plot.4km.sph[[7]], vario.plot.4km.sph[[8]],vario.plot.4km.sph[[9]], vario.plot.4km.sph[[10]],vario.plot.4km.sph[[11]], vario.plot.4km.sph[[12]],vario.plot.4km.sph[[13]], vario.plot.4km.sph[[14]],vario.plot.4km.sph[[15]], vario.plot.4km.sph[[16]],vario.plot.4km.sph[[17]], vario.plot.4km.sph[[18]],vario.plot.4km.sph[[19]], vario.plot.4km.sph[[20]],vario.plot.4km.sph[[21]], vario.plot.4km.sph[[22]],vario.plot.4km.sph[[23]], vario.plot.4km.sph[[24]],vario.plot.4km.sph[[25]], vario.plot.4km.sph[[26]],vario.plot.4km.sph[[27]], vario.plot.4km.sph[[28]],vario.plot.4km.sph[[29]], vario.plot.4km.sph[[30]],vario.plot.4km.sph[[31]], vario.plot.4km.sph[[32]],vario.plot.4km.sph[[33]], vario.plot.4km.sph[[34]],vario.plot.4km.sph[[35]])


cowplot::plot_grid(vario.plot.4kR.sph[[71]], vario.plot.4kR.sph[[72]], vario.plot.4kR.sph[[73]], vario.plot.4kR.sph[[74]],vario.plot.4kR.sph[[75]], vario.plot.4kR.sph[[76]],vario.plot.4kR.sph[[77]], vario.plot.4kR.sph[[78]],vario.plot.4kR.sph[[79]], vario.plot.4kR.sph[[80]],vario.plot.4kR.sph[[81]], vario.plot.4kR.sph[[82]],vario.plot.4kR.sph[[83]], vario.plot.4kR.sph[[84]],vario.plot.4kR.sph[[85]], vario.plot.4kR.sph[[86]],vario.plot.4kR.sph[[87]], vario.plot.4kR.sph[[88]],vario.plot.4kR.sph[[89]], vario.plot.4kR.sph[[90]],vario.plot.4kR.sph[[91]], vario.plot.4kR.sph[[92]],vario.plot.4kR.sph[[93]], vario.plot.4kR.sph[[94]], 
               ncol = 6, nrow = 6, label_size = 5)

pdf("test2.pdf")

a <- grid.arrange(vario.plot.4km.sph[[1]], vario.plot.4km.sph[[2]], vario.plot.4km.sph[[3]], vario.plot.4km.sph[[4]], ncol = 2, nrow = 2)



savePlot(vario.plot.4km.gau[[1]], filename = "test.png", type = "png")
ggsave(filename = "test.png", plot = vario.plot.4km.gau[[1]])

all_vario

library(ncf)

ncf::correlog(test_xy_mammal[[1]][,1], test_xy_mammal[[1]][,2], test_xy_mammal[[1]][,3],
              resamp = 10, increment = 1000)
```


```{r}
dnn.sph.4kM.test.good <- vector("list", length = 19)
dnn.sph.4kM.test.good <- dnn.sph.4kM.test[c(1,2,5,7,9,10,12,13,17,18,19,20,22,23,24,25,27,30,33,34)]




sf_test_rndcv4k_mammal.good <- list()
sf_test_rndcv4k_mammal.good <-  sf_test_rndcv4k_mammal[c(1,5,7,9,10,12,13,17,18,19,20,22,23,24,25,27,30,33,34)]

```




```{r}

library(spdep)

dnn.sph.4kM[[1]] <- spdep::dnearneigh(a, 0, 168.4539)

moran()

```



```{r}
library(spatstat)
library(geosphere)





# most far away
a <- vector()
for(i in 1:length(test_rndcv4k_mammal.df)){
a[i] <- which.max(spatstat.geom::nndist(X = test_rndcv4k_mammal.df[[i]]$long,
                                                              Y = test_rndcv4k_mammal.df[[i]]$lat,
                                                              k = 1))}


# most far away's neightbor

b <- vector()
for( i in 1:length(test_rndcv4k_mammal.df)){
b[i] <- spatstat.geom::nnwhich(X = test_rndcv4k_mammal.df[[i]]$long,
                       Y = test_rndcv4k_mammal.df[[i]]$lat,
                       k = 1)[a[i]]}


# Their distance 
c <- vector()
for(i in 1:length(test_rndcv4k_mammal.df)){
c[i] <- geosphere::distGeo(test_rndcv4k_mammal.df[[i]][a[i],2:3], test_rndcv4k_mammal.df[[i]][b[i],2:3])
}
c <- c/1000





dnn.min.4kM.test <- furrr::future_map2(sf_test_rndcv4k_mammal, 
                                       c, 
                                       dnn,
                                       .options = furrr_options(seed = TRUE),
                                       .progress = T)

listw.min_test_mammal <- furrr::future_map(dnn.min.4kM.test, 
                                           nb2listw_zero, 
                                           .options = furrr_options(seed = TRUE),
                                           .progress = T)

moran_rnd.min_mammal   <- future_map2(sf_test_rndcv4k_mammal, 
                         listw.min_test_mammal, 
                         my.moran.test, 
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

moran_rnd.min_M <- vector()
for(i in 1:length(moran_rnd.min_mammal)){
moran_rnd.min_M[i] <- moran_rnd.min_mammal[[i]][3][[1]][[1]]
}


#

a <- vector()
for(i in 1:length(test_rndcv4k_reptile.df)){
a[i] <- which.max(spatstat.geom::nndist(X = test_rndcv4k_reptile.df[[i]]$long,
                                                              Y = test_rndcv4k_reptile.df[[i]]$lat,
                                                              k = 1))}


# most far away's neightbor

b <- vector()
for( i in 1:length(test_rndcv4k_reptile.df)){
b[i] <- spatstat.geom::nnwhich(X = test_rndcv4k_reptile.df[[i]]$long,
                       Y = test_rndcv4k_reptile.df[[i]]$lat,
                       k = 1)[a[i]]}


# Their distance 
c <- vector()
for(i in 1:length(test_rndcv4k_reptile.df)){
c[i] <- geosphere::distGeo(test_rndcv4k_reptile.df[[i]][a[i],2:3], test_rndcv4k_reptile.df[[i]][b[i],2:3])
}
c <- c/1000

dnn.min.4kR.test <- furrr::future_map2(sf_test_rndcv4k_reptile, 
                                       c, 
                                       dnn,
                                       .options = furrr_options(seed = TRUE),
                                       .progress = T)

listw.min_test_reptile <- furrr::future_map(dnn.min.4kR.test, 
                                           nb2listw_zero, 
                                           .options = furrr_options(seed = TRUE),
                                           .progress = T)

moran_rnd.min_reptile   <- future_map2(sf_test_rndcv4k_reptile, 
                         listw.min_test_reptile, 
                         my.moran.test, 
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

moran_rnd.min_R <- vector()
for(i in 1:length(moran_rnd.min_reptile)){
moran_rnd.min_R[i] <- moran_rnd.min_reptile[[i]][3][[1]][[1]]
}

#


a <- vector()
for(i in 1:length(test_rndcv4k_amphibian.df)){
a[i] <- which.max(spatstat.geom::nndist(X = test_rndcv4k_amphibian.df[[i]]$long,
                                                              Y = test_rndcv4k_amphibian.df[[i]]$lat,
                                                              k = 1))}


# most far away's neightbor

b <- vector()
for( i in 1:length(test_rndcv4k_amphibian.df)){
b[i] <- spatstat.geom::nnwhich(X = test_rndcv4k_amphibian.df[[i]]$long,
                       Y = test_rndcv4k_amphibian.df[[i]]$lat,
                       k = 1)[a[i]]}


# Their distance 
c <- vector()
for(i in 1:length(test_rndcv4k_amphibian.df)){
c[i] <- geosphere::distGeo(test_rndcv4k_amphibian.df[[i]][a[i],2:3], test_rndcv4k_amphibian.df[[i]][b[i],2:3])
}
c <- c/1000

dnn.min.4kA.test <- furrr::future_map2(sf_test_rndcv4k_amphibian, 
                                       c, 
                                       dnn,
                                       .options = furrr_options(seed = TRUE),
                                       .progress = T)

listw.min_test_amphibian <- furrr::future_map(dnn.min.4kA.test, 
                                           nb2listw_zero, 
                                           .options = furrr_options(seed = TRUE),
                                           .progress = T)

moran_rnd.min_amphibian   <- future_map2(sf_test_rndcv4k_amphibian, 
                         listw.min_test_amphibian, 
                         my.moran.test, 
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

moran_rnd.min_A <- vector()
for(i in 1:length(moran_rnd.min_amphibian)){
moran_rnd.min_A[i] <- moran_rnd.min_amphibian[[i]][3][[1]][[1]]
}




```



```{r get your Spatial Autocorrelation with gstat:: }
#make neightbour to check Moran's I 
library(spdep)

#1. Variogram range based neighborhood 

dnn <- function(x, y){spdep::dnearneigh(x, 0, y, longlat = TRUE)}

dnn.sph.4kM.test <- furrr::future_map2(sf_test_rndcv4k_mammal, 
                                       range.vm.sph.4kM, 
                                       dnn,
                                       .options = furrr_options(seed = TRUE),
                                       .progress = T)

dnn.sph.4kR.test <- furrr::future_map2(sf_test_rndcv4k_reptile, 
                                       range.vm.sph.4kR, 
                                       dnn,
                                       .options = furrr_options(seed = TRUE),
                                       .progress = T)

dnn.sph.4kA.test <- furrr::future_map2(sf_test_rndcv4k_amphibian,
                                       range.vm.sph.4kA, 
                                       dnn,
                                       .options = furrr_options(seed = TRUE),
                                       .progress = T)

nb2listw_zero <- function(x){nb2listw(x, zero.policy = T)}



listw_test_mammal <- furrr::future_map(dnn.sph.4kM.test, 
                                           nb2listw_zero, 
                                           .options = furrr_options(seed = TRUE),
                                           .progress = T)

listw_test_reptile <- furrr::future_map(dnn.sph.4kR.test, 
                                           nb2listw_zero, 
                                           .options = furrr_options(seed = TRUE),
                                           .progress = T)

listw_test_amphibian <- furrr::future_map(dnn.sph.4kA.test, 
                                           nb2listw_zero, 
                                           .options = furrr_options(seed = TRUE),
                                           .progress = T)


my.moran.test <- function(x,y){spdep::moran.test(x$residual, listw = y, zero.policy = T)}



moran_rndsph_mammal   <- future_map2(sf_test_rndcv4k_mammal, 
                         listw_test_mammal, 
                         my.moran.test, 
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

moran_rndsph_reptile   <- future_map2(sf_test_rndcv4k_reptile, 
                         listw_test_reptile, 
                         my.moran.test, 
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

moran_rndsph_amphibian   <- future_map2(sf_test_rndcv4k_amphibian, 
                         listw_test_amphibian, 
                         my.moran.test, 
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

moran_rndsph_M <- vector()
for(i in 1:length(moran_rndsph_mammal)){
moran_rndsph_M[i] <- moran_rndsph_mammal[[i]][3][[1]][[1]]
}

moran_rndsph_R <- vector()
for(i in 1:length(moran_rndsph_reptile)){
moran_rndsph_R[i] <- moran_rndsph_reptile[[i]][3][[1]][[1]]
}

moran_rndsph_A <- vector()
for(i in 1:length(moran_rndsph_amphibian)){
moran_rndsph_A[i] <- moran_rndsph_amphibian[[i]][3][[1]][[1]]
}

result_mammal <- tibble(rsac_rndcv4k_sph = moran_rndsph_M)
result_reptile <- tibble(rsac_rndcv4k_sph = moran_rndsph_R)
result_amphibian <- tibble(rsac_rndcv4k_sph = moran_rndsph_A)

#set true for fucking stupid error and its old style function
set.ZeroPolicyOption(TRUE)




```

```{r}
result_mammal <- readr::read_csv("result_mammal.csv")

result_mammal %>% mutate(log_train_rndcv4k = map_dbl(map(map(train_rndcv4k_mammal.df,4), log),mean),
                         log_test_rndcv4k = map_dbl(map(map(test_rndcv4k_mammal.df,4), log),mean)) -> result_mammal

result_mammal %>% mutate(log_train_rndcv4k.median = map_dbl(map(map(train_rndcv4k_mammal.df,4), log),median),
                         log_test_rndcv4k.median = map_dbl(map(map(test_rndcv4k_mammal.df,4), log),median)) -> result_mammal


result_mammal %>% mutate(resid_train_rndcv4k = map_dbl(map(train_rndcv4k_mammal.df,4), mean),
                         resid_test_rndcv4k = map_dbl(map(test_rndcv4k_mammal.df,4), mean)) -> result_mammal


result_reptile %>% mutate(resid_train_rndcv4k = map_dbl(map(train_rndcv4k_reptile.df,4), mean),
                         resid_test_rndcv4k = map_dbl(map(test_rndcv4k_reptile.df,4), mean)) -> result_reptile

result_amphibian <- readr::read_csv("result_amphibian.csv")
result_amphibian %>% mutate(resid_train_rndcv4k = map_dbl(map(train_rndcv4k_amphibian.df,4), mean),
                         resid_test_rndcv4k = map_dbl(map(test_rndcv4k_amphibian.df,4), mean)) -> result_amphibian


readr::write_csv(result_mammal, file = "result_mammal.csv")
readr::write_csv(result_reptile, file = "result_reptile.csv")
readr::write_csv(result_amphibian, file = "result_amphibian.csv")

```

# raster DNN

```{r}
library(spdep)
library(raster)
library(furrr)
future::plan(multisession, workers = 30)

biopath <- paste0("bios/10m/", list.files(pattern="*.tif", path = "bios/10m/"))
stackbio <- raster::stack(biopath)

bbox_mammal <- readr::read_rds(file = "bbox_mammal.rds")
bbox_reptile <- readr::read_rds(file = "bbox_reptile.rds")
bbox_amphibian <- readr::read_rds(file = "bbox_amphibian.rds")
data.mammal <- readr::read_rds(file = "data/data.mammal.rds")
name.M <- list()
  for(i in 1:length(data.mammal)){
    name.M[[i]] <- paste0("bios/crop/10m/",data.mammal[[i]]@species,".tif")}

stackbio_M <- list()
for(i in 1:length(bbox_mammal)){stackbio_M[[i]] <- stackbio}
stackbio_R <- list()
for(i in 1:length(bbox_reptile)){stackbio_R[[i]] <- stackbio}
stackbio_A <- list()
for(i in 1:length(bbox_amphibian)){stackbio_A[[i]] <- stackbio}

crop <- raster::crop

cropbio_M <- furrr::future_map2(stackbio_M, bbox_mammal, crop)
wr <- function(x, y){raster::writeRaster(x, filename = y)}
furrr::future_map2(cropbio_M, name.M ,wr)

a <-raster::raster("test.tif")
b <- elsa::moran(a, 1, 90)


spdep::cell2nb(stackbio_M[[1]])
a <- spdep::dnearneigh(raster::as.matrix(cropbio_M[[1]]), 1, range.vm.sph.4kM[1,])
b <- spdep::nb2listwdist(a)

sf::st_as_sf(stackbio_M[[1]])




```

```{r}
library(glmnet)
library(maxnet)
biopath <- paste0("bios/crop/", list.files(pattern="*.tif", path = "bios/crop/"))

bio_mammal <- list()
for(i in 1:length(biopath)){
bio_mammal[[i]] <- raster::stack(biopath[i])
}
for(i in 1:length(bio_mammal)){
names(cropbio_M[[i]]) <- c("wc2.1_30s_bio_01",  "wc2.1_30s_bio_02",  "wc2.1_30s_bio_03",  "wc2.1_30s_bio_04", "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07",  "wc2.1_30s_bio_08", "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12", "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16", "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")}

model_rndcv4k_mammal <- readr::read_rds("model/model_rndcv4k_mammal.rds")
sdm.predict <- function(x,y){SDMtune::predict(x, y, type = "cloglog")}

raster.rndcv4k.mammal <- map2(model_rndcv4k_mammal, cropbio_M, sdm.predict)

name.M <- list()
  for(i in 1:length(data.mammal)){
    name.M[[i]] <- paste0("bios/pred/10m/",data.mammal[[i]]@species,".tif")}


pred_mammal <- paste0("bios/pred/10m/",list.files(path = "bios/pred/10m", pattern = "*.tif"))

predict_mammal <- list()
for(i in 1:length(pred_mammal)){
predict_mammal[[i]] <- raster::raster(pred_mammal[i])}

my.moran <- function(x){Moran(x, w=matrix(c(1,1,1,1,0,1,1,1,1), 3,3))}

moran.mammal <- future_map(predict_mammal, my.moran)


pred_mammal <- list()
for(i in 1:length(predict_mammal)){
  pred_mammal[[i]] <- raster::rasterToPoints(predict_mammal[[i]])
}

pred_mammal <- map(pred_mammal, as.data.frame)

sf_df <- function(x){sf::st_as_sf(x, 
                                  coords = c("x", "y"), 
                                  crs = 4326,
                                  remove = T)
}

pred_mammal <- map(pred_mammal, sf_df)

for(i in 1:length(pred_mammal)){
names(pred_mammal[[i]]) <- c("prediction","geometry")}

dnn.pred.mammal <- furrr::future_map2(pred_mammal, range.vm.sph.4kM ,dnn)








raster::Moran()
ncf::correlog(test_xy_mammal[[1]][,1], test_xy_mammal[[1]][,2], test_xy_mammal[[1]][,3],
              resamp = 10, increment = 1000)

```


```{r Moran's Correlogram}

  
a <- sp.correlogram(knb5_mammal[[6]], sf_resid_fold5_mammal[[6]]$residual_1, order = 9, method = "I")
plot(a)


```

```{r tmap visualize}
#Some variogram is weird. Checking tmap

library(tmap)
tm_shape(sf_resid_fold5_mammal[[1]]) + tm_dots("pred")
tm_shape(sf_resid_fold5_mammal[[2]]) + tm_dots("pred")
tm_shape(sf_resid_fold5_mammal[[13]]) + tm_dots("pred")
#4 is fucked up 
tm_shape(sf_resid_fold5_mammal[[14]]) + tm_dots("pred")

```


```{r Y axis, Moran's I with different "K", different "Threshold"}

mammal_Y_moran <- tibble::tibble(moran_0.5k1 = unlist(map(moran_mammal_0.5k1, c(3,1))),
                                  moran_0.5k2 = unlist(map(moran_mammal_0.5k2, c(3,1))),
                                  moran_0.5k3 = unlist(map(moran_mammal_0.5k3, c(3,1))),
                                  moran_0.5k4 = unlist(map(moran_mammal_0.5k4, c(3,1))),
                                  moran_0.5k5 = unlist(map(moran_mammal_0.5k5, c(3,1))),
                                  moran_0.5k6 = unlist(map(moran_mammal_0.5k6, c(3,1))),
                                  moran_0.5k7 = unlist(map(moran_mammal_0.5k7, c(3,1))),
                                  moran_0.5k8 = unlist(map(moran_mammal_0.5k8, c(3,1))),
                                  moran_0.5k9 = unlist(map(moran_mammal_0.5k9, c(3,1))),
                                  moran_0.5k10 = unlist(map(moran_mammal_0.5k10, c(3,1))),
                                  moran_0.6k1 = unlist(map(moran_mammal_0.6k1, c(3,1))),
                                  moran_0.6k2 = unlist(map(moran_mammal_0.6k2, c(3,1))),
                                  moran_0.6k3 = unlist(map(moran_mammal_0.6k3, c(3,1))),
                                  moran_0.6k4 = unlist(map(moran_mammal_0.6k4, c(3,1))),
                                  moran_0.6k5 = unlist(map(moran_mammal_0.6k5, c(3,1))),
                                  moran_0.6k6 = unlist(map(moran_mammal_0.6k6, c(3,1))),
                                  moran_0.6k7 = unlist(map(moran_mammal_0.6k7, c(3,1))),
                                  moran_0.6k8 = unlist(map(moran_mammal_0.6k8, c(3,1))),
                                  moran_0.6k9 = unlist(map(moran_mammal_0.6k9, c(3,1))),
                                  moran_0.6k10 = unlist(map(moran_mammal_0.6k10, c(3,1))),
                                  moran_0.7k1 = unlist(map(moran_mammal_0.7k1, c(3,1))),
                                  moran_0.7k2 = unlist(map(moran_mammal_0.7k2, c(3,1))),
                                  moran_0.7k3 = unlist(map(moran_mammal_0.7k3, c(3,1))),
                                  moran_0.7k4 = unlist(map(moran_mammal_0.7k4, c(3,1))),
                                  moran_0.7k5 = unlist(map(moran_mammal_0.7k5, c(3,1))),
                                  moran_0.7k6 = unlist(map(moran_mammal_0.7k6, c(3,1))),
                                  moran_0.7k7 = unlist(map(moran_mammal_0.7k7, c(3,1))),
                                  moran_0.7k8 = unlist(map(moran_mammal_0.7k8, c(3,1))),
                                  moran_0.7k9 = unlist(map(moran_mammal_0.7k9, c(3,1))),
                                  moran_0.7k10 = unlist(map(moran_mammal_0.7k10, c(3,1))),
                                  moran_0.8k1 = unlist(map(moran_mammal_0.8k1, c(3,1))),
                                  moran_0.8k2 = unlist(map(moran_mammal_0.8k2, c(3,1))),
                                  moran_0.8k3 = unlist(map(moran_mammal_0.8k3, c(3,1))),
                                  moran_0.8k4 = unlist(map(moran_mammal_0.8k4, c(3,1))),
                                  moran_0.8k5 = unlist(map(moran_mammal_0.8k5, c(3,1))),
                                  moran_0.8k6 = unlist(map(moran_mammal_0.8k6, c(3,1))),
                                  moran_0.8k7 = unlist(map(moran_mammal_0.8k7, c(3,1))),
                                  moran_0.8k8 = unlist(map(moran_mammal_0.8k8, c(3,1))),
                                  moran_0.8k9 = unlist(map(moran_mammal_0.8k9, c(3,1))),
                                  moran_0.8k10 = unlist(map(moran_mammal_0.8k10, c(3,1))),
                                  moran_0.9k1 = unlist(map(moran_mammal_0.9k1, c(3,1))),
                                  moran_0.9k2 = unlist(map(moran_mammal_0.9k2, c(3,1))),
                                  moran_0.9k3 = unlist(map(moran_mammal_0.9k3, c(3,1))),
                                  moran_0.9k4 = unlist(map(moran_mammal_0.9k4, c(3,1))),
                                  moran_0.9k5 = unlist(map(moran_mammal_0.9k5, c(3,1))),
                                  moran_0.9k6 = unlist(map(moran_mammal_0.9k6, c(3,1))),
                                  moran_0.9k7 = unlist(map(moran_mammal_0.9k7, c(3,1))),
                                  moran_0.9k8 = unlist(map(moran_mammal_0.9k8, c(3,1))),
                                  moran_0.9k9 = unlist(map(moran_mammal_0.9k9, c(3,1))),
                                  moran_0.9k10 = unlist(map(moran_mammal_0.9k10, c(3,1))),
                                  moran_1k1 = unlist(map(moran_mammal_1k1, c(3,1))),
                                  moran_1k2 = unlist(map(moran_mammal_1k2, c(3,1))),
                                  moran_1k3 = unlist(map(moran_mammal_1k3, c(3,1))),
                                  moran_1k4 = unlist(map(moran_mammal_1k4, c(3,1))),
                                  moran_1k5 = unlist(map(moran_mammal_1k5, c(3,1))),
                                  moran_1k6 = unlist(map(moran_mammal_1k6, c(3,1))),
                                  moran_1k7 = unlist(map(moran_mammal_1k7, c(3,1))),
                                  moran_1k8 = unlist(map(moran_mammal_1k8, c(3,1))),
                                  moran_1k9 = unlist(map(moran_mammal_1k9, c(3,1))),
                                  moran_1k10 = unlist(map(moran_mammal_1k10, c(3,1))),
                                  nb_0.05 = nb_mammal,
                                  moran_1dnb = unlist(map(moran_mammal_1dnb, c(3,1))),
                                  moran_0.9dnb = unlist(map(moran_mammal_0.9dnb, c(3,1))),
                                  moran_0.8dnb = unlist(map(moran_mammal_0.8dnb, c(3,1))),
                                  moran_0.7dnb = unlist(map(moran_mammal_0.7dnb, c(3,1))),
                                  moran_0.6dnb = unlist(map(moran_mammal_0.6dnb, c(3,1))),
                                  moran_0.5dnb = unlist(map(moran_mammal_0.5dnb, c(3,1))))



```

```{r X axis, Niche breadth from different Quantile, different ordination}
#from 0.05~0.95 

nb_mammal <- nb_mammal[-31]
nb_mammal <- nb_mammal[-26]

mammal_X_nb <- data.frame(nb_mammal)


```

```{r X + Y...}

mammal_XY <- mutate(mammal_Y_moran, nb_0.05 = mammal_X_nb)
mammal_XY_df <- as.data.frame(mammal_XY)
```

```{r}
p <- ggplot(data = mammal_Y_moran) +
    geom_point(aes(x = nb_0.05, y = moran_1k1, color = 1)) +
    geom_smooth(aes(x = nb_0.05, y= moran_1k10), method = "lm") +
    geom_point(aes(x = nb_0.05, y = moran_1k2, color = 2)) +
    geom_point(aes(x = nb_0.05, y = moran_1k3, color = 3)) +
    geom_point(aes(x = nb_0.05, y = moran_1k4, color = 4)) +
    geom_point(aes(x = nb_0.05, y = moran_1k5, color = 5)) +
    geom_point(aes(x = nb_0.05, y = moran_1k6, color = 6)) +
    geom_point(aes(x = nb_0.05, y = moran_1k7, color = 7)) +
    geom_point(aes(x = nb_0.05, y = moran_1k8, color = 8)) +
    geom_point(aes(x = nb_0.05, y = moran_1k9, color = 9)) +
    geom_point(aes(x = nb_0.05, y = moran_1k10, color = 10)) +
    labs(title = "Residuals of SDM", 
    subtitle = "Can niche breadth predict ",
    x = "Climatic Niche Breadth", y = "Residual Moran's I") +
    theme_bw()



```

```{r}
q <- ggplot(data = mammal_Y_moran) +
    geom_point(aes(x = nb_0.05, y = moran_1k1, color = 1)) +
    geom_smooth(aes(x = nb_0.05, y= moran_1k10), method = "lm") +
    geom_point(aes(x = nb_0.05, y = moran_1k10, color = 10)) +
    labs(title = "Residuals of SDM", 
    subtitle = "Can niche breadth predict?",
    x = "Climatic Niche Breadth", y = "Residual Moran's I") +
    theme_bw()

scatter_morannb <- function(y){ggplot(data = mammal_Y_moran) +
                                  geom_point(aes(x = nb_0.05, y = y)) + 
                                  geom_smooth(aes(x = nb_0.05, y = y), method = "lm")}
  
all_scatter<- map(mammal_Y_moran[1:60], scatter_morannb)


map2(ggname_save, all_scatter, ggsave)

a <- list(1:60)

ggname <- read_csv("E:/residual_sdm/species_mammal.csv", col_names = "scientificName", skip = 1)
ggname<- ggname[-31,]
ggname <- ggname[-26,]
ggname <- as.list(ggname)
ggname_save <- paste0("E:/",1:60,".png")
```

```{r}
p <- ggplot(data = mammal_Y_moran) +
    geom_point(aes(x = nb_0.05, y = moran_1dnb, size = , color = "red")) +
    geom_point(aes(x = nb_0.05, y = moran_1k10, size = 0.002, color = "blue")) +
    geom_smooth(aes(x = nb_0.05, y= moran_1dnb), method = "lm") +
    geom_smooth(aes(x = nb_0.05, y= moran_1k10), method = "lm") +
    labs(title = "Residuals of SDM", 
    subtitle = "Can niche breadth predict ",
    x = "Climatic Niche Breadth", y = "Residual Moran's I", fill="wtf") +
    theme(legend.position = 'none') 
    

    





```

```{r}
p <- ggplot(data = mammal_Y_moran) +
    geom_point(aes(x = nb_0.05, y = moran_1k1, color = 1)) +
    geom_smooth(aes(x = nb_0.05, y= moran_1k10), method = "lm") +
    geom_point(aes(x = nb_0.05, y = moran_1k2, color = 2)) +
    geom_point(aes(x = nb_0.05, y = moran_1k3, color = 3)) +
    geom_point(aes(x = nb_0.05, y = moran_1k4, color = 4)) +
    geom_point(aes(x = nb_0.05, y = moran_1k5, color = 5)) +
    geom_point(aes(x = nb_0.05, y = moran_1k6, color = 6)) +
    geom_point(aes(x = nb_0.05, y = moran_1k7, color = 7)) +
    geom_point(aes(x = nb_0.05, y = moran_1k8, color = 8)) +
    geom_point(aes(x = nb_0.05, y = moran_1k9, color = 9)) +
    geom_point(aes(x = nb_0.05, y = moran_1k10, color = 10)) +
    labs(title = "Residuals of SDM", 
    subtitle = "Can niche breadth predict ",
    x = "Climatic Niche Breadth", y = "Residual Moran's I") +
    theme_bw()



```



```{r}
p <- ggplot(data = mammal_Y_moran) +
    geom_smooth(aes(x = nb_0.05, y= moran_0.9k10), method = "lm") +
    labs(title = "Residuals of SDM", 
    subtitle = "Can niche breadth predict ",
    x = "Climatic Niche Breadth", y = "Residual Moran's I") +
    theme_bw()
```

```{r histogram}
tibble4ggp1 <- tibble(rsac = mammal_Y_moran$moran_1dnb)
tibble4ggp2 <- tibble(rsac = mammal_Y_moran$moran_1k10)
tibble4ggp <- rbind(tibble4ggp1, tibble4ggp2)                     
tibble4ggp <- tibble4ggp %>% mutate(neighbor = length(1:100))
tibble4ggp$neighbor[1:50] <- "variogram distance"
tibble4ggp$neighbor[51:100] <- "10NN"


ggplot(tibble4ggp, aes(x=rsac, color=neighbor, )) +
  geom_histogram(aes(fill=neighbor), color="gray50", position="dodge", alpha=0.2, binwidth=0.09)+
  geom_density(alpha=0.5)+
  theme_gray()


```



```{r}

for (i in 1:length(sf_train_rndcv4k_amphibian)){
sf::write_sf(sf_train_rndcv4k_amphibian[[i]], paste0("train_test/t.t_rndcv4k_amphibian/train_",t.t.amphibian_0.2[[i]][[1]]@species[1],".shp"))
}

for (i in 1:length(sf_test_rndcv4k_amphibian)){
sf::write_sf(sf_test_rndcv4k_amphibian[[i]], paste0("train_test/t.t_rndcv4k_amphibian/test/test_",t.t.amphibian_0.2[[i]][[1]]@species[1],".shp"))
}

#

for (i in 1:length(sf_train_rndcv4k_mammal)){
sf::write_sf(sf_train_rndcv4k_mammal[[i]], paste0("train_test/t.t_rndcv4k_mammal/train/train_",t.t.mammal_0.2[[i]][[1]]@species[1],".shp"))
}

for (i in 1:length(sf_test_rndcv4k_mammal)){
sf::write_sf(sf_test_rndcv4k_mammal[[i]], paste0("train_test/t.t_rndcv4k_mammal/test/testn_",t.t.mammal_0.2[[i]][[1]]@species[1],".shp"))
}

#

for (i in 1:length(sf_train_rndcv4k_reptile)){
sf::write_sf(sf_train_rndcv4k_reptile[[i]], paste0("train_test/t.t_rndcv4k_reptile/train/train_",t.t.reptile_0.2[[i]][[1]]@species[1],".shp"))
}

for (i in 1:length(sf_test_rndcv4k_reptile)){
sf::write_sf(sf_test_rndcv4k_reptile[[i]], paste0("train_test/t.t_rndcv4k_reptile/test/testn_",t.t.reptile_0.2[[i]][[1]]@species[1],".shp"))
}

```
