==> 4_niche_breadth.Rmd <==
---
title: "4_niche_breadth"
author: "JEON"
date: '2021 3 9 '
output: html_document
---
```{r}
library(ggcorrplot)
library(FactoMineR)
library(factoextra)
library(vegan)
library(listviewer)
library(MicroNiche)
library(corrr)
library(mapr)
library(ggplot2)
```

```{r}
library(tidyverse)

```

```{r}
library(furrr)
future::plan(multisession, workers = 30)
```


# Aims. 
Capture the niche breadth of every species.
Methodology include 3 ways,
1. classical Levin's Niche Breadth with my own function(다 더하는 것은 어렵. samp intensity에 대한 비판 방어는?)
2. Levin's Niche breadth with ENMeval
3. DBSCAN - 자의적 판단기준?
4. Broennimann et al. 2012 ecospat... var, PCA, Sum of Variance
5. 0.05 ~ 0.95 boxplot
6. 전문가 데이터와 비교 
7. 순위로 더하기
8. 다 해보고 순위가 비슷한 애들만 뽑아보기?
9. 


# 1. Classical Levin's niche breadth
This represent relative frequency of individuals in the jth Category
which catch the Eveness along the gradients

0 < B < J (number of Jth category)
0 represent specialist, J reprensent generalist.

until bio11, it is temperature. each category can be 2
from 12 - 19, they are precipitation. each category can be 10 each.

```{r}
data.mammal <- readr::read_rds(file = "data/data.mammal.rds")
data.reptile <- readr::read_rds(file = "data/data.reptile.rds")
data.amphibian <- readr::read_rds(file = "data/data.amphibian.rds")
```


```{r}

reaarange <- function(x){dplyr::select(x@data, c(1,12,13,14,15,16,17,18,19,2,3,4,5,6,7,8,9,10,11))}

for (i in 1:length(data.mammal)){
data.mammal[[i]]@data <- reaarange(data.mammal[[i]])
}

for (i in 1:length(data.mammal)){
names(data.mammal[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                  "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                  "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                  "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                  "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}

for (i in 1:length(data.reptile)){
data.reptile[[i]]@data <- reaarange(data.reptile[[i]])
}

for (i in 1:length(data.reptile)){
names(data.reptile[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                  "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                  "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                  "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                  "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}

for (i in 1:length(data.amphibian)){
data.amphibian[[i]]@data <- reaarange(data.amphibian[[i]])
}

for (i in 1:length(data.amphibian)){
names(data.amphibian[[i]]@data) <- c("wc2.1_30s_bio_01", "wc2.1_30s_bio_02", "wc2.1_30s_bio_03", "wc2.1_30s_bio_04",
                                  "wc2.1_30s_bio_05", "wc2.1_30s_bio_06", "wc2.1_30s_bio_07", "wc2.1_30s_bio_08",
                                  "wc2.1_30s_bio_09", "wc2.1_30s_bio_10", "wc2.1_30s_bio_11", "wc2.1_30s_bio_12",
                                  "wc2.1_30s_bio_13", "wc2.1_30s_bio_14", "wc2.1_30s_bio_15", "wc2.1_30s_bio_16",
                                  "wc2.1_30s_bio_17", "wc2.1_30s_bio_18", "wc2.1_30s_bio_19")
}

```


```{r}
library(raster)

bios.name <- paste0("C:/Users/lophital/Documents/rsac_x_overfit/bios/2.5m/", list.files(pattern="*.tif", path = "C:/Users/lophital/Documents/rsac_x_overfit/bios/2.5m/"))

bios <- list()
for (i in 1:length(bios.name)){
  bios[[i]] <- raster::raster(bios.name[i])
}

bio.min <- vector()
for (i in 1:19){
bio.min[i] <- ceiling(bios[[i]]@data@min)
}

bio.max <- vector()
for (i in 1:19){
bio.max[i] <- ceiling(bios[[i]]@data@max)
}

biobreak <- list()
for(i in 1:11){
biobreak[[i]] <- bio.min[i]:bio.max[i]
}

for(i in 12:19){
biobreak[[i]] <- (bio.min[i]:bio.max[i]*10)  
}

bchist <- function(x, y){hist(x, breaks = y)}
biohist_reptile <- list()
for (i in 1:94){
biohist_reptile[[i]] <- map2(data.reptile[[i]]@data, biobreak, bchist)
}

levin.nb <- function(x){
  1/sum((x$counts/sum(x$counts))^2)
  }

levin.nb_reptile <- list()
for(i in 1:length(data.reptile)){
levin.nb_reptile[[i]] <- map(biohist_reptile[[i]], levin.nb)
}

nichebreadth.df <- data.frame(bio1 = map_dbl(levin.nb_reptile, 1),
                              bio2 = map_dbl(levin.nb_reptile, 2),
                              bio3 = map_dbl(levin.nb_reptile, 3),
                              bio4 = map_dbl(levin.nb_reptile, 4),
                              bio5 = map_dbl(levin.nb_reptile, 5),
                              bio6 = map_dbl(levin.nb_reptile, 6),
                              bio7 = map_dbl(levin.nb_reptile, 7),
                              bio8 = map_dbl(levin.nb_reptile, 8),
                              bio9 = map_dbl(levin.nb_reptile, 9),
                              bio10 = map_dbl(levin.nb_reptile, 10),
                              bio11 = map_dbl(levin.nb_reptile, 11),
                              bio12 = map_dbl(levin.nb_reptile, 12),
                              bio13 = map_dbl(levin.nb_reptile, 13),
                              bio14 = map_dbl(levin.nb_reptile, 14),
                              bio15 = map_dbl(levin.nb_reptile, 15),
                              bio16 = map_dbl(levin.nb_reptile, 16),
                              bio17 = map_dbl(levin.nb_reptile, 17),
                              bio18 = map_dbl(levin.nb_reptile, 18),
                              bio19 = map_dbl(levin.nb_reptile, 19))

```

# 2. Levin's niche breadth with ENMeval

```{r}
library(RInSp)

a <- RInSp::import.RInSp(data.reptile[[2]]@data, data.type = "double")
b <- RInSp::like.Wi(a)
```


# 3. Broennimann et al. 2012

plot in 2 PC axis, Z axis of SDM, Kernel density smoothing


```{r{}
# this will set up disk.frame with multiple workers
setup_disk.frame(workers = 60)

# this will allow unlimited amount of data to be passed from worker to worker
options(future.globals.maxSize = Inf)
```



```{r}
library(FactoMineR)
library(data.table)

stackbio <- raster::stack(bios)

memory.limit(9999999)

bio_df <- data.frame(na.omit(raster::values(stackbio)))
readr::write_csv(bio_df, "bios/bio_df.csv")

bio_df <- fread("bio_df.csv")



my.PCA <- function(x){FactoMineR::PCA(x@data, scale.unit = TRUE)}

memory.limit(999999999)
PCA_bio <- FactoMineR::PCA(bio_df, scale.unit = T, ncp = 2)

PCA_mammal <- future_map(data.mammal, 
                         my.PCA,
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

PCA_reptile <- future_map(data.reptile, 
                         my.PCA,
                         .options = furrr_options(seed = TRUE),
                         .progress = T)

PCA_amphibian <- future_map(data.amphibian, 
                         my.PCA,
                         .options = furrr_options(seed = TRUE),
                         .progress = T)


```



```{r}
library(ecospat)
```


```{r}
mcp <- function (xy) {
  xy <- as.data.frame(coordinates(xy))
  coords.t <- chull(xy[, 1], xy[, 2])
  xy.bord <- xy[coords.t, ]
  xy.bord <- rbind(xy.bord[nrow(xy.bord), ], xy.bord)
  return(SpatialPolygons(list(Polygons(list(Polygon(as.matrix(xy.bord))), 1))))
  }
```



```{r}
my.ecospat.kd <- function (x, ext, R = 100, th = 0, env.mask = c(), method = "adehabitat") 
{
  if (method == "adehabitat") {
    if (ncol(x) == 2) {
      xr <- data.frame(cbind((x[, 1] - ext[1])/abs(ext[2] - 
        ext[1]), (x[, 2] - ext[3])/abs(ext[4] - ext[3])))
      mask <- adehabitatMA::ascgen(sp::SpatialPoints(cbind((0:(R))/R, 
        (0:(R)/R))), nrcol = R - 2, count = FALSE)
      x.dens <- adehabitatHR::kernelUD(sp::SpatialPoints(xr[, 
        1:2]), h = "href", grid = mask, kern = "bivnorm")
      x.dens <- raster::raster(xmn = ext[1], xmx = ext[2], 
        ymn = ext[3], ymx = ext[4], matrix(x.dens$ud, 
          nrow = R))
      if (!is.null(th)) {
        th.value <- quantile(raster::extract(x.dens, 
          x), th, na.rm = T)
        x.dens[x.dens < th.value] <- 0
      }
      if (!is.null(env.mask)) {
        x.dens <- x.dens * env.mask
      }
    }
    else if (ncol(x) == 1) {
      xr <- seq(from = min(ext), to = max(ext), length.out = R)
      x.dens <- density(x[, 1], kernel = "gaussian", from = min(xr), 
        to = max(xr), n = R, cut = 0)
      if (!is.null(env.mask)) {
        x.dens$y <- x.dens$y * env.mask
      }
      if (!is.null(th)) {
        xr <- sapply(x, findInterval, x.dens$x)
        th.value <- quantile(x.dens$y[xr], th, na.rm = T)
        sprm <- which(x.dens$y < th.value)
        x.dens$y[sprm] <- 0
      }
    }
  }
  if (method == "ks") {
    if (ncol(x) == 2) {
      x.dens <- ks::kde(x, xmin = ext[c(1, 3)], xmax = ext[c(2, 
        4)], gridsize = c(R, R))
      x.dens <- raster::flip(raster::t(raster::raster(x.dens$estimate)), 
        direction = "y")
      raster::extent(x.dens) <- c(xmn = ext[1], xmx = ext[2], 
        ymn = ext[3], ymx = ext[4])
      if (!is.null(th)) {
        th.value <- quantile(raster::extract(x.dens, 
          x), th, na.rm = T)
        x.dens[x.dens < th.value] <- 0
      }
      if (!is.null(env.mask)) {
        x.dens <- x.dens * env.mask
      }
    }
    else if (ncol(x) == 1) {
      x.dens <- ks::kde(x, xmin = min(ext), xmax = max(ext), 
        gridsize = c(R, R))
      x.dens$y <- x.dens$estimate
      x.dens$x <- x.dens$eval.points
      if (!is.null(env.mask)) {
        x.dens$y <- x.dens$y * env.mask
      }
      if (!is.null(th)) {
        xr <- sapply(x, findInterval, x.dens$x)
        th.value <- quantile(x.dens$y[xr], th, na.rm = T)
        sprm <- which(x.dens$y < th.value)
        x.dens$y[sprm] <- 0
      }
    }
  }
  return(x.dens)
}


my.ecospat.grid <- function (glob, glob1, sp, R = 100, th.sp = 0, th.env = 0, geomask = NULL, 
  kernel.method = "adehabitat", extend.extent = c(0, 0, 0, 
    0)) 
{
  if (is.null(kernel.method) | (kernel.method != "ks" & kernel.method != 
    "adehabitat")) {
    stop("supply a kernel method ('adehabitat' or 'ks')")
  }
  glob <- as.matrix(glob)
  glob1 <- as.matrix(glob1)
  sp <- as.matrix(sp)
  l <- list()
  if (ncol(glob) > 2) {
    stop("cannot calculate overlap with more than two axes")
  }
  if (ncol(glob) == 1) {
    xmin <- min(glob[, 1]) + extend.extent[1]
    xmax <- max(glob[, 1]) + extend.extent[2]
    glob1.dens <- my.ecospat.kd(x = glob1, ext = c(xmin, xmax), 
      method = kernel.method, th = 0)
    sp.dens <- my.ecospat.kd(x = sp, ext = c(xmin, xmax), method = kernel.method, 
      th = 0, env.mask = glob1.dens$y > 0)
    x <- sp.dens$x
    y <- sp.dens$y
    z <- sp.dens$y * nrow(sp)/sum(sp.dens$y)
    Z <- glob1.dens$y * nrow(glob)/sum(glob1.dens$y)
    z.uncor <- z/max(z)
    z.cor <- z/Z
    z.cor[is.na(z.cor)] <- 0
    z.cor[z.cor == "Inf"] <- 0
    z.cor <- z.cor/max(z.cor)
  }
  if (ncol(glob) == 2) {
    xmin <- apply(glob, 2, min, na.rm = T)
    xmax <- apply(glob, 2, max, na.rm = T)
    ext <- c(xmin[1], xmax[1], xmin[2], xmax[2]) + extend.extent
    glob1.dens <- my.ecospat.kd(x = glob1, ext = ext, method = kernel.method, 
      th = 0)
    if (!is.null(geomask)) {
      sp::proj4string(geomask) <- NA
      glob1.dens <- raster::mask(glob1.dens, geomask, 
        updatevalue = 0)
    }
    sp.dens <- my.ecospat.kd(x = sp, ext = ext, method = kernel.method, 
      th = 0, env.mask = glob1.dens > 0)
    x <- seq(from = ext[1], to = ext[2], length.out = 100)
    y <- seq(from = ext[3], to = ext[4], length.out = 100)
    l$y <- y
    Z <- glob1.dens * nrow(glob1)/raster::cellStats(glob1.dens, 
      "sum")
    z <- sp.dens * nrow(sp)/raster::cellStats(sp.dens, "sum")
    z.uncor <- z/raster::cellStats(z, "max")
    z.cor <- z/Z
    z.cor[is.na(z.cor)] <- 0
    z.cor <- z.cor/raster::cellStats(z.cor, "max")
  }
  w <- z.uncor
  w[w > 0] <- 1
  l$x <- x
  l$z <- z
  l$z.uncor <- z.uncor
  l$z.cor <- z.cor
  l$Z <- Z
  l$glob <- glob
  l$glob1 <- glob1
  l$sp <- sp
  l$w <- w
  return(l)
}
```

```{r}
my.ecospat.grid2 <- function(x){
                     my.ecospat.grid(glob = PCA_bio$ind$coord[,1:2],
                                    glob1 = PCA_bio$ind$coord[,1:2],
                                    sp = x$ind$coord[,1:2],
                                    R = 1000
                                    )}

nb_mammal <- map(PCA_mammal, my.ecospat.grid2)
nb_reptile <- future_map(PCA_reptile, my.ecospat.grid2)
nb_amphibian <- future_map(PCA_amphibian, my.ecospat.grid2)

nb.01 <- function(x){length(Which(x$z.uncor>0.1, cells=T))}
nb.03 <- function(x){length(Which(x$z.uncor>0.3, cells=T))}
nb.05 <- function(x){length(Which(x$z.uncor>0.5, cells=T))}
nb.09 <- function(x){length(Which(x$z.uncor>0.9, cells=T))}

nbb.01_mammal <- map_int(nb_mammal, nb.01) 
nbb.03_mammal <- map_int(nb_mammal, nb.03)
nbb.05_mammal <- map_int(nb_mammal, nb.05)
nbb.09_mammal <- map_int(nb_mammal, nb.09)

nbb.01_reptile <- map_int(nb_reptile, nb.01) 
nbb.03_reptile <- map_int(nb_reptile, nb.03)
nbb.05_reptile <- map_int(nb_reptile, nb.05)
nbb.09_reptile <- map_int(nb_reptile, nb.09)

nbb.01_amphibian <- map_int(nb_amphibian, nb.01) 
nbb.03_amphibian <- map_int(nb_amphibian, nb.03)
nbb.05_amphibian <- map_int(nb_amphibian, nb.05)
nbb.09_amphibian <- map_int(nb_amphibian, nb.09)

plot(a$z.uncor, main = "Sp1 with default kernel")
length(Which(x$z.uncor>0.5, cells=T))

```


```{r}
result_mammal <-readr:::read_csv(file = "result_mammal.csv")
result_mammal %>% mutate(nb_01 = nbb.01_mammal,
                         nb_03 = nbb.03_mammal,
                         nb_05 = nbb.05_mammal,
                         nb_09 = nbb.09_mammal) -> result_mammal
readr::write_csv(result_mammal, file =  "result_mammal.csv")

result_reptile <-readr:::read_csv(file = "result_reptile.csv")
result_reptile %>% mutate(nb_01 = nbb.01_reptile,
                         nb_03 = nbb.03_reptile,
                         nb_05 = nbb.05_reptile,
                         nb_09 = nbb.09_reptile) -> result_reptile
readr::write_csv(result_reptile, file =  "result_reptile.csv")

result_amphibian <-readr:::read_csv(file = "result_amphibian.csv")
result_amphibian %>% mutate(nb_01 = nbb.01_amphibian,
                         nb_03 = nbb.03_amphibian,
                         nb_05 = nbb.05_amphibian,
                         nb_09 = nbb.09_amphibian) -> result_amphibian
readr::write_csv(result_amphibian, file =  "result_amphibian.csv")


qq <- ggplot(data = result_mammal) +
    geom_point(aes(x = nb_01, y = rsac_rndcv4k_sph, color = 1)) +
    geom_smooth(aes(x = nb_01, y= rsac_rndcv4k_sph), method = "lm") +
    geom_point(aes(x = nb_03, y = rsac_rndcv4k_sph, color = 2)) +
    geom_smooth(aes(x = nb_03, y= rsac_rndcv4k_sph), method = "lm") +
    geom_point(aes(x = nb_05, y = rsac_rndcv4k_sph, color = 3)) +
    geom_smooth(aes(x = nb_05, y= rsac_rndcv4k_sph), method = "lm") +
    geom_point(aes(x = nb_09, y = rsac_rndcv4k_sph, color = 4)) +
    geom_smooth(aes(x = nb_09, y= rsac_rndcv4k_sph), method = "lm")

q <- ggplot(data = result_mammal) +
    geom_point(aes(x = nb_01, y = rsac_rndcv4k_sph, color = 1)) +
    geom_smooth(aes(x = nb_01, y= rsac_rndcv4k_sph), method = "lm") 

q <- ggplot(data = result_reptile) +
    geom_point(aes(x = nb_01, y = rsac_rndcv4k_sph, color = 1)) +
    geom_smooth(aes(x = nb_01, y= rsac_rndcv4k_sph), method = "lm") 

qq <- ggplot(data = result_reptile) +
    geom_point(aes(x = nb_01, y = rsac_rndcv4k_sph, color = 1)) +
    geom_smooth(aes(x = nb_01, y= rsac_rndcv4k_sph), method = "lm") 
```










```{r}
map(out_reptile$gbif$data, map_ggplot)
```
```{r mapping occurences with mapr::mapggplot}
map(out_aves$gbif$data, map_ggplot)
mapr::map_ggplot(out_reptile$gbif$data$Terrapene_carolina, map = "usa", size = 1)
```

```{r standardized extracted points values}
standard <- function(x){scale(x, center = T, scale = T)}

standard_amphibian <- extract_spatial_amphibian_lapply %>%  map(.f = scale(., center = T, scale = T))
standard_amphibian <- map(extract_spatial_amphibian_lapply, standard) 
standard_amphibian_df <- map(standard_amphibian, as.data.frame)


```



```{r PCA extracted points with FactoMineR::PCA }
#----------------------------
# PCA extracted points
#----------------------------
list_PCA_mammal <- lapply(
  extract_spatial_mammal_lapply,
  function(x){
    PCA(x)
  }
)

list_PCA_reptile <- lapply(
  extract_spatial_reptile_lapply,
  function(x){
    PCA(x)
  }
)

list_PCA_amphibian <- lapply(
  extract_spatial_amphibian_lapply,
  function(x){
    PCA(x)
  }
)

list_PCA_aves <- lapply(
  extract_spatial_aves_furrr,
  function(x){
    PCA(x)
  }
)

```

```{r PCA extracted points from saved}
#----------------------------
# PCA extracted points - saved
#----------------------------
list_PCA_mammal <- lapply(
  extract_mammal_saved,
  function(x){
    PCA(x)
  }
)

list_PCA_reptile <- lapply(
  extract_reptile_saved,
  function(x){
    PCA(x)
  }
)

list_PCA_amphibian <- lapply(
  extract_spatial_amphibian_lapply,
  function(x){
    PCA(x)
  }
)

list_PCA_aves <- lapply(
  extract_aves_saved,
  function(x){
    PCA(x)
  }
)

```


```{r PCA for the standardized extracted points value}
list_PCA_amphibian_standard <- lapply(
  standard_amphibian,
  function(x){
    PCA(x)
  }
)

```

```{r caculating niche breadth}
#----------------------------
# Niche breadth - 0.05 ~ 0.95
#----------------------------

cord_out_1 <- list()
cord_out_2 <- list()

for (i in seq_along(1:length(list_PCA_mammal))){
  cord_out_1[[i]] <- quantile(list_PCA_mammal[[i]]$ind$coord[,1], c(0.05, .95))
  cord_out_2[[i]] <- quantile(list_PCA_mammal[[i]]$ind$coord[,2], c(0.05, .95))
}

cord_mammal_df_pc1 <- do.call(rbind, cord_out_1)
cord_mammal_df_pc2 <- do.call(rbind, cord_out_2)
nb_mammal_pc1_df <- abs(cord_mammal_df_pc1[,1]) + cord_mammal_df_pc1[,2]
nb_mammal_pc2_df <- abs(cord_mammal_df_pc2[,2]) + cord_mammal_df_pc2[,2]
nb_mammal <- nb_mammal_pc1_df + nb_mammal_pc2_df 

##########################################################



cord_out_1 <- list()
cord_out_2 <- list()

for (i in seq_along(1:length(list_PCA_reptile))){
  cord_out_1[[i]] <- quantile(list_PCA_reptile[[i]]$ind$coord[,1], c(0.05, .95))
  cord_out_2[[i]] <- quantile(list_PCA_reptile[[i]]$ind$coord[,2], c(0.05, .95))
}

cord_reptile_df_pc1 <- do.call(rbind, cord_out_1)
cord_reptile_df_pc2 <- do.call(rbind, cord_out_2)
nb_reptile_pc1_df <- abs(cord_reptile_df_pc1[,1]) + cord_reptile_df_pc1[,2]
nb_reptile_pc2_df <- abs(cord_reptile_df_pc2[,2]) + cord_reptile_df_pc2[,2]
nb_reptile <- nb_reptile_pc1_df + nb_reptile_pc2_df

##########################################################

cord_out_1 <- list()
cord_out_2 <- list()
# A for loop to extract the quantiles and same
for (i in seq_along(1:length(list_PCA_amphibian))){
  cord_out_1[[i]] <- quantile(list_PCA_amphibian[[i]]$ind$coord[,1], c(0.05, .95))
  cord_out_2[[i]] <- quantile(list_PCA_amphibian[[i]]$ind$coord[,2], c(0.05, .95))
}

# Create two corresponding DF's
cord_amphibian_df_pc1 <- do.call(rbind, cord_out_1)
cord_amphibian_df_pc2 <- do.call(rbind, cord_out_2)
nb_amphibian_pc1_df <- abs(cord_amphibian_df_pc1[,1]) + cord_amphibian_df_pc1[,2]
nb_amphibian_pc2_df <- abs(cord_amphibian_df_pc2[,2]) + cord_amphibian_df_pc2[,2]
nb_amphibian <- nb_amphibian_pc1_df + nb_amphibian_pc2_df 

##########################################################

cord_out_1 <- list()
cord_out_2 <- list()
# A for loop to extract the quantiles and same
for (i in seq_along(1:length(list_PCA_aves))){
  cord_out_1[[i]] <- quantile(list_PCA_aves[[i]]$ind$coord[,1], c(0.05, .95))
  cord_out_2[[i]] <- quantile(list_PCA_aves[[i]]$ind$coord[,2], c(0.05, .95))
}

# Create two corresponding DF's
cord_aves_df_pc1 <- do.call(rbind, cord_out_1)
cord_aves_df_pc2 <- do.call(rbind, cord_out_2)
nb_aves_pc1_df <- abs(cord_aves_df_pc1[,1]) + cord_aves_df_pc1[,2]
nb_aves_pc2_df <- abs(cord_aves_df_pc2[,2]) + cord_aves_df_pc2[,2]
nb_aves <- nb_aves_pc1_df + nb_aves_pc2_df 

##########################################################
```

```{r niche breadth with standardzied}
#----------------------------
# Niche breadth with standardized 
#----------------------------

cord_out_1 <- list()
cord_out_2 <- list()

for (i in seq_along(1:length(list_PCA_amphibian_standard))){
  cord_out_1[[i]] <- quantile(list_PCA_amphibian_standard[[i]]$ind$coord[,1], c(0.05, .95))
  cord_out_2[[i]] <- quantile(list_PCA_amphibian_standard[[i]]$ind$coord[,2], c(0.05, .95))
}

cord_amphibian_df_pc1_sd <- do.call(rbind, cord_out_1)
cord_amphibian_df_pc2_sd <- do.call(rbind, cord_out_2)
nb_amphibian_pc1_df_sd <- abs(cord_amphibian_df_pc1_sd[,1]) + cord_amphibian_df_pc1_sd[,2]
nb_amphibian_pc2_df_sd <- abs(cord_amphibian_df_pc2_sd[,2]) + cord_amphibian_df_pc2_sd[,2]
nb_amphibian_sd <- nb_amphibian_pc1_df_sd + nb_amphibian_pc2_df_sd

```

Do really nb is different with taxons?

```{r save nb}
nb_aves_df <- data.frame(nb_aves)
nb_reptile_df <- data.frame(nb_reptile)
nb_mammal_df <- data.frame(nb_mammal)

write_csv(nb_aves_df, "E:/nichebreadth/nb_aves_df.csv")
write_csv(nb_reptile_df, "E:/nichebreadth/nb_reptile_df.csv")
write_csv(nb_mammal_df, "E:/nichebreadth/nb_mammal_df.csv")

```

```{r TOP 4 PCA Variables Visualize}
pca_var_g <- fviz_pca_var(pkmon_pca, select.var = list(cos2 = 0.5), repel = TRUE)

sweep(list_PCA_mammal$Peromyscus_maniculatus$var$coord,2,sqrt(list_PCA_mammal$Peromyscus_maniculatus$eig[1:ncol(list_PCA_mammal$Peromyscus_maniculatus$var$coord),1]),FUN="/")

pcm1 <- PCA(m1bio)
pca_var_contrib <- factoextra::fviz_pca_var(pcm1, select.var = list(contrib = 3), repel = TRUE)

```

```{r TOP 4 PCA Variables}
top4_pc1_mammal <- list()
top4_pc2_mammal <- list()
for (i in seq_along(1:length(list_PCA_mammal))){
 top4_pc1_mammal[[i]] <- names(tail(sort(list_PCA_mammal[[i]]$var$cos2[,1]), 4))
 top4_pc2_mammal[[i]] <- names(tail(sort(list_PCA_mammal[[i]]$var$cos2[,2]), 4))
}

top4_pc1_reptile <- list()
top4_pc2_reptile <- list()
for (i in seq_along(1:length(list_PCA_reptile))){
  top4_pc1_reptile[[i]] <- names(tail(sort(list_PCA_reptile[[i]]$var$cos2[,1]), 4))
  top4_pc2_reptile[[i]] <- names(tail(sort(list_PCA_reptile[[i]]$var$cos2[,2]), 4))
}

top4_pc1_amphibian <- list()
top4_pc2_amphibian <- list()
for (i in seq_along(1:length(list_PCA_amphibian))){
  top4_pc1_amphibian[[i]] <- names(tail(sort(list_PCA_amphibian[[i]]$var$cos2[,1]), 4))
  top4_pc2_amphibian[[i]] <- names(tail(sort(list_PCA_amphibian[[i]]$var$cos2[,2]), 4))
}

top4_pc1_aves <- list()
top4_pc2_aves <- list()
for (i in seq_along(1:length(list_PCA_aves))){
  top4_pc1_aves[[i]] <- names(tail(sort(list_PCA_aves[[i]]$var$cos2[,1]), 4))
  top4_pc2_aves[[i]] <- names(tail(sort(list_PCA_aves[[i]]$var$cos2[,2]), 4))
}

```

```{r TOP 4 PCA standardized variables}
##################### 
# TOP 4 PCA standardized 
######################
top4_pc1_amphibian_sd <- list()
top4_pc2_amphibian_sd <- list()
for (i in seq_along(1:length(list_PCA_amphibian_standard))){
  top4_pc1_amphibian_sd[[i]] <- names(tail(sort(list_PCA_amphibian_standard[[i]]$var$cos2[,1]), 4))
  top4_pc2_amphibian_sd[[i]] <- names(tail(sort(list_PCA_amphibian_standard[[i]]$var$cos2[,2]), 4))
}

```

```{r stacking rasters }

paste2 <- function(x){paste("E:/residual_sdm/" ,x, ".tif", sep = "")}



######################################################
top4_pc1_mammal.tif <- top4_pc1_mammal %>% map(paste2)
top4_pc2_mammal.tif <- top4_pc2_mammal %>% map(paste2)

top4_pc2_mammal.tif.stack <- top4_pc2_mammal.tif %>% map(stack)
top4_pc1_mammal.tif.stack <- top4_pc1_mammal.tif %>% map(stack)
top4_pc12_mammal.stack <- map2(top4_pc1_mammal.tif.stack, top4_pc2_mammal.tif.stack, stack)

##############################################################################
top4_pc1_reptile.tif <- top4_pc1_reptile %>% map(paste2)
top4_pc2_reptile.tif <- top4_pc2_reptile %>% map(paste2)

top4_pc2_reptile.tif.stack <- top4_pc2_reptile.tif %>% map(stack)
top4_pc1_reptile.tif.stack <- top4_pc1_reptile.tif %>% map(stack)
top4_pc12_reptile.stack <- map2(top4_pc1_reptile.tif.stack, top4_pc2_reptile.tif.stack, stack)

##############################################################################
top4_pc1_amphibian.tif <- top4_pc1_amphibian %>% map(paste2)
top4_pc2_amphibian.tif <- top4_pc2_amphibian %>% map(paste2)

top4_pc2_amphibian.tif.stack <- top4_pc2_amphibian.tif %>% map(stack)
top4_pc1_amphibian.tif.stack <- top4_pc1_amphibian.tif %>% map(stack)
top4_pc12_amphibian.stack <- map2(top4_pc1_amphibian.tif.stack, top4_pc2_amphibian.tif.stack, stack)

##############################################################################
top4_pc1_aves.tif <- top4_pc1_aves %>% map(paste2)
top4_pc2_aves.tif <- top4_pc2_aves %>% map(paste2)

top4_pc2_aves.tif.stack <- top4_pc2_aves.tif %>% map(stack)
top4_pc1_aves.tif.stack <- top4_pc1_aves.tif %>% map(stack)
top4_pc12_aves.stack <- map2(top4_pc1_aves.tif.stack, top4_pc2_aves.tif.stack, stack)

```

```{r get the extent}
#extent 
extent_mammal <- map(list_spatial_mammal, extent)
extent_reptile <- map(list_spatial_reptile, extent)
extent_amphibian <- map(list_spatial_amphibian, extent)
extent_aves <- map(list_spatial_aves, extent)
```

```{r}
#----------------------------
#raster crop to save times
#----------------------------

tic("mammal_bios") #mammal_bios: 5026.39 sec elapsed
mammal_bios <- furrr::future_map2(top4_pc12_mammal.stack, extent_mammal, crop)
toc()

tic("reptile_bios") #reptile_bios: 2834.84 sec elapsed
reptile_bios <- furrr::future_map2(top4_pc12_reptile.stack, extent_reptile, crop)
toc()

tic("amphibian_bios") #amphibian_bios: 762.35 sec elapsed
amphibian_bios <- furrr::future_map2(top4_pc12_amphibian.stack, extent_amphibian, crop)
toc()

tic("aves_bios") #aves_bios: 3765.14 sec elapsed
aves_bios <- furrr::future_map2(top4_pc12_aves.stack, extent_aves, crop)
toc()


```


```{r save raster}

reptile_bios %>% furrr::future_map2(paste0("J:/reptile_bios_", list_origin_reptile,".gri"), writeRaster)

amphibian_bios %>% furrr::future_map2(paste0("J:/amphibian_bios_", list_origin_amphibian,".gri"), writeRaster)

aves_bios %>% 
furrr::future_map2(paste0("J:/aves_bios_", list_origin_aves,".gri"), writeRaster)

```

```{r read saved rasters}

bios_mammal <- list.files(path = "J:/", pattern="*.gri")
bios_mammal <- paste0("J:/mammal_bios_", list_origin_mammal, ".gri")
bios_mammal <- map(bios_mammal, brick)

bios_amphibian <- list.files(path = "J:/", pattern="*.gri")
bios_amphibian <- paste0("J:/amphibian/amphibian_bios_", list_origin_amphibian, ".gri")
bios_amphibian <- map(bios_amphibian, brick)

bios_aves <- list.files(path = "J:/", pattern="*.gri")
bios_aves <- paste0("J:/aves_bios_", list_origin_aves, ".gri")
bios_aves <- map(bios_aves, brick)

bios_reptile <- list.files(path = "J:/", pattern="*.gri")
bios_reptile <- paste0("J:/reptile_bios_", list_origin_reptile, ".gri")
bios_reptile <- map(bios_reptile, brick)


```
Old face is name face. Classic PCA is nice as well.  


```{r}

```


t-SNE was fancy for last few years. Check it too.  

```{r}

```

But UMAP is even more fancier these days.

```{r}

```

Don't forget the environmental grid.
```{r}
dismo::gridSample()

```



```{r}

```

Before PCA, Let's check correlation matrix of variables
```{r}
extract_spatial_amphibian_lapply[[1]] %>% 
  mutate_if(is.numeric, scale) %>% 
  select_if(is.numeric) %>% 
  correlate() %>% # (2)
  shave(upper = TRUE) %>% # (3)
  stretch(na.rm = TRUE) %>% 
  arrange(-r) %>% 
  DT::datatable() %>% 
  DT::formatRound("r", digits = 2)

```


```{r}
dismo::threshold()
```

t-SNE, UMAP, PCA, Environment filter? 

## 추가 연구 Further study 
```{r}
library("sampbias")
```
